### 2024-08-17 #7113 @editorialbot
[PRE REVIEW]: ler : LVK (LIGO-Virgo-KAGRA collaboration) event (compact-binary mergers) rate calculator and simulator
**Submitting author:** @hemantaph (<a href="http://orcid.org/0000-0002-0471-3724">Hemantakumar  Phurailatpam</a>)
**Repository:** https://github.com/hemantaph/ler/
**Branch with paper.md** (empty if default branch): main
**Version:** v0.4.0
