### 2024-08-17 #7113 @dfm
In the meantime, can you clean up the `paper` directory and subdirectory? There are a lot of temporary and autogenerated files in there that shouldn't be. I think that's part of the reason why it's not successfully compiling. Then try commenting `@editorialbot generate pdf` on this thread to make sure it works. Thanks!
