+++ arxiv paper/arxiv/paper.bib
@@ -1,0 +1,183 @@
+@article{Janquart2023,
+      author = {Janquart, J and Wright, M and Goyal, S and Chan, J C L and Ganguly, A and Garrón, Á and Keitel, D and Li, A K Y and Liu, A and Lo, R K L and Mishra, A and More, A and Phurailatpam, H and Prasia, P and Ajith, P and Biscoveanu, S and Cremonese, P and Cudell, J R and Ezquiaga, J M and Garcia-Bellido, J and Hannuksela, O A and Haris, K and Harry, I and Hendry, M and Husa, S and Kapadia, S and Li, T G F and Magaña Hernandez, I and Mukherjee, S and Seo, E and Van Den Broeck, C and Veitch, J},
+      title = "{Follow-up analyses to the O3 LIGO–Virgo–KAGRA lensing searches}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+      volume = {526},
+      number = {3},
+      pages = {3832-3860},
+      year = {2023},
+      month = {09},
+      abstract = "{Along their path from source to observer, gravitational waves may be gravitationally lensed by massive objects leading to distortion in the signals. Searches for these distortions amongst the observed signals from the current detector network have already been carried out, though there have as yet been no confident detections. However, predictions of the observation rate of lensing suggest detection in the future is a realistic possibility. Therefore, preparations need to be made to thoroughly investigate the candidate lensed signals. In this work, we present some follow-up analyses that could be applied to assess the significance of such events and ascertain what information may be extracted about the lens-source system by applying these analyses to a number of O3 candidate events, even if these signals did not yield a high significance for any of the lensing hypotheses. These analyses cover the strong lensing, millilensing, and microlensing regimes. Applying these additional analyses does not lead to any additional evidence for lensing in the candidates that have been examined. However, it does provide important insight into potential avenues to deal with high-significance candidates in future observations.}",
+      issn = {0035-8711},
+      doi = {10.1093/mnras/stad2909},
+      url = {https://doi.org/10.1093/mnras/stad2909},
+      eprint = {https://academic.oup.com/mnras/article-pdf/526/3/3832/52108394/stad2909.pdf},
+}
+
+@article{Ashton2019,
+      title={Bilby: A User-friendly Bayesian Inference Library for Gravitational-wave Astronomy},
+      volume={241},
+      ISSN={1538-4365},
+      url={http://dx.doi.org/10.3847/1538-4365/ab06fc},
+      DOI={10.3847/1538-4365/ab06fc},
+      number={2},
+      journal={The Astrophysical Journal Supplement Series},
+      publisher={American Astronomical Society},
+      author={Ashton, Gregory and Hübner, Moritz and Lasky, Paul D. and Talbot, Colm and Ackley, Kendall and Biscoveanu, Sylvia and Chu, Qi and Divakarla, Atul and Easter, Paul J. and Goncharov, Boris and Vivanco, Francisco Hernandez and Harms, Jan and Lower, Marcus E. and Meadors, Grant D. and Melchor, Denyz and Payne, Ethan and Pitkin, Matthew D. and Powell, Jade and Sarin, Nikhil and Smith, Rory J. E. and Thrane, Eric},
+      year={2019},
+      month=apr, pages={27}
+}
+
[…]
