+++ arxiv paper/paper.md
@@ -1,0 +1,125 @@
+---
+title: '$ler$ : LVK (LIGO-Virgo-KAGRA collaboration) event (compact-binary mergers) rate calculator and simulator'
+tags:
+  - Python
+  - astrophysics
+  - statistics
+  - gravitational waves
+  - LIGO
+authors:
+  - name: Hemantakumar Phurailatpam
+    orcid: 0000-0000-0000-0000
+    # equal-contrib: true
+    affiliation: "1"
+  - name: Anupreeta More
+    # equal-contrib: False
+    affiliation: "2,3"
+  - name: Harsh Narola
+    # equal-contrib: False
+    affiliation: "4,5"
+  - name: Ng Chung Yin (Leo)
+    # equal-contrib: False 
+    affiliation: "1"
+  - name: Justin Janquart
+    # equal-contrib: False 
+    affiliation: "4,5,6,7"
+  - name: Chris Van Den Broeck
+    # equal-contrib: False 
+    affiliation: "4,5"
+  - name: Otto Akseli Hannuksela
+    # equal-contrib: False 
+    affiliation: "1"
+  - name: Neha Singh
+    orcid: 0000-0002-1135-3456 
+    # equal-contrib: False
+    affiliation: "8"
+  - name: David Keitel
+    orcid: 0000-0002-2824-626X
+    # equal-contrib: False
+    affiliation: "8"
+affiliations:
+  - name: Department of Physics, The Chinese University of Hong Kong, Shatin, New Territories, Hong Kong
+    index: 1
+  - name:  The Inter-University Centre for Astronomy and Astrophysics (IUCAA), Post Bag 4, Ganeshkhind, Pune 411007, India
+    index: 2
+  - name: Kavli Institute for the Physics and Mathematics of the Universe (IPMU), 5-1-5 Kashiwanoha, Kashiwa-shi, Chiba 277-8583, Japan
+    index: 3
+  - name: Department of Physics, Utrecht University, Princetonplein 1, 3584 CC Utrecht, The Netherlands
+    index: 4
+  - name: Nikhef – National Institute for Subatomic Physics, Science Park, 1098 XG Amsterdam, The Netherlands
+    index: 5
+  - name: Center for Cosmology, Particle Physics and Phenomenology - CP3, Universit\'e Catholique de Louvain, Louvain-La-Neuve, B-1348, Belgium
+    index: 6
+  - name: Royal Observatory of Belgium, Avenue Circulaire, 3, 1180 Uccle, Belgium
+    index: 7
+  - name: Departament de Física, Universitat de les Illes Balears, IAC3-IEEC, Crta. Valldemossa km 7.5, E-07122 Palma, Spain
+    index: 8
+date: 19 June 2024
+bibliography: paper.bib
+---
+
+# Summary
+
[…]
