### 2024-08-20 #7113 @hemantaph
@dfm I have made the necessary changes but the '@editorialbot' still gives an error "An error happened when generating the pdf." Can you please clarify and help me deal with this confusing issue? Thanks in advance.
