+++ .arxiv/paper_ver2.md
@@ -1,115 +1,0 @@
----
-title: '***ler*** : LVK (LIGO-Virgo-KAGRA collaboration) Event (compact-binary mergers) Rate calculator and simulator'
-tags:
-  - Python
-  - astrophysics
-  - statistics
-  - gravitational waves
-  - LIGO
-authors:
-  - name: Phurailatpam Hemantakumar
-    orcid: 0000-0000-0000-0000
-    # equal-contrib: true
-    affiliation: "1"
-  - name: Anupreeta More
-    # equal-contrib: False
-    affiliation: "2,3"
-  - name: Harsh Narola
-    # equal-contrib: False
-    affiliation: "4,5"
-  - name: Ng Chung Yin (Leo)
-    # equal-contrib: False 
-    affiliation: "1"
-  - name: Justin Janquart
-    # equal-contrib: False 
-    affiliation: "4,5"
-  - name: Chris Van Den Broeck
-    # equal-contrib: False 
-    affiliation: "4,5"
-  - name: Otto Akseli HANNUKSELA 
-    # equal-contrib: False 
-    affiliation: "1"
-  - name: Neha Singh
-    # equal-contrib: False
-    affiliation: "6"
-  - name: David Keitel
-    # equal-contrib: False
-    affiliation: "6"
-affiliations:
-  - name: Department of Physics, The Chinese University of Hong Kong, Shatin, New Territories, Hong Kong
-    index: 1
-  - name:  The Inter-University Centre for Astronomy and Astrophysics (IUCAA), Post Bag 4, Ganeshkhind, Pune 411007, India
-    index: 2
-  - name: Kavli Institute for the Physics and Mathematics of the Universe (IPMU), 5-1-5 Kashiwanoha, Kashiwa-shi, Chiba 277-8583, Japan
-    index: 3
-  - name: Department of Physics, Utrecht University, Princetonplein 1, 3584 CC Utrecht, The Netherlands
-    index: 4
-  - name: Nikhef – National Institute for Subatomic Physics, Science Park, 1098 XG Amsterdam, The Netherlands
-    index: 5
-  - name: Departament de F\'isica, Universitat de les Illes Balears, IAC3-IEEC, Crta. Valldemossa km 7.5, E-07122 Palma, Spain
-    index: 6
-  
-date: 19 June 2024
-bibliography: paper.bib
----
-
-### Summary
-
-Gravitational waves (GWs) are ripples in the fabric of space and time caused by the acceleration of unevenly distributed mass or masses. Observable GWs are created especially during the violent cosmic events of merging compact binaries, such as 'binary black holes' (BBH), 'binary neutron stars' (BNS), and 'neutron star and black hole pair' (NSBH). GWs emitted by these events can be distorted or magnified by the gravitational fields of massive objects such as galaxies or galaxy clusters, a phenomenon known as gravitational lensing. Profound comprehension of gravitational lensing's impact on GW signals is imperative to their accurate interpretation and the extraction of astrophysical insights therein. For this purpose, statistical modeling of GWs lensing can provide valuable insights into the properties of the lensing objects and GW sources. Such statistics require accurate and efficient means to calculate the detectable lensing rates, which depend on up-to-date modeling and implementation of lens and source properties and their distribution. The outcomes of these computational analyses not only contribute to generating dependable forecasts but also play an important role in validating forthcoming lensing events [arXiv:2306.03827 [gr-qc]](https://arxiv.org/abs/2306.03827).
-
[…]
