+++ docs/autoapi/ler/gw_source_population/index.rst
@@ -2222,7 +2222,7 @@ Functions
            default: True
 
        **spin_precession** : `bool`
-           If spin_zero=True and spin_precession=True, spin parameters are sampled for precessing binaries.
+           If spin_zero=False and spin_precession=True, spin parameters are sampled for precessing binaries.
            if spin_zero=True and spin_precession=False, spin parameters are sampled for aligned/anti-aligned spin binaries.
            default: False
 
