+++ ler/gw_source_population/cbc_source_parameter_distribution.py
@@ -58,7 +58,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         If True, spin parameters are completely ignore in the sampling.
         default: True
     spin_precession : `bool`
-        If spin_zero=True and spin_precession=True, spin parameters are sampled for precessing binaries.
+        If spin_zero=False and spin_precession=True, spin parameters are sampled for precessing binaries.
         if spin_zero=True and spin_precession=False, spin parameters are sampled for aligned/anti-aligned spin binaries.
         default: False
     directory : `str`
