+++ docs/autoapi/ler/image_properties/image_properties/index.rst
@@ -85,7 +85,7 @@ Classes
 
        **create_new_interpolator** : `dict`
            dictionary to create new interpolator pickle files
-           default: dict(Dl_to_z=dict(create_new=False, resolution=1000))
+           default: dict(luminosity_distance_to_z=dict(create_new=False, resolution=1000))
 
 
 

[…]
