+++ ler/gw_source_population/cbc_source_parameter_distribution.py
@@ -59,7 +59,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         default: True
     spin_precession : `bool`
         If spin_zero=False and spin_precession=True, spin parameters are sampled for precessing binaries.
-        if spin_zero=True and spin_precession=False, spin parameters are sampled for aligned/anti-aligned spin binaries.
+        if spin_zero=False and spin_precession=False, spin parameters are sampled for aligned/anti-aligned spin binaries.
         default: False
     directory : `str`
         Directory to store the interpolator pickle files
