### 2024-09-17 #7113 @hemantaph
> As for possible reviewers, I think the bot found at least 2 good matches of previous JOSS papers with similarities in scope / type of software: deeplenstronomy and LEGWORK. Their authors could be good candidates. (For disclosure, I was a reviewer of the latter; not sure if such a "closed loop" would count as a conflict of interest.)

Thanks for the suggestion @dbkeitel . I will try contacting the authors of deeplenstronomy and LEGWORK, or is there a formal channel where I can ask them to be reviewers?
