+++ ler/rates/ler.py
@@ -488,7 +488,7 @@ class LeR(LensGalaxyParameterDistribution):
         Function to print all the parameters.
         """
         # print all relevant functions and sampler priors
-        print("\n LeR set up params:")
+        print("\n # LeR set up params:")
         print(f'npool = {self.npool},')
         print(f'z_min = {self.z_min},')
         print(f'z_max = {self.z_max},')
@@ -501,23 +501,23 @@ class LeR(LensGalaxyParameterDistribution):
         if self.pdet:
             print(f'pdet_finder = {self.pdet},')
         print(f'json_file_names = {self.json_file_names},')
-        print(f'interpolator_directory = {self.interpolator_directory},')
-        print(f'ler_directory = {self.ler_directory},')
+        print(f"interpolator_directory = '{self.interpolator_directory}',")
+        print(f"ler_directory = '{self.ler_directory}',")
 
-        print("\n LeR also takes CBCSourceParameterDistribution class params as kwargs, as follows:")
+        print("\n # LeR also takes CBCSourceParameterDistribution class params as kwargs, as follows:")
         print(f"source_priors = {self.gw_param_sampler_dict['source_priors']},")
         print(f"source_priors_params = {self.gw_param_sampler_dict['source_priors_params']},")
         print(f"spin_zero = {self.gw_param_sampler_dict['spin_zero']},")
         print(f"spin_precession = {self.gw_param_sampler_dict['spin_precession']},")
         print(f"create_new_interpolator = {self.gw_param_sampler_dict['create_new_interpolator']},")
         
-        print("\n LeR also takes LensGalaxyParameterDistribution class params as kwargs, as follows:")
+        print("\n # LeR also takes LensGalaxyParameterDistribution class params as kwargs, as follows:")
         print(f"lens_type = '{self.gw_param_sampler_dict['lens_type']}',")
         print(f"lens_functions = {self.gw_param_sampler_dict['lens_functions']},")
         print(f"lens_priors = {self.gw_param_sampler_dict['lens_priors']},")
         print(f"lens_priors_params = {self.gw_param_sampler_dict['lens_priors_params']},")
         
-        print("\n LeR also takes ImageProperties class params as kwargs, as follows:")
+        print("\n # LeR also takes ImageProperties class params as kwargs, as follows:")
         print(f"n_min_images = {self.n_min_images},")
         print(f"n_max_images = {self.n_max_images},")
         print(f"geocent_time_min = {self.geocent_time_min},")
@@ -525,7 +525,7 @@ class LeR(LensGalaxyParameterDistribution):
         print(f"lens_model_list = {self.lens_model_list},")
         
         if self.gwsnr:
-            print("\n LeR also takes gwsnr.GWSNR params as kwargs, as follows:")
+            print("\n # LeR also takes gwsnr.GWSNR params as kwargs, as follows:")
             print(f"mtot_min = {self.snr_calculator_dict['mtot_min']},")
             print(f"mtot_max = {self.snr_calculator_dict['mtot_max']},")
             print(f"ratio_min = {self.snr_calculator_dict['ratio_min']},")
@@ -539,7 +539,6 @@ class LeR(LensGalaxyParameterDistribution):
             print(f"psds = {self.snr_calculator_dict['psds']},")
             print(f"ifos = {self.snr_calculator_dict['ifos']},")
             print(f"interpolator_dir = '{self.snr_calculator_dict['interpolator_dir']}',")
-            print(f"create_new_interpolator = {self.snr_calculator_dict['create_new_interpolator']},")
             print(f"gwsnr_verbose = {self.snr_calculator_dict['gwsnr_verbose']},")
[…]
