+++ ler/rates/ler.py
@@ -76,6 +76,23 @@ class LeR(LensGalaxyParameterDistribution):
     interpolator_directory : `str`
         directory to store the interpolators.
         default interpolator_directory = './interpolator_pickle'. This is used for storing the various interpolators related to `ler` and `gwsnr` package.
+    create_new_interpolator : `bool` or `dict`
+        default create_new_interpolator = False.
+        if True, the all interpolators (including `gwsnr`'s)will be created again.
+        if False, the interpolators will be loaded from the interpolator_directory if they exist.
+        if dict, you can specify which interpolators to create new. Complete example (change any of them to True), create_new_interpolator = create_new_interpolator = dict(
+            redshift_distribution=dict(create_new=False, resolution=1000),
+            z_to_luminosity_distance=dict(create_new=False, resolution=1000),
+            velocity_dispersion=dict(create_new=False, resolution=1000),
+            axis_ratio=dict(create_new=False, resolution=1000),
+            optical_depth=dict(create_new=False, resolution=200),
+            z_to_Dc=dict(create_new=False, resolution=1000),
+            Dc_to_z=dict(create_new=False, resolution=1000),
+            angular_diameter_distance=dict(create_new=False, resolution=1000),
+            differential_comoving_volume=dict(create_new=False, resolution=1000),
+            Dl_to_z=dict(create_new=False, resolution=1000),
+            gwsnr=False,
+        )
     ler_directory : `str`
         directory to store the parameters.
         default ler_directory = './ler_data'. This is used for storing the parameters of the simulated events.
@@ -194,7 +211,7 @@ class LeR(LensGalaxyParameterDistribution):
     |                                     | ratio between lensed and         |
     |                                     | unlensed events.                 |
     +-------------------------------------+----------------------------------+
-    |:meth:`~rate_comparision_with_rate_calculation                          |
+    |:meth:`~rate_comparison_with_rate_calculation                          |
     +-------------------------------------+----------------------------------+
     |                                     | Function to calculate rates for  |
     |                                     | unleesed and lensed events and   |
@@ -437,6 +454,7 @@ class LeR(LensGalaxyParameterDistribution):
         list_of_detectors=None,
         json_file_names=None,
         interpolator_directory="./interpolator_pickle",
+        create_new_interpolator=False,
         ler_directory="./ler_data",
         verbose=True,
         **kwargs,
@@ -453,6 +471,7 @@ class LeR(LensGalaxyParameterDistribution):
         if json_file_names:
             self.json_file_names.update(json_file_names)
         self.interpolator_directory = interpolator_directory
+        kwargs["create_new_interpolator"] = create_new_interpolator
         self.ler_directory = ler_directory
         # create directory if not exists
         if not os.path.exists(ler_directory):
@@ -820,6 +839,21 @@ class LeR(LensGalaxyParameterDistribution):
                 if key in input_params:
                     input_params[key] = value
         self.snr_calculator_dict = input_params
+
+        # dealing with create_new_interpolator param
+        if isinstance(input_params["create_new_interpolator"], bool):
+            pass
+        elif isinstance(input_params["create_new_interpolator"], dict):
+            # check input_params["gwsnr"] exists
+            if "gwsnr" in input_params["create_new_interpolator"]:
+                if isinstance(input_params["create_new_interpolator"]["gwsnr"], bool):
+                    input_params["create_new_interpolator"] = input_params["create_new_interpolator"]["gwsnr"]
+                else:
+                    raise ValueError("create_new_interpolator['gwsnr'] should be a boolean.")
+            else:
[…]
