+++ ler/image_properties/image_properties.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """
 This module contains the LensGalaxyPopulation class, which is used to sample lens galaxy parameters, source parameters conditioned on the source being strongly lensed, image properties, and lensed SNRs. \n
-The class inherits from the CompactBinaryPopulation class, which is used to sample source parameters. \n
+The class inherits from the CBCSourceParameterDistribution class, which is used to sample source parameters. \n
 """
 
 import warnings
