+++ docs/autoapi/ler/lens_galaxy_population/index.rst
@@ -94,7 +94,7 @@ Functions
            Dictionary of prior sampler functions and its input parameters.
            Check for available priors and corresponding input parameters by running,
            >>> from ler.gw_source_population import CBCSourceParameterDistribution
-           >>> cbc = CompactBinaryPopulation()
+           >>> cbc = CBCSourceParameterDistribution()
            >>> cbc.available_gw_prior_list_and_its_params()
            # To check the current chosen priors and its parameters, run,
            >>> print("default priors=",cbc.gw_param_samplers)
@@ -140,7 +140,7 @@ Functions
 
    Instance Attributes
    ----------
-   CompactBinaryPopulation has the following instance attributes:
+   CBCSourceParameterDistribution has the following instance attributes:
 
    +-------------------------------------+----------------------------------+
    | Atrributes                          | Type                             |
@@ -178,7 +178,7 @@ Functions
 
    Instance Methods
    ----------
-   CompactBinaryPopulation has the following instance methods:
+   CBCSourceParameterDistribution has the following instance methods:
 
    +-------------------------------------+----------------------------------+
    | Methods                             | Type                             |
@@ -297,7 +297,7 @@ Functions
       .. rubric:: Examples
 
       >>> from ler.gw_source_population import CBCSourceParameterDistribution
-      >>> cbc = CompactBinaryPopulation()
+      >>> cbc = CBCSourceParameterDistribution()
       >>> priors = cbc.available_gw_prior_list_and_its_params
       >>> priors.keys()  # type of priors
       dict_keys(['merger_rate_density', 'source_frame_masses', 'spin', 'geocent_time', 'ra', 'phase', 'psi', 'theta_jn'])
@@ -3259,9 +3259,9 @@ Functions
    .. py:attribute:: cbc_pop
 
       
-      :class:`~CompactBinaryPopulation` class
+      :class:`~CBCSourceParameterDistribution` class
 
-      This is an already initialized class that contains a function (CompactBinaryPopulation.sample_gw_parameters) that actually samples the source parameters.
+      This is an already initialized class that contains a function (CBCSourceParameterDistribution.sample_gw_parameters) that actually samples the source parameters.
 
 
 
