+++ docs/index.rst
@@ -20,7 +20,8 @@ ler
 ``ler`` is closely integarted with a gravitational waves SNR calculator package `gwsnr <https://gwsnr.readthedocs.io/en/latest/>`_.
 
 | ``ler`` main developer: `Hemanta Ph. <https://hemantaph.notion.site/Hemanta-Ph-b4c52e857d424c0db2a68f4e63d6097f>`_
-| ``ler`` developer and analyst: Anupreeta More, Harsh Narola, Ng Chung Yin (Leo), Justin Janquart, Chris Van Den Broeck, Otto Akseli Hannuksela, Neha Singh, David Keitel, 	Yiwei Li (Lisa).
+| ``ler`` developer and analyst: Anupreeta More, Harsh Narola, Ng Chung Yin (Leo), Justin Janquart, Chris Van Den Broeck, Otto Akseli Hannuksela, Neha Singh, David Keitel
+.. Yiwei Li (Lisa).
 
 
 Glossary
