### 2024-12-10 #7420 @hemantaph
@michellegurevich

> ### General checks
> * [ ]  **Contribution and authorship:** Has the submitting author (@hemantaph) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> Author's reply: The authors and contributors are specified in the paper draft and also in the documentation. Are there other places where I need to specify the author list?

> * [ ]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
> Author's reply: 1. The package has commits older than 3 months. 2. How many commits are eligible? 3. ler has 9 authors which are specified in the paper draft. 4. ler has more than 1000 lines of code. 5. The software has been reference in https://arxiv.org/abs/2409.07298 and https://arxiv.org/abs/2407.07526 . 6. Whether the software is sufficiently useful that it is likely to be cited by your peer group. Yes, it has been already cited. Also it has been undergone PnP review in the LIGO community. What else should I take care in this section?
