+++ ler/lens_galaxy_population/lens_galaxy_parameter_distribution.py
@@ -10,7 +10,7 @@ warnings.filterwarnings("ignore")
 import numpy as np
 from numba import njit
 from scipy.integrate import quad
-from lenstronomy.Util.param_util import phi_q2_ellipticity
+# from lenstronomy.Util.param_util import phi_q2_ellipticity
 
 # for redshift to luminosity distance conversion
 from astropy.cosmology import LambdaCDM
@@ -20,7 +20,7 @@ from ..gw_source_population import CBCSourceParameterDistribution
 from .optical_depth import OpticalDepth
 from ..image_properties import ImageProperties
 from ..utils import add_dictionaries_together, trim_dictionary
-from .jit_functions import phi_cut_SIE, velocity_dispersion_z_dependent, lens_redshift_SDSS_catalogue
+from .jit_functions import phi_cut_SIE, velocity_dispersion_z_dependent, lens_redshift_SDSS_catalogue, phi_q2_ellipticity_hemanta
 
 
 class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImageProperties, OpticalDepth):
@@ -216,7 +216,7 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
         z_max=10.0,
         cosmology=None,
         event_type="BBH",
-        lens_type="epl_galaxy",
+        lens_type="epl_shear_galaxy",
         lens_functions= None,
         lens_priors=None,
         lens_priors_params=None,
@@ -375,7 +375,7 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
         )
         input_params_image.update(params)
 
-        print("input_params_image", input_params_image)
+        # print("input_params_image", input_params_image)
         ImageProperties.__init__(
             self,
             npool=self.npool,
@@ -423,7 +423,7 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
             dictionary of lens functions
         """
 
-        if lens_type == "epl_galaxy":
+        if lens_type == "epl_shear_galaxy":
             lens_priors_ = dict(
                 source_redshift_sl="strongly_lensed_source_redshifts",
                 lens_redshift="lens_redshift_SDSS_catalogue",
@@ -487,7 +487,7 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
             size=size, lens_parameters_input=lens_parameters_input
         )
 
-    def sample_all_routine(self, size=1000, lens_parameters_input=None):
+    def sample_all_routine_SIE(self, size=1000, lens_parameters_input=None):
         """
         Function to sample galaxy lens parameters along with the source parameters.
 
@@ -590,7 +590,133 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
             lens_parameters["phi"] = self.sample_axis_rotation_angle(size=size)
 
             # Transform the axis ratio and the angle, to ellipticities e1, e2, using lenstronomy
-            lens_parameters["e1"], lens_parameters["e2"] = phi_q2_ellipticity(
+            lens_parameters["e1"], lens_parameters["e2"] = phi_q2_ellipticity_hemanta(
+                lens_parameters["phi"], lens_parameters["q"]
+            )
+
+            # Sample shears
+            lens_parameters["gamma1"], lens_parameters["gamma2"] = self.sample_shear(
+                size=size)
+
+            # Sample the spectral index of the mass density distribution
+            lens_parameters["gamma"] = self.sample_mass_density_spectral_index(
+                size=size)
+
+            # sample gravitional waves source parameter
+            param = dict(zs=lens_parameters["zs"])
+            if samplers_params["source_parameters"]:
[…]
