+++ ler/lens_galaxy_population/optical_depth.py
@@ -4,6 +4,8 @@ import numpy as np
 from scipy.integrate import quad
 from scipy.stats import gengamma
 from astropy.cosmology import LambdaCDM
+from tqdm import tqdm
+
 from ..utils import  interpolator_from_pickle, cubic_spline_interpolator, inverse_transform_sampler
 from .jit_functions import phi_cut_SIE, axis_ratio_rayleigh, axis_ratio_SIS, phi, phi_loc_bernardi
 
@@ -156,27 +158,37 @@ class OpticalDepth():
         npool=4,
         z_min=0.001,
         z_max=10.0,
+        lens_type="sie_galaxy",
         optical_depth_function=None,
         sampler_priors=None,
         sampler_priors_params=None,
         cosmology=None,
         directory="./interpolator_pickle",
         create_new_interpolator=False,
+        verbose=False,
     ):
 
         self.npool = npool
         self.z_min = z_min
         self.z_max = z_max
         self.cosmo = cosmology if cosmology else LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+        self.lens_type = lens_type
 
         self.sampler_priors = dict(
-            velocity_dispersion="velocity_dispersion_bernardi",
+            velocity_dispersion="velocity_dispersion_gengamma",
             axis_ratio="axis_ratio_rayleigh",
         )
         if sampler_priors:
             self.sampler_priors.update(sampler_priors)
+
+        vd_name = self.sampler_priors["velocity_dispersion"]  # velocity dispersion sampler name
+        if vd_name in self.available_velocity_dispersion_list_and_its_params:
+            vd_priors_params = self.available_velocity_dispersion_list_and_its_params[vd_name]
+        else:
+            vd_priors_params = None
+
         self.sampler_priors_params = dict(
-            velocity_dispersion=dict(vd_min=0., vd_max=350.),
+            velocity_dispersion=vd_priors_params,
             axis_ratio=dict(q_min=0.2, q_max=1.),
         )
         if sampler_priors_params:
@@ -207,9 +219,6 @@ class OpticalDepth():
                 differential_comoving_volume=dict(create_new=True, resolution=1000),
             )
 
-        vd_name = self.sampler_priors["velocity_dispersion"]  # velocity dispersion sampler name
-        tau_name = optical_depth_function  # optical depth function name
-
         self.create_lookup_table_fuction(self.z_max)
 
         # setting velocity dispersion sampler method
@@ -219,7 +228,16 @@ class OpticalDepth():
         if callable(optical_depth_function):
             self.strong_lensing_optical_depth = optical_depth_function
         else:
-            self.initialize_optical_depth_function(tau_name, vd_name);
+            self.initialize_optical_depth_function(optical_depth_function);
+    
+        if verbose:
+            print(f"Chosen velocity dispersion sampler: {vd_name}")
+            print(f"Chosen velocity dispersion sampler params: {self.sampler_priors_params['velocity_dispersion']}")
+            print(f"Chosen optical depth function: {optical_depth_function}")
+            print(f"Chosen axis ratio sampler: {self.sampler_priors['axis_ratio']}")
+            print(f"Chosen axis ratio sampler params: {self.sampler_priors_params['axis_ratio']}")
+            print(f"Chosen cosmology: {self.cosmo}")
+            print(f"Chosen lens_type: {self.lens_type}")
         
     def initialize_velocity_dispersion_sampler(self, vd_name):
         """
@@ -231,32 +249,56 @@ class OpticalDepth():
             name of velocity dispersion sampler
         """
 
[…]
