+++ ler/image_properties/multiprocessing_routine.py
@@ -7,8 +7,8 @@ import numpy as np
 from lenstronomy.LensModel.lens_model import LensModel
 from lenstronomy.LensModel.Solver.lens_equation_solver import LensEquationSolver
 from lenstronomy.LensModel.Solver.epl_shear_solver import caustics_epl_shear
-from lenstronomy.Util.param_util import phi_q2_ellipticity
-from ..lens_galaxy_population.jit_functions import axis_ratio_rayleigh
+# from lenstronomy.Util.param_util import phi_q2_ellipticity
+# from ..lens_galaxy_population.jit_functions import axis_ratio_rayleigh
 
 # For sampling from caustic
 from shapely.geometry import Polygon
@@ -38,7 +38,7 @@ def solve_lens_equation(lens_parameters):
     Returns
     -------
     x_source : `float`
-        x position of the source in the source plane
+        x position of the source in the source plane, unit: arcsec
     y_source : `float`
         y position of the source in the source plane
     x0_image_position : `float`
@@ -242,208 +242,3 @@ def solve_lens_equation(lens_parameters):
     )
 
 
-# def solve_lens_equation(lens_parameters):
-#     """
-#     Function to solve the lens equation (min_image = 2)
-
-#     Parameters
-#     ----------
-#     lens_parameters : `list`
-#         a list of parameters
-#         lens_parameters[0] = min_images : minimum number of images
-#         lens_parameters[1] = e1 : ellipticity
-#         lens_parameters[2] = e2 : ellipticity
-#         lens_parameters[3] = gamma : power-law index
-#         lens_parameters[4] = gamma1 : shear
-#         lens_parameters[5] = gamma2 : shear
-#         lens_parameters[6] = zl : redshift of the lens
-#         lens_parameters[7] = zs : redshift of the source
-#         lens_parameters[8] = einstein_radius : Einstein radius
-#         lens_parameters[9] = iteration : iteration number
-#         lens_parameters[10:] = lens_model_list : numpy array of lens models
-
-#     Returns
-#     -------
-#     x_source : `float`
-#         x position of the source in the source plane
-#     y_source : `float`
-#         y position of the source in the source plane
-#     x0_image_position : `float`
-#         x position of the images in the source plane
-#     x1_image_position : `float`
-#         y position of the images in the source plane
-#     magnifications : `float`
-#         magnification of the images
-#     time_delays : `float`
-#         time-delay of the images
-#     nImages : `int`
-#         number of images
-#     determinant : `float`
-#         determinant of the hessian matrix
-#     trace : `float`
-#         trace of the hessian matrix
-#     iteration : `int`
-#         iteration number
-
-#     Examples
-#     --------
-#     >>> from ler.multiprocessing_routine import solve_lens_equation1
-#     >>> import numpy as np
-#     >>> from multiprocessing import Pool
-#     >>> # lens parameters input contains 12 parameters [e1, e2, gamma, gamma1, gamma2, zl, zs, einstein_radius, iteration, lens_model_list]
-#     >>> lens_parameters1 = np.array([2, 0.024069457093642648, -0.016002190961948142, 1.8945414936459974, 0.10117465203892329, 0.09600089396968613, 0.2503743800068136, 0.9418211055453296, 2.5055790287104725e-06, 0, 'EPL_NUMBA', 'SHEAR'], dtype=object)
-#     >>> lens_parameters2 = np.array([2, -0.04030088581646998, -0.01419438113690042, 2.0068239327017, 0.08482718989370612, -0.015393332086560785, 1.0952303138971118, 2.5534097159384417, 1.0125570159563301e-06, 1, 'EPL_NUMBA', 'SHEAR'], dtype=object)
[…]
