### 2025-02-06 hemantaph/ler#12 @kazewong
JOSS paper writing review
https://github.com/openjournals/joss-reviews/issues/7420

This is an issue created for reviewing the writing of the linked JOSS paper. I am going through the papers and marking some of the adjustments down. I am copying the checklist here as a reminder for review items
