### 2025-02-06 hemantaph/ler#13 @kazewong
Joss review - Missing unit test
https://github.com/openjournals/joss-reviews/issues/7420

I do not see any test in the current main branch. Here is the Joss review guideline regarding testing:

Tests

Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

    Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
    OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
    Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works

Here are a list of items I wish to see done on the testing side:

- [ ] Implement a testing suite. This should at least have unit tests that assert the correctness of individual components of the code. On top of that, integration tests are useful to ensure the code can serve its intended usage. https://github.com/kazewong/flowMC/tree/main/test can be a (non-exhaustive) reference point.
- [ ] Implement a Github action that will trigger testing whenever there are changes pushed to the main branch.
