+++ requirements.txt
@@ -1,13 +1,15 @@
 setuptools>=67.8.0
+matplotlib>=3.4.2
+pycbc>=2.0.4
 numpy>=1.18
-numba>=0.57.1
+numba>=0.60.0
 bilby>=1.0.2
-gwsnr>=0.3.0
-scipy>=1.9.0
+gwsnr>=0.3.2
+scipy<1.14.0
 lenstronomy>=1.10.4
 astropy>=5.1
 tqdm>=4.64.1
 pointpats>=2.3
 shapely>=2.0.1
-gwcosmo
+gwcosmo==2.1.0
 #gwcosmo@git+https://git.ligo.org/lscsoft/gwcosmo.git@v1.0.0
