+++ test/LeR_example.py
@@ -1,0 +1,196 @@
+# This is a short example to simulate lensed and unlensed binary black hole mergers and calculate their rates ($yr^{-1}$) and finally compare the results.
+# LeR settings and important results (like simulated events file names and location, rates and rate ratios, etc.) will be saved in the `./ler_data/ler_params.json` file.
+# The unlensed events will be saved in the `./ler_data/unlensed_param.json` file and the detectable unlensed events will be saved in the `./ler_data/unlensed_param_detectable.json` file.
+# The lensed events will be saved in the `./ler_data/lensed_param.json` file and the detectable lensed events will be saved in the `./ler_data/lensed_param_detectable.json` file.
+# A plot of the redshift distributions of lensed and unlensed events will be saved in the current directory.
+
+
+# call the LeR class
+from ler.rates import LeR
+
+# other necessary imports
+from astropy.cosmology import LambdaCDM
+
+# define the main function to avoid multiprocessing issues
+def main():
+    """
+    Function to simulate lensed and unlensed binary black hole mergers and calculate their rates ($yr^{-1}$) and finally compare the results. The outputs are saved in the `ler_data` directory by default.
+    """
+
+    # if you want to initialize with the default values, you can simply call the class without any arguments; ler = LeR()
+    # below I am showing an example with all the arguments, including the arguments of the parent classes (CBCSourceParameterDistribution, LensGalaxyParameterDistribution, ImageProperties) and the arguments of the gwsnr package.
+    ler = LeR(
+        # LeR setup arguments
+        npool=4, # number of processors to use
+        z_min=0.0, # minimum redshift
+        z_max=10.0, # maximum redshift
+        event_type='BBH', # event type
+        size=100000, # number of events to simulate
+        batch_size=50000, # batch size
+        cosmology=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7), # cosmology
+        snr_finder=None, # snr calculator from 'gwsnr' package will be used
+        pdet_finder=None,  # will not be consider unless specified
+        list_of_detectors=None, # list of detectors that will be considered when calculating snr or pdet for lensed events. if None, all the detectors from 'gwsnr' will be considered
+        json_file_names=dict(
+            ler_params="ler_params.json", # to store initialization parameters and important results
+            unlensed_param="unlensed_param.json", # to store all unlensed events
+            unlensed_param_detectable="unlensed_param_detectable.json", # to store only detectable unlensed events
+            lensed_param="lensed_param.json", # to store all lensed events 
+            lensed_param_detectable="lensed_param_detectable.json"), # to store only detectable lensed events
+        interpolator_directory='./interpolator_pickle', # directory to store the interpolator pickle files. 'ler' uses interpolation to get values of various functions to speed up the calculations (relying on numba njit).
+        ler_directory='./ler_data', # directory to store all the outputs
+        verbose=True, # if True, will print all information at initialization
+
+        # CBCSourceParameterDistribution class arguments
[…]
