+++ requirements.txt
@@ -12,4 +12,5 @@ tqdm>=4.64.1
 pointpats>=2.3
 shapely>=2.0.1
 gwcosmo==2.1.0
+
 #gwcosmo@git+https://git.ligo.org/lscsoft/gwcosmo.git@v1.0.0
