+++ ler/lens_galaxy_population/lens_galaxy_parameter_distribution.py
@@ -44,7 +44,7 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
         default: 'BBH'
     lens_type : `str`
         Type of lens galaxy to generate.
-        default: 'epl_galaxy'
+        default: 'lens_functions'
     lens_functions, lens_priors, lens_priors_params : `dict`, `dict`, `dict`
         dictionary of lens functions, priors, and priors parameters
         Check for default/available lens functions, priors and corresponding input parameters by running,\n
@@ -486,134 +486,8 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
         return self.sample_lens_parameters_routine(
             size=size, lens_parameters_input=lens_parameters_input
         )
-
-    def sample_all_routine_SIE(self, size=1000, lens_parameters_input=None):
-        """
-        Function to sample galaxy lens parameters along with the source parameters.
-
-        Parameters
-        ----------
-        size : `int`
-            number of lens parameters to sample
-        lens_parameters_input : `dict`
-            dictionary of lens parameters to sample
-
-        Returns
-        -------
-        lens_parameters : `dict`
-            dictionary of lens parameters and source parameters (lens conditions applied): \n
-            zl: lens redshifts \n
-            zs: source redshifts, lensed condition applied\n
-            sigma: velocity dispersions \n
-            q: axis ratios \n
-            theta_E: Einstein radii \n
-            phi: axis rotation angle \n
-            e1: ellipticity component 1 \n
-            e2: ellipticity component 2 \n
-            gamma1: shear component 1 \n
-            gamma2: shear component 2 \n
-            gamma: spectral index of the mass density distribution \n
-            geocent_time: time of arrival of the unlensed signal\n
-            phase: phase of the unlensed signal\n
-            psi: polarization angle of the unlensed signal\n
-            theta_jn: inclination angle of the unlensed signal\n
-            luminosity_distance: luminosity distance of the source\n
-            mass_1_source: mass 1 (larger) of the source\n
-            mass_2_source: mass 2 (smaller) of the source\n
-            ra: right ascension of the source\n
-            dec: declination of the source\n
-
-        Examples
-        --------
-        >>> from ler.lens_galaxy_population import LensGalaxyParameterDistribution
-        >>> lens = LensGalaxyParameterDistribution()
-        >>> lens.sample_all_routine(size=1000)
-        """
-
-        if lens_parameters_input is None:
-            lens_parameters_input = dict()
-        samplers_params = self.lens_param_samplers_params.copy()
-
-        # Sample source redshifts from the source population
-        # rejection sampled with optical depth
-        zs = self.sample_source_redshift_sl(size=size)
-
-        # Sample lens redshifts
-        zl = self.sample_lens_redshift(zs=zs)
-
-        # Sample velocity dispersions
-        try:
-            sigma = self.sample_velocity_dispersion(len(zs))
-        except:
-            sigma = self.sample_velocity_dispersion(len(zs), zl)
-
-        # Sample axis ratios
-        try:
-            q = self.sample_axis_ratio(sigma)
-        except:
-            q = self.sample_axis_ratio(len(sigma))
-
-        # Compute the Einstein radii
-        theta_E = self.compute_einstein_radii(sigma, zl, zs)
-
-        # Create a dictionary of the lens parameters
-        lens_parameters = {
-            "zl": zl,
-            "zs": zs,
-            "sigma": sigma,
-            "q": q,
-            "theta_E": theta_E,
-        }
-
-        # Rejection sample based on the lensing probability, that is, rejection sample wrt theta_E
-        lens_parameters = self.rejection_sample_sl(
-            lens_parameters
-        )  # proportional to pi theta_E^2
-
[…]
