+++ ler/gw_source_population/cbc_source_redshift_distribution.py
@@ -11,15 +11,11 @@ import warnings
 warnings.filterwarnings("ignore")
 
 import numpy as np
-from numba import njit
-
-# from gwcosmo import priors as p
-from scipy.integrate import quad
 
 # for redshift to luminosity distance conversion
 from astropy.cosmology import LambdaCDM
 
-from ..utils import  interpolator_from_pickle, cubic_spline_interpolator, inverse_transform_sampler
+from ..utils import  FunctionConditioning
 from .jit_functions import merger_rate_density_bbh_popI_II_oguri2018, star_formation_rate_madau_dickinson2014, merger_rate_density_bbh_popIII_ken2022, merger_rate_density_bbh_primordial_ken2022
 
 
@@ -71,6 +67,8 @@ class CBCSourceRedshiftDistribution(object):
     +-------------------------------------+----------------------------------+
     |:attr:`~z_max`                       | `float`                          |
     +-------------------------------------+----------------------------------+
+    |:attr:`~directory`                   | `str`                            |
+    +-------------------------------------+----------------------------------+
     |:attr:`~event_type`                  | `str`                            |
     +-------------------------------------+----------------------------------+
     |:attr:`~cosmo`                       | `astropy.cosmology`              |
@@ -85,9 +83,15 @@ class CBCSourceRedshiftDistribution(object):
     |                                     | merger rate density functions    |
     |                                     | and its parameters               |
     +-------------------------------------+----------------------------------+
-    |:attr:`~sample_source_redshift`      | Function to sample source        |
-    |                                     | redshifts                        |
+    |:attr:`~merger_rate_density`         | `class object`                   |
     +-------------------------------------+----------------------------------+
+    |:attr:`~source_redshift`             | `class object`                   |
+    +-------------------------------------+----------------------------------+
+    |:attr:`~luminosity_distance`         | `class object`                   |
+    +-------------------------------------+----------------------------------+
+    |:attr:`~differential_comoving_volume`| `class object`                   |
+    +-------------------------------------+----------------------------------+
+
 
     Instance Methods
     ----------
@@ -95,16 +99,7 @@ class CBCSourceRedshiftDistribution(object):
     +-------------------------------------+----------------------------------+
     | Methods                             | Type                             |
     +=====================================+==================================+
-    |:attr:`~merger_rate_density`         | `function`                       |
-    +-------------------------------------+----------------------------------+
-    |:attr:`~z_to_luminosity_distance`    | `function`                       |
-    +-------------------------------------+----------------------------------+
-    |:attr:`~differential_comoving_volume`| `function`                       |
-    +-------------------------------------+----------------------------------+
-    |:meth:`~pdf_z`                       | Function to compute the pdf      |
-    |                                     | p(z)                             |
-    +-------------------------------------+----------------------------------+
-    |:meth:`~merger_rate_density_detector_frame`                                  |
+    |:meth:`~merger_rate_density_detector_frame`                             |
     +-------------------------------------+----------------------------------+
     |                                     | Function to compute the merger   |
     |                                     | rate density (detector frame)    |
@@ -167,9 +162,9 @@ class CBCSourceRedshiftDistribution(object):
     Dictionary of merger rate density function input parameters
     """
 
-    c_n_i = None
[…]
