+++ ler/lens_galaxy_population/lens_galaxy_parameter_distribution.py
@@ -8,175 +8,172 @@ Either the class takes in initialized CBCSourceParameterDistribution class as in
 import warnings
 warnings.filterwarnings("ignore")
 import numpy as np
-from numba import njit
 from scipy.integrate import quad
 # from lenstronomy.Util.param_util import phi_q2_ellipticity
-
-# for redshift to luminosity distance conversion
-from astropy.cosmology import LambdaCDM
+from multiprocessing import Pool
 
 # the following .py file will be called if they are not given in the class initialization
 from ..gw_source_population import CBCSourceParameterDistribution
 from .optical_depth import OpticalDepth
 from ..image_properties import ImageProperties
-from ..utils import add_dictionaries_together, trim_dictionary
-from .jit_functions import phi_cut_SIE, velocity_dispersion_z_dependent, lens_redshift_SDSS_catalogue, phi_q2_ellipticity_hemanta
-
+from ..utils import add_dictionaries_together, trim_dictionary, FunctionConditioning, interpolator_pickle_path
+from .jit_functions import phi_cut_SIE, phi_q2_ellipticity_hemanta
+from .mp import cross_section_mp
 
 class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImageProperties, OpticalDepth):
     """
-    Class to sample lens galaxy parameters, source parameters conditioned on the source being strongly lensed, and image properties \n
+        Class to sample lens galaxy parameters, source parameters conditioned on the source being strongly lensed, and image properties \n
 
-    Parameters
-    ----------
-    npool : `int`
-        number of processors to use
-    z_min : `float`
-        minimum redshift
-    z_max : `float`
-        maximum redshift    
-    cosmology : `astropy.cosmology`
-        Cosmology to use
-        default: None/astropy.cosmology.FlatLambdaCDM(H0=70, Om0=0.3)
-    event_type : `str`
-        Type of event to generate.
-        e.g. 'BBH', 'BNS', 'NSBH'
-        default: 'BBH'
-    lens_type : `str`
-        Type of lens galaxy to generate.
-        default: 'epl_galaxy'
-    lens_functions, lens_priors, lens_priors_params : `dict`, `dict`, `dict`
-        dictionary of lens functions, priors, and priors parameters
-        Check for default/available lens functions, priors and corresponding input parameters by running,\n
+        Parameters
+        ----------
+        npool : `int`
+            number of processors to use
+        z_min : `float`
+            minimum redshift
+        z_max : `float`
+            maximum redshift    
+        cosmology : `astropy.cosmology`
+            Cosmology to use
+            default: None/astropy.cosmology.FlatLambdaCDM(H0=70, Om0=0.3)
+        event_type : `str`
+            Type of event to generate.
+            e.g. 'BBH', 'BNS', 'NSBH'
+            default: 'BBH'
+        lens_type : `str`
+            Type of lens galaxy to generate.
+            default: 'epl_shear_galaxy'
+        lens_functions, lens_priors, lens_priors_params : `dict`, `dict`, `dict`
+            dictionary of lens functions, priors, and priors parameters
+            Check for default/available lens functions, priors and corresponding input parameters by running,\n
+            >>> from ler.lens_galaxy_population import LensGalaxyParameterDistribution
+            >>> lens = LensGalaxyParameterDistribution()
+            >>> print(lens.lens_functions)
+            >>> print(lens.lens_priors)
+            >>> print(lens.lens_priors_params)
+        directory : `str`
+            directory to store the interpolators
+            default: './interpolator_pickle'
+        **kwargs : 
+            keyword arguments to pass to the parent classes
+
+        Examples
+        --------
         >>> from ler.lens_galaxy_population import LensGalaxyParameterDistribution
[…]
