+++ ler/rates/ler_UML.txt
@@ -1,66 +1,0 @@
-@startuml
-class LeR {
-    snr_finder()
-    pdet_finder()
-    initialization()
-    print_all_params()
-    snr()
-    snr()
-    unlensed_param()
-    unlensed_param_detectable()
-    lensed_param()
-    lensed_param_detectable()
-    class_initialization()
-    gwsnr_intialization()
-    store_ler_params()
-    unlensed_cbc_statistics()
-    unlensed_sampling_routine()
-    unlensed_rate()
-    _load_param()
-    _recalculate_snr_unlensed()
-    _find_detectable_index_unlensed()
-    rate_function()
-    _save_detectable_params()
-    _append_ler_param()
-    lensed_cbc_statistics()
-    lensed_sampling_routine()
-    lensed_rate()
-    _check_snr_threshold_lensed()
-    _recalculate_snr_lensed()
-    _find_detectable_index_lensed()
-    rate_comparison_with_rate_calculation()
-    rate_ratio()
-    selecting_n_unlensed_detectable_events()
-    selecting_n_lensed_detectable_events()
-    _initial_setup_for_n_event_selection()
-    _trim_results_to_size()
-    _append_meta_data()
-        npool
-    z_min
-    z_max
-    event_type
-    cosmo
-    size
-    batch_size
-    json_file_names
-    interpolator_directory
-    ler_directory
-    gwsnr
-    pdet
-    snr
-    pdet
-    gwsnr
-    list_of_detectors
-    _snr
-    gw_param_sampler_dict
-    snr_calculator_dict
-    snr
-    list_of_detectors
-    snr_bilby
-    dict_buffer
-    dict_buffer
-    batch_size
-    __init__()
-}
-LensGalaxyParameterDistribution <|-- LeR
-@enduml

[…]
