+++ ler/rates/ler.py
@@ -460,6 +460,7 @@ class LeR(LensGalaxyParameterDistribution):
         **kwargs,
     ):
         
+        print("\nInitializing LeR class...\n")
         self.npool = npool
         self.z_min = z_min
         self.z_max = z_max
@@ -497,12 +498,12 @@ class LeR(LensGalaxyParameterDistribution):
         # if verbose, prevent anything from printing
         if verbose:
             initialization()
-            self.print_all_params()
+            self.print_all_params_ler()
         else:
             with contextlib.redirect_stdout(None):
                 initialization()
             
-    def print_all_params(self):
+    def print_all_params_ler(self):
         """
         Function to print all the parameters.
         """
@@ -533,8 +534,8 @@ class LeR(LensGalaxyParameterDistribution):
         print("\n # LeR also takes LensGalaxyParameterDistribution class params as kwargs, as follows:")
         print(f"lens_type = '{self.gw_param_sampler_dict['lens_type']}',")
         print(f"lens_functions = {self.gw_param_sampler_dict['lens_functions']},")
-        print(f"lens_priors = {self.gw_param_sampler_dict['lens_priors']},")
-        print(f"lens_priors_params = {self.gw_param_sampler_dict['lens_priors_params']},")
+        print(f"sampler_priors = {self.gw_param_sampler_dict['sampler_priors']},")
+        print(f"sampler_priors_params = {self.gw_param_sampler_dict['sampler_priors_params']},")
         
         print("\n # LeR also takes ImageProperties class params as kwargs, as follows:")
         print(f"n_min_images = {self.n_min_images},")
@@ -604,14 +605,16 @@ class LeR(LensGalaxyParameterDistribution):
         print("axis_ratio_params = ", self.lens_param_samplers_params["axis_ratio"])
         print(f"axis_rotation_angle = '{self.lens_param_samplers['axis_rotation_angle']}'")
         print("axis_rotation_angle_params = ", self.lens_param_samplers_params["axis_rotation_angle"])
-        print(f"shear = '{self.lens_param_samplers['shear']}'")
-        print("shear_params = ", self.lens_param_samplers_params["shear"])
-        print(f"mass_density_spectral_index = '{self.lens_param_samplers['mass_density_spectral_index']}'")
-        print("mass_density_spectral_index_params = ", self.lens_param_samplers_params["mass_density_spectral_index"])
+        print(f"shear = '{self.lens_param_samplers['external_shear']}'")
+        print("shear_params = ", self.lens_param_samplers_params['external_shear'])
+        print(f"density_profile_slope = '{self.lens_param_samplers['density_profile_slope']}'")
+        print("density_profile_slope_params = ", self.lens_param_samplers_params["density_profile_slope"])
         # lens functions
         print("Lens functions:")
         print(f"strong_lensing_condition = '{self.lens_functions['strong_lensing_condition']}'")
         print(f"optical_depth = '{self.lens_functions['optical_depth']}'")
+        print(f"optical_depth_params = '{self.lens_functions_params['optical_depth']}'")
+        print(f"param_sampler_type = '{self.lens_functions['param_sampler_type']}'")
 
     @property
     def snr(self):
@@ -743,10 +746,12 @@ class LeR(LensGalaxyParameterDistribution):
             z_max=self.z_max,
             cosmology=self.cosmo,
             event_type=self.event_type,
-            lens_type="epl_galaxy",
+            lens_type="epl_shear_galaxy",
             lens_functions= None,
-            lens_priors=None,
-            lens_priors_params=None,
+            lens_functions_params=None,
+            sampler_priors=None,
+            sampler_priors_params=None,
+            buffer_size=int(2e5),
             # ImageProperties class params
[…]
