+++ ler/gw_source_population/cbc_source_parameter_distribution.py
@@ -20,12 +20,12 @@ from gwcosmo import priors as p
 
 # for multiprocessing
 # Import helper routines
-from ..utils import interpolator_from_pickle, cubic_spline_interpolator
+from ..utils import FunctionConditioning
 
 # import redshift distribution sampler
 from .cbc_source_redshift_distribution import CBCSourceRedshiftDistribution
 
-from .jit_functions import lognormal_distribution_2D, inverse_transform_sampler_m1m2
+from .jit_functions import lognormal_distribution_2D, bns_bimodal_pdf, inverse_transform_sampler_m1m2
 
 
 class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
@@ -72,7 +72,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
     ----------
     >>> from ler.gw_source_population import CBCSourceParameterDistribution
     >>> cbc = CBCSourceParameterDistribution()
-    >>> params = cbc.sample_gw_parameters(size=1000)
+    >>> params = cbc.gw_parameters(size=1000)
     >>> print("sampled parameters=",list(params.keys()))
 
     Instance Attributes
@@ -128,46 +128,46 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
     |                                     | table for converting redshift    |
     |                                     | to luminosity distance           |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_gw_parameters`        | Function to sample all the       |
+    |:meth:`~gw_parameters`        | Function to sample all the       |
     |                                     | intrinsic and extrinsic          |
     |                                     | parameters of compact binaries   |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_source_frame_masses`  | Function to sample source mass1  |
+    |:meth:`~source_frame_masses`  | Function to sample source mass1  |
     |                                     | and mass2                        |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_geocent_time`         | Function to sample geocent time  |
+    |:meth:`~geocent_time`         | Function to sample geocent time  |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_zs`                   | Function to sample source        |
+    |:meth:`~zs`                   | Function to sample source        |
     |                                     | redshift                         |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_ra`                   | Function to sample right         |
+    |:meth:`~ra`                   | Function to sample right         |
     |                                     | ascension (sky position)         |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_dec`                  | Function to sample declination   |
+    |:meth:`~dec`                  | Function to sample declination   |
     |                                     | (sky position)                   |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_phase`                | Function to sample coalescence   |
+    |:meth:`~phase`                | Function to sample coalescence   |
     |                                     | phase                            |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_psi`                  | Function to sample polarization  |
+    |:meth:`~psi`                  | Function to sample polarization  |
     |                                     | angle                            |
     +-------------------------------------+----------------------------------+
-    |:meth:`~sample_theta_jn`             | Function to sample inclination   |
+    |:meth:`~theta_jn`             | Function to sample inclination   |
[…]
