+++ ler/utils/__init__.py
@@ -1,3 +1,4 @@
 from .utils import *
 from .plots import *
 from .gwsnr_training_data_generator import *
+from .function_interpolation import *

[…]
