### 2025-02-18 hemantaph/ler#12 @kazewong
Regarding summary:

The summary of this paper read a bit mixed in terms of the purpose of the software. I understand the code can be in principle used for modeling an unlensed population, but the majority of the codebase is geared toward modeling a population of lensed gravitational wave event. For example, the summary page of the [doc ](https://ler.readthedocs.io/en/latest/Summary.html ) starts with strong lensing of gravitational wave.

I think the summary should more accurately reflect the main intended usage of the code instead of encapsulating all the possible use cases. This helps the user better identify the purpose of the software.
