### 2025-02-18 hemantaph/ler#12 @kazewong
Regarding statement of need, I believe the paper can be benefited from similar focusing as the summary
