### 2025-02-24 hemantaph/ler#14 @kazewong
JOSS Review - missing community guide
https://github.com/openjournals/joss-reviews/issues/7420

Per JOSS review guideline :

- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Currently there is no community guideline I could find on the repository. Please adjust accordingly. As a starting point, I can open a PR pasting a template for the community guide I used in the past, and the authors are welcome to either modified it or craft one from scratch.
This is the community guide I used for a code base I made. https://github.com/kazewong/flowMC/blob/main/CONTRIBUTING.md
