### 2025-02-24 b90f29e
Add community guidelines
Related to #14

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/hemantaph/ler/issues/14?shareId=XXXX-XXXX-XXXX-XXXX).
