### 2025-02-27 hemantaph/ler#12 @kazewong
While the quality of the writing is decent, I do have some suggestions regarding statements related to the performance of the code.

- [ ] In both the summary and the statement of need, the authors mentioned the calculation needed is computational intensive, but there is not quantitative description of the computational load. Considering some of the example contains these estimates and the speedup provided by this code, I believe quoting an estimate in the paper can help the readers understand the strength of this package. For example there are some estimate hidden in [this notebook](https://ler.readthedocs.io/en/latest/examples/rates/LeR%20complete%20examples.html)

- [ ] Also, I am not exactly sure whether equations section is needed for the JOSS paper. They seem quite out of place and were not referenced directly in the paper. Furthermore, more precise versions of the equations are detailed in the docs, so I will suggest to remove the equations.
