+++ ler/lens_galaxy_population/__init__.py
@@ -2,3 +2,5 @@ from .lens_galaxy_parameter_distribution import *
 from .optical_depth import *
 from .mp import *
 from .jit_functions import *
+
+from .import lens_param_data
