+++ ler/gw_source_population/cbc_source_redshift_distribution.py
@@ -11,12 +11,16 @@ import warnings
 warnings.filterwarnings("ignore")
 
 import numpy as np
+from multiprocessing import Pool
+from tqdm import tqdm
+from scipy.interpolate import CubicSpline
 
 # for redshift to luminosity distance conversion
 from astropy.cosmology import LambdaCDM
 
-from ..utils import  FunctionConditioning
-from .jit_functions import merger_rate_density_bbh_popI_II_oguri2018, star_formation_rate_madau_dickinson2014, merger_rate_density_bbh_popIII_ken2022, merger_rate_density_bbh_primordial_ken2022
+from ..utils import  FunctionConditioning, interpolator_pickle_path
+from .jit_functions import merger_rate_density_bbh_popI_II_oguri2018, sfr_madau_dickinson2014, merger_rate_density_bbh_popIII_ken2022, merger_rate_density_bbh_primordial_ken2022
+from .sfr_with_time_delay import sfr_with_time_delay
 
 
 class CBCSourceRedshiftDistribution(object):
@@ -43,7 +47,7 @@ class CBCSourceRedshiftDistribution(object):
         for others see instance method in :class:`~ler.ler.merger_rate_density_model_list`
     merger_rate_density_param : `dict`
         Dictionary of merger rate density function parameters
-        default: None/dict(R0=25 * 1e-9, b2=1.6, b3=2.0, b4=30)
+        default: None/dict(R0=25 * 1e-9, b2=1.6, b3=2.1, b4=30)
     directory : `str`
         Directory to store the interpolator pickle files
         default: './interpolator_pickle'
@@ -115,7 +119,7 @@ class CBCSourceRedshiftDistribution(object):
     |                                     | rate density (PopI/PopII)        |
     |                                     | from Oguri et al. (2018)         |
     +-------------------------------------+----------------------------------+
-    |:meth:`~star_formation_rate_madau_dickinson2014`                        |
+    |:meth:`~sfr_madau_dickinson2014`                        |
     +-------------------------------------+----------------------------------+
     |                                     | Function to compute star         |
     |                                     | formation rate as given in       |
@@ -175,6 +179,7 @@ class CBCSourceRedshiftDistribution(object):
 
     def __init__(
         self,
+        npool=4,
         z_min=0.001,
         z_max=10.0,
         event_type="BBH",
@@ -186,6 +191,7 @@ class CBCSourceRedshiftDistribution(object):
     ):
         print("\nInitializing CBCSourceRedshiftDistribution...\n")
         # set attributes
+        self.npool = npool
         self.z_min = z_min
         self.z_max = z_max
         self.directory = directory
@@ -194,7 +200,7 @@ class CBCSourceRedshiftDistribution(object):
         self.cosmo = cosmology if cosmology else LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
 
         # setting up the interpolator creation parameters
-        self.create_new_interpolator = self.setup_decision_dictionary(create_new_interpolator)
+        self.create_new_interpolator = self.setup_decision_dictionary(create_new_interpolator, merger_rate_density)
         
         # creating of interpolators for redshift dependent quantities
         self.create_lookup_table()
@@ -204,11 +210,12 @@ class CBCSourceRedshiftDistribution(object):
         self.merger_rate_density = merger_rate_density # this is an initialization, not a variable assignment
 
         # source redshift distribution initialization
-        self.source_redshift = self.merger_rate_density_detector_frame(zs=None, get_attribute=True)
-        self.normalization_pdf_z = self.source_redshift.pdf_norm_const
+        self.merger_rate_density_detector_frame = self.merger_rate_density_detector_frame(zs=None, get_attribute=True)
[…]
