+++ ler/lens_galaxy_population/lens_param_data/number_density_zl_zs.txt
@@ -1,0 +1,48 @@
+0.000000000000000000e+00 1.184021318642246768e-10 1.517714157452121117e-10 1.893871169965377463e-10 2.286502377350245833e-10 2.756404056040109914e-10 3.172028372609239429e-10 3.653882265381016822e-10 4.006362515307286579e-10 4.506372940650550242e-10 5.025140156154955322e-10 5.431608191167116715e-10 5.777533491877313117e-10 6.200681704162470304e-10 6.700322941805261246e-10 7.050179387760153017e-10 7.292195562487669361e-10 7.680435928638710589e-10 7.943790102350091834e-10 8.273123601517104846e-10 8.379732450198277600e-10 8.618237396912116583e-10 8.707296365643810128e-10 8.979996437164168715e-10 9.053751847771496701e-10 8.961279536787109850e-10 8.819758504274547932e-10 9.014723049047749895e-10 8.759735952701586499e-10 8.700809978377016608e-10 8.514585942656093251e-10 8.309974899781962540e-10 8.194003185048257154e-10 7.996180196843310710e-10 7.607327831325739204e-10 7.319128504885263425e-10 6.944185115779021784e-10 6.632979101126287782e-10 6.163587971826539145e-10 5.802202620180069950e-10 5.287886543823089742e-10 5.035446839659991962e-10 4.523905359060897460e-10 4.082477214308089837e-10 3.613642522569752689e-10 3.109738832000095957e-10 2.748645240146644345e-10 2.286320363474559260e-10 1.895647623160928912e-10 1.522283229618994360e-10 1.166631678197521044e-10 0.000000000000000000e+00
+0.000000000000000000e+00 1.231484552415651498e-10 1.700533436717635288e-10 2.238776921139677482e-10 2.810212711916125563e-10 3.497859542456546171e-10 4.129986878482738669e-10 4.857134820376148517e-10 5.416835513672015767e-10 6.185788221884773966e-10 6.978240151492407256e-10 7.623753275550404194e-10 8.176893229279829669e-10 8.847691493098502043e-10 9.610337399560045922e-10 1.018301300828336277e-09 1.057945784832401826e-09 1.120183635885028986e-09 1.161761917582384334e-09 1.213341712474455006e-09 1.233037292618023990e-09 1.271228900651606590e-09 1.284235043048808585e-09 1.327847038099060957e-09 1.337624317262736300e-09 1.325286355374261877e-09 1.304541196701794836e-09 1.333706771364521298e-09 1.295242821494767392e-09 1.284286171450331148e-09 1.254852527177463257e-09 1.223239551140092148e-09 1.201959493619821001e-09 1.169927599420355431e-09 1.108521214172497354e-09 1.062431868173786499e-09 1.003405595298761584e-09 9.515695748538697787e-10 8.793502979924039715e-10 8.208749928836600566e-10 7.415748096651844581e-10 6.979820308869607990e-10 6.203642292428699430e-10 5.516976670519075441e-10 4.800246621387611278e-10 4.052873732742156004e-10 3.484035907448181079e-10 2.812776171684348052e-10 2.239133182529146807e-10 1.704991327914755123e-10 1.212624849412349532e-10 0.000000000000000000e+00
[…]

[…]
