+++ ler/lens_galaxy_population/lens_galaxy_parameter_distribution.py
@@ -11,6 +11,7 @@ import numpy as np
 from scipy.integrate import quad
 # from lenstronomy.Util.param_util import phi_q2_ellipticity
 from multiprocessing import Pool
+from tqdm import tqdm
 
 # the following .py file will be called if they are not given in the class initialization
 from ..gw_source_population import CBCSourceParameterDistribution
@@ -18,7 +19,7 @@ from .optical_depth import OpticalDepth
 from ..image_properties import ImageProperties
 from ..utils import add_dictionaries_together, trim_dictionary, FunctionConditioning, interpolator_pickle_path
 from .jit_functions import phi_cut_SIE, phi_q2_ellipticity_hemanta
-from .mp import cross_section_mp
+from .mp import cross_section_mp, rjs_sie_mp
 
 class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImageProperties, OpticalDepth):
     """
@@ -220,7 +221,7 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
         lens_param_samplers_params=None,
         directory="./interpolator_pickle",
         create_new_interpolator=False,
-        buffer_size=int(2e5),
+        buffer_size=1000,
         **kwargs  # for initialization of CBCSourceParameterDistribution class and ImageProperties class
     ):
         print("\nInitializing LensGalaxyParameterDistribution class...\n")          
@@ -245,19 +246,19 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
 
         # function to sample source redshifts conditioned on the source being strongly lensed 
         self.sample_source_redshift_sl = getattr(self,self.lens_param_samplers["source_redshift_sl"])
-        # create kde for strong lensing paramters: gamma, gamma1, gamma2
-        # this is done to speed up the sampling lens parameters
-        self.create_kde_sl();
+
+        # interpolated cross section function is not very accurate for rejection sampling
+        if self.lens_functions["cross_section"] == "interpolated_cross_section_function":
+            self.cross_section = self.cross_section_caustic_area
         # function to sample lens parameters
         self.sample_lens_parameters_routine = getattr(self, self.lens_functions['param_sampler_type'])  
         # function to rejection sample wrt lens cross section
         self.rejection_sample_sl = getattr(self, self.lens_functions['strong_lensing_condition']) 
-        
 
         # To find the normalization constant of the pdf p(z)
         # this under the assumption that the event is strongly lensed
         # Define the merger-rate density function
-        pdf_unnormalized_ = lambda z: self.merger_rate_density_detector_frame(np.array([z]), param=self.merger_rate_density_param) * self.optical_depth.function(np.array([z]))
+        pdf_unnormalized_ = lambda z: self.merger_rate_density_detector_frame(np.array([z])) * self.optical_depth.function(np.array([z]))
         pdf_unnormalized = lambda z: pdf_unnormalized_(z)[0]
 
         self.normalization_pdf_z_lensed = quad(
@@ -265,89 +266,6 @@ class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImagePrope
             self.z_min,
             self.z_max
         )[0]
-
-    def create_kde_sl(self):
-        """
-            Creates kernel density estimates (KDE) for strong lensing parameters: gamma, gamma1, and gamma2.
-            
-            This function checks if the necessary interpolator files exist for the parameters 
-            associated with strongly lensed gamma and gamma1, gamma2. If they do not exist, 
-            or if a new creation is requested, it generates the interpolators by sampling 
-            lens parameters under strong lensing conditions. The KDEs are computed to speed 
[…]
