+++ docs/autoapi/ler/gw_source_population/index.rst
@@ -36,31 +36,51 @@ Functions
 
 .. autoapisummary::
 
-   ler.gw_source_population.interpolator_from_pickle
-   ler.gw_source_population.cubic_spline_interpolator
-   ler.gw_source_population.inverse_transform_sampler
+   ler.gw_source_population.interpolator_pickle_path
    ler.gw_source_population.merger_rate_density_bbh_popI_II_oguri2018
-   ler.gw_source_population.star_formation_rate_madau_dickinson2014
+   ler.gw_source_population.sfr_madau_dickinson2014
    ler.gw_source_population.merger_rate_density_bbh_popIII_ken2022
    ler.gw_source_population.merger_rate_density_bbh_primordial_ken2022
-   ler.gw_source_population.interpolator_from_pickle
-   ler.gw_source_population.cubic_spline_interpolator
+   ler.gw_source_population.sfr_with_time_delay
    ler.gw_source_population.lognormal_distribution_2D
    ler.gw_source_population.inverse_transform_sampler_m1m2
+   ler.gw_source_population.sample_powerlaw_gaussian_source_bbh_masses
+   ler.gw_source_population.sample_broken_powerlaw_nsbh_masses
    ler.gw_source_population.inverse_transform_sampler
+   ler.gw_source_population.sample_from_powerlaw_distribution
+   ler.gw_source_population.cumulative_trapezoid
    ler.gw_source_population.merger_rate_density_bbh_popI_II_oguri2018
    ler.gw_source_population.merger_rate_density_bbh_popIII_ken2022
-   ler.gw_source_population.star_formation_rate_madau_dickinson2014
+   ler.gw_source_population.sfr_madau_fragos2017_with_bbh_td
+   ler.gw_source_population.sfr_madau_dickinson2014_with_bbh_td
+   ler.gw_source_population.sfr_madau_fragos2017_with_bns_td
+   ler.gw_source_population.sfr_madau_dickinson2014_with_bns_td
+   ler.gw_source_population.sfr_madau_fragos2017
+   ler.gw_source_population.sfr_madau_dickinson2014
    ler.gw_source_population.merger_rate_density_bbh_primordial_ken2022
    ler.gw_source_population.lognormal_distribution_2D
    ler.gw_source_population.inverse_transform_sampler_m1m2
+   ler.gw_source_population.powerlaw_with_smoothing
+   ler.gw_source_population.sample_broken_powerlaw
+   ler.gw_source_population.sample_broken_powerlaw_nsbh_masses
+   ler.gw_source_population.broken_powerlaw_pdf
+   ler.gw_source_population.broken_powerlaw_unormalized
+   ler.gw_source_population.powerlaw_B
+   ler.gw_source_population.gaussian_G
+   ler.gw_source_population.powerlaw_gaussian_pdf
+   ler.gw_source_population.powerlaw_gaussian_cdf
+   ler.gw_source_population.sample_powerlaw_gaussian
+   ler.gw_source_population.sample_powerlaw_gaussian_source_bbh_masses
+   ler.gw_source_population.powerlaw_gaussian_unnormalized
+   ler.gw_source_population.sfr_madau_fragos2017
+   ler.gw_source_population.sfr_with_time_delay
 
 
 
-.. py:function:: interpolator_from_pickle(param_dict_given, directory, sub_directory, name, x, pdf_func=None, y=None, conditioned_y=None, dimension=1, category='pdf', create_new=False)
+.. py:function:: interpolator_pickle_path(param_dict_given, directory, sub_directory, interpolator_name)
 
    
-   Function to decide which interpolator to use.
+   Function to create the interpolator pickle file path.
 
 
    :Parameters:
@@ -74,34 +94,16 @@ Functions
        **sub_directory** : `str`
[…]
