+++ docs/autoapi/ler/gw_source_population/jit_functions/index.rst
@@ -20,16 +20,58 @@ Functions
 
 .. autoapisummary::
 
+   ler.gw_source_population.jit_functions.cumulative_trapezoid
    ler.gw_source_population.jit_functions.merger_rate_density_bbh_popI_II_oguri2018
    ler.gw_source_population.jit_functions.merger_rate_density_bbh_popIII_ken2022
-   ler.gw_source_population.jit_functions.star_formation_rate_madau_dickinson2014
+   ler.gw_source_population.jit_functions.sfr_madau_fragos2017_with_bbh_td
+   ler.gw_source_population.jit_functions.sfr_madau_dickinson2014_with_bbh_td
+   ler.gw_source_population.jit_functions.sfr_madau_fragos2017_with_bns_td
+   ler.gw_source_population.jit_functions.sfr_madau_dickinson2014_with_bns_td
+   ler.gw_source_population.jit_functions.sfr_madau_fragos2017
+   ler.gw_source_population.jit_functions.sfr_madau_dickinson2014
    ler.gw_source_population.jit_functions.merger_rate_density_bbh_primordial_ken2022
    ler.gw_source_population.jit_functions.lognormal_distribution_2D
    ler.gw_source_population.jit_functions.inverse_transform_sampler_m1m2
+   ler.gw_source_population.jit_functions.powerlaw_with_smoothing
+   ler.gw_source_population.jit_functions.inverse_transform_sampler
+   ler.gw_source_population.jit_functions.sample_broken_powerlaw
+   ler.gw_source_population.jit_functions.sample_broken_powerlaw_nsbh_masses
+   ler.gw_source_population.jit_functions.broken_powerlaw_pdf
+   ler.gw_source_population.jit_functions.broken_powerlaw_unormalized
+   ler.gw_source_population.jit_functions.powerlaw_B
+   ler.gw_source_population.jit_functions.gaussian_G
+   ler.gw_source_population.jit_functions.powerlaw_gaussian_pdf
+   ler.gw_source_population.jit_functions.powerlaw_gaussian_cdf
+   ler.gw_source_population.jit_functions.sample_powerlaw_gaussian
+   ler.gw_source_population.jit_functions.sample_powerlaw_gaussian_source_bbh_masses
+   ler.gw_source_population.jit_functions.powerlaw_gaussian_unnormalized
 
 
 
-.. py:function:: merger_rate_density_bbh_popI_II_oguri2018(zs, R0=23.9 * 1e-09, b2=1.6, b3=2.0, b4=30)
+.. py:function:: cumulative_trapezoid(y, x=None, dx=1.0, initial=0.0)
+
+   
+   Compute the cumulative integral of a function using the trapezoidal rule.
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+   ..
+       !! processed by numpydoc !!
+
+.. py:function:: merger_rate_density_bbh_popI_II_oguri2018(zs, R0=23.9 * 1e-09, b2=1.6, b3=2.1, b4=30)
 
    
    Function to compute the merger rate density (PopI/PopII). Reference: Oguri et al. (2018). The output is in detector frame and is unnormalized.
@@ -50,7 +92,7 @@ Functions
 
        **b3** : `float`
            Fitting paramters
-           default: 2.0
+           default: 2.1
 
[…]
