+++ setup.py
@@ -27,19 +27,22 @@ setup(
       },
     python_requires='>=3.10',
     install_requires=[
-        "setuptools>=65.5.0",
-        "matplotlib>=3.4.2",
-        "pycbc>=2.0.4",
-        "numpy>=1.18",
-        "numba>=0.60.0",
-        "bilby>=1.0.2",
-        "gwsnr>=0.3.2",
-        "scipy<1.14.0",
-        "lenstronomy>=1.10.4",
-        "astropy>=5.1",
-        "tqdm>=4.64.1",
-        "pointpats>=2.3",
-        "shapely>=2.0.1",
-        "gwcosmo==2.1.0",
+        # "setuptools>=65.5.0",
+        # "matplotlib>=3.4.2",
+        # "pycbc>=2.0.4",
+        # "numpy>=1.18",
+        # "numba>=0.60.0",
+        # "bilby>=1.0.2",
+        # "gwsnr>=0.3.2",
+        # "scipy<1.14.0",
+        # "lenstronomy>=1.10.4",
+        # "astropy>=5.1",
+        # "tqdm>=4.64.1",
+        # "pointpats>=2.3",
+        # "shapely>=2.0.1",
+        # "gwcosmo==2.1.0",
+        "lenstronomy",
+        "pointpats",
+        "shapely",
     ],
 )
