+++ ler/gw_source_population/jit_functions.py
@@ -11,7 +11,23 @@ cosmo = LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
 from scipy.integrate import quad
 from scipy.optimize import fsolve
 
-from ler.utils import inverse_transform_sampler
+from ..utils import inverse_transform_sampler, sample_from_powerlaw_distribution
+
+@njit
+def cumulative_trapezoid(y, x=None, dx=1.0, initial=0.0):
+    """
+    Compute the cumulative integral of a function using the trapezoidal rule.
+    """
+    if x is None:
+        x = np.arange(len(y)) * dx
+
+    # Calculate the cumulative integral using trapezoidal rule
+    cumsum = np.zeros_like(y)
+    cumsum[0] = initial
+    for i in range(1, len(y)):
+        cumsum[i] = cumsum[i - 1] + (y[i - 1] + y[i]) * (x[i] - x[i - 1]) / 2.0
+
+    return cumsum
 
 # import pickle
 # # call the interpolator
@@ -397,6 +413,260 @@ def bns_bimodal_pdf(m, w=0.643, muL=1.352, sigmaL=0.08, muR=1.88, sigmaR=0.3, mm
 
     return pdf
 
+@njit
+def smoothing_S(m, mmin, delta_m, threshold=709.0):
+    s = np.zeros_like(m)
+
+    # Region where smoothing is not needed: m >= mmin + delta_m
+    idx_2 = m >= mmin + delta_m
+    s[idx_2] = 1.0
+
+    # Region where smoothing is applied: mmin <= m < mmin + delta_m
+    idx_1 = (m >= mmin) & (m < mmin + delta_m)
+    mprime = m[idx_1] - mmin
+    exponent = delta_m / mprime + delta_m / (mprime - delta_m)
+
+    # Safe exponentiation only where the exponent is below threshold
+    safe_idx = exponent <= threshold
+    s_vals = np.zeros_like(mprime)
+    s_vals[safe_idx] = 1.0 / (np.exp(exponent[safe_idx]) + 1.0)
+
+    # Assign back to main array
+    s[idx_1] = s_vals
+
+    return s
+
+@njit
+def powerlaw_with_smoothing(m, mmin, alpha, delta_m):
+    """
+    Power law with smoothing applied.
+    """
+    s = smoothing_S(m, mmin, delta_m)
+    return m ** (-alpha) * s
+
+@njit
+def inverse_transform_sampler(size, cdf, x):
+    """
+    Function to sample from the inverse transform method.
+    """
+
+    u = np.random.uniform(0, 1, size)
+    idx = np.searchsorted(cdf, u)
+    x1, x0, y1, y0 = cdf[idx], cdf[idx-1], x[idx], x[idx-1]
+    samples = y0 + (y1 - y0) * (u - x0) / (x1 - x0)
+    return samples
+
+@njit
+def broken_powerlaw_cdf(size=1000, mminbh=26,mmaxbh=125,alpha_1=6.75,alpha_2=0.0,b=0.5,delta_m=5):
+
+    # find normalization
+    m_try = np.linspace(mminbh, mmaxbh, size)
+    pdf_unnormalized = broken_powerlaw_unormalized(m_try, mminbh=26., mmaxbh=125., alpha_1=6.75, alpha_2=0., b=0.5, delta_m=5.)
+    # Compute the CDF using cumulative trapezoid integration
+    cdf_values = cumulative_trapezoid(y=pdf_unnormalized, x=m_try, dx=1.0, initial=0.0)
+    # Normalize the CDF
+    normalization = cdf_values[size-1]
+    # Normalize the CDF
+    cdf_values /= normalization
+
+    return cdf_values
+
+@njit
[…]
