+++ docs/index.rst
@@ -24,56 +24,70 @@ ler
 .. Yiwei Li (Lisa).
 
 
-.. _glossary:
-
 Glossary
-========
+==================
 
-Gravitational waves
+.. glossary::
+ 
+   Gravitational waves
 
-.. image:: _static/gw.gif
-   :width: 480px
-   :alt: Animation showing the propagation of gravitational waves
+      .. image:: _static/gw.gif
+         :align: left
+         :width: 480px
+         :alt: ler logo
 
-*Animation showing the propagation of gravitational waves from inspiraling binary black holes.*  
-`Source <https://community.wolfram.com/groups/-/m/t/790989>`_ : Jeffrey Bryant, Wolfram|Alpha, LLC.
+      *Animation showing the propagation of gravitational waves from inspiraling binary black holes.* 
+      `Source <https://community.wolfram.com/groups/-/m/t/790989>`_ : Jeffrey Bryant, Wolfram|Alpha, LLC.
 
-Ripples in the fabric of space-time caused by some of the most violent and energetic processes in the Universe, such as mergers of compact binaries (e.g., black holes and neutron stars) and supernova explosions. Albert Einstein predicted the existence of gravitational waves in 1916 in his general theory of relativity, but it took a century to detect them directly. The first detection was made in 2015 by the LIGO and Virgo collaborations, which won the 2017 Nobel Prize in Physics. Gravitational waves are invisible, yet incredibly fast; they travel at the speed of light, squeezing and stretching anything in their path.
+      |
+      Ripples in the fabric of space-time caused by some of the most violent and energetic processes in the Universe, such as merger of compact binaries (e.g., black holes and neutron stars) and supernovae explosions. Albert Einstein predicted the existence of gravitational waves in 1916 in his general theory of relativity, but it took a century to detect them directly. The first detection was made in 2015 by the LIGO and Virgo collaborations, which won the 2017 Nobel Prize in Physics. Gravitational waves are invisible, yet incredibly fast; they travel at the speed of light, squeezing and stretching anything in their path.
+      
+      |
 
-Lensing of gravitational waves
+   Lensing of gravitational waves
 
-.. image:: _static/Light_Bending.gif
-   :width: 480px
-   :alt: Animation showing the bending of light by a massive object
+      .. raw:: html
 
-*Animation showing the bending of light by a massive object.*  
-`Source <https://science.nasa.gov/universe/how-gravity-warps-light/>`_ : NASA, ESA, and Goddard Space Flight Center/K. Jackson.
+         <div style="text-align: center;">
+         <iframe src="_static/gwlensing.html"
+            width="90%"
+            height="500"
+            frameborder="0"></iframe>
+         </div>
 
-A process similar to the gravitational lensing of light, where gravitational waves emitted from distant astrophysical events are bent and split (strong-lensing case) into multiple images by the gravity of intervening massive objects, such as galaxies and galaxy clusters. This can magnify and change the arrival time of the gravitational waves.
+      *Interactive animation showing the lensing of gravitational waves by a massive object.*
 
-.. raw:: html
+      |
+      A process similar to the gravitational lensing of light, where gravitational waves emitted from distant astrophysical events are bent and splitted (strong-lensing case) into multiple images by the gravity of intervening massive objects, such as galaxy and galaxy cluster. This can magnify and change the arrival time of the gravitational waves.
 
-   <iframe src="_static/gwlensing.html"
-       width="100%"
-       height="600"
-       frameborder="0"
-       allowfullscreen
-       style="border:1px solid #ccc; border-radius:10px;"></iframe>
+      |
+      .. image:: _static/Light_Bending.gif
+         :align: right
+         :width: 480px
[…]
