+++ ler/utils/test.py
@@ -10,9 +10,8 @@ import numpy as np
 import json
 from scipy.interpolate import interp1d
 from scipy.interpolate import CubicSpline
-from scipy.integrate import quad
+from scipy.integrate import quad, cumtrapz
 from numba import njit
-from importlib import resources
 # import datetime
 
 
@@ -140,20 +139,9 @@ def save_json(file_name, param):
         json file name for storing the parameters.
     param : `dict`
         dictionary to be saved as a json file.
-
-    Example
-    ----------
-    >>> from ler.utils import save_json, load_json
-    >>> param = dict_list = [{"name": "lens_mass", "type": "cubic_spline", "x": np.array([0, 1, 2, 3]), "y": [0, 1, 0, 1]},]
-    >>> save_json("param.json", param)
-    >>> param = load_json("param.json")
     """
-    try:
-        with open(file_name, "w", encoding="utf-8") as write_file:
-            json.dump(param, write_file)
-    except:
-        with open(file_name, "w", encoding="utf-8") as write_file:
-            json.dump(param, write_file, indent=4, cls=NumpyEncoder)
+    with open(file_name, "w", encoding="utf-8") as write_file:
+        json.dump(param, write_file)
 
 def append_json(file_name, new_dictionary, old_dictionary=None, replace=False):
     """
@@ -258,28 +246,6 @@ def get_param_from_json(json_file):
         param[key] = np.array(value)
     return param
 
-def load_txt_from_module(package, directory, filename):
-    """
-    Function to load a specific dataset from a .txt file within the package
-
-    Parameters
-    ----------
-    package : str
-        name of the package
-    directory : str
-        name of the directory within the package
-    filename : str
-        name of the .txt file
-
-    Returns
-    ----------
-    data : `dict`
-        Dictionary loaded from the .txt file
-    """
-
-    with resources.path(package + '.' + directory, filename) as txt_path:
-        return np.loadtxt(txt_path)
-        
 def rejection_sample(pdf, xmin, xmax, size=100, chunk_size=10000):
     """
     Helper function for rejection sampling from a pdf with maximum and minimum arguments.
@@ -521,7 +487,7 @@ def create_func_pdf_invcdf(x, y, category="function"):
         pdf = interp1d(x, y, kind="cubic", fill_value="extrapolate")
         return pdf
     # cdf
-    cdf_values = cumulative_trapezoid(y, x, initial=0)
+    cdf_values = cumtrapz(y, x, initial=0)
     idx = np.argwhere(cdf_values > 0)[0][0]
     cdf_values = cdf_values[idx:]
     x = x[idx:]
@@ -647,7 +613,7 @@ def create_inv_cdf_array(x, y):
     idx = np.argwhere(np.isnan(y))
     x = np.delete(x, idx)
     y = np.delete(y, idx)
-    cdf_values = cumulative_trapezoid(y, x, initial=0)
+    cdf_values = cumtrapz(y, x, initial=0)
     cdf_values = cdf_values / cdf_values[-1]
     # to remove duplicate values on x-axis before interpolation
     # idx = np.argwhere(cdf_values > 0)[0][0]
@@ -702,9 +668,8 @@ def create_conditioned_inv_cdf_array(x, conditioned_y, pdf_func):
     """
 
     list_ = []
-    size = len(x)
     for y in conditioned_y:
-        phi = pdf_func(x,y*np.ones(size))
+        phi = pdf_func(x,y)
[…]
