+++ .github/workflows/documentation.yml
@@ -1,0 +1,73 @@
+# Simple workflow for deploying Sphinx documentation to GitHub Pages
+name: Deploy Sphinx Docs to Pages
+
+on:
+  # Runs on pushes targeting the default branch
+  push:
+    branches: ["main"] # Or "master", depending on your repo
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
+# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
+# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
+concurrency:
+  group: "pages"
+  cancel-in-progress: false
+
+jobs:
+  # Build job
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+      - name: Install system dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y pandoc
+      - name: Install Python dependencies
+        run: |
+          pip install -r requirements_sphinx.txt  # Or pip install sphinx sphinx-rtd-theme
+      - name: Install project dependencies
+        run: |
+          pip install -e .
+      - name: Build Sphinx documentation
+        run: |
+          cd docs
+          rm -r autoapi/*
+          make clean
+          make html
+      - name: List build directory contents
+        run: |
+          ls -la docs/
+          ls -la docs/_build/ || echo "_build directory not found"
+          ls -la docs/_build/html/ || echo "html directory not found"
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v3
+        with:
+          # Upload entire directory
+          path: './docs/_build/html' # Correct path based on Makefile
+  
+
+  # Deployment job
+  deploy:
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    runs-on: ubuntu-latest
+    needs: build
+    steps:
+      - name: Deploy to GitHub Pages
+        id: deployment
+        uses: actions/deploy-pages@v4
