+++ requirements_sphinx.txt
@@ -1,0 +1,30 @@
+html5lib==1.0b8
+setuptools
+nbsphinx==0.9.3
+sphinx-autoapi==3.0.0
+autoapi==2.0.1
+nbsphinx==0.9.3
+sphinx-copybutton==0.5.2
+sphinx-rtd-theme==3.0.2
+numpydoc==1.6.0
+sphinx-tabs==3.4.1
+sphinx-jupyterbook-latex==1.0.0
+myst-parser==3.0.1
+linkify-it-py==2.0.3
+sphinx-rtd-dark-mode==1.3.0
+
+# html5lib
+# setuptools
+# sphinx
+# sphinx-autoapi
+# autoapi
+# nbsphinx
+# sphinx-copybutton
+# sphinx-rtd-theme
+# numpydoc
+# sphinx-tabs
+# sphinx-jupyterbook-latex
+# myst-parser
+# linkify-it-py
+# sphinx-rtd-dark-mode
+# # npm install -g @mermaid-js/mermaid-cli
