+++ ler/image_properties/image_properties.py
@@ -368,7 +368,7 @@ class ImageProperties():
         -------
             snrs : `dict`
                 signal to noise ratio for each image in each event.
-                (dictionary containing 'H1', 'L1', ..., and 'optimal_snr_net', which is the network snr, for each image as an array with dimensions (number_of_lensed_events,n_max_images) )
+                (dictionary containing 'H1', 'L1', ..., and 'snr_net', which is the network snr, for each image as an array with dimensions (number_of_lensed_events,n_max_images) )
 
         """
         # needed to calculate effective luminosity distance and effective time delay
@@ -412,7 +412,7 @@ class ImageProperties():
         # setting up snr dictionary
         result_dict = dict()
         if snr_calculator:
-            result_dict["optimal_snr_net"] = (
+            result_dict["snr_net"] = (
                 np.ones((number_of_lensed_events, n_max_images)) * np.nan
             )
         # setting up pdet dictionary
@@ -518,7 +518,7 @@ class ImageProperties():
                         ),
                         output_jsonfile=False,
                     )
-                    result_dict["optimal_snr_net"][idx, i] = optimal_snr["optimal_snr_net"]
+                    result_dict["snr_net"][idx, i] = optimal_snr["snr_net"]
 
                     if list_of_detectors:
                         for detector in list_of_detectors:
