+++ ler/rates/ler.py
@@ -4,6 +4,7 @@ This module contains the main class for calculating the rates of detectable grav
 """
 
 import os
+# os.environ['OMP_NESTED'] = 'FALSE'
 import warnings
 warnings.filterwarnings("ignore")
 import logging
@@ -53,7 +54,7 @@ class LeR(LensGalaxyParameterDistribution):
         def snr_finder(gw_param_dict):
             ...
             return optimal_snr_dict
-        where optimal_snr_dict.keys = ['optimal_snr_net']. Refer to `gwsnr` package's GWSNR.snr attribute for more details.
+        where optimal_snr_dict.keys = ['snr_net']. Refer to `gwsnr` package's GWSNR.snr attribute for more details.
     pdet_finder : `function`
         default pdet_finder = None.
         The rate calculation uses either the pdet_finder or the snr_finder to calculate the detectable events. The custom pdet finder function should follow the following signature:
@@ -159,7 +160,7 @@ class LeR(LensGalaxyParameterDistribution):
     |:meth:`~class_initialization`        | Function to initialize the       |
     |                                     | parent classes                   |
     +-------------------------------------+----------------------------------+
-    |:meth:`~gwsnr_intialization`         | Function to initialize the       |
+    |:meth:`~gwsnr_initialization`         | Function to initialize the       |
     |                                     | gwsnr class                      |
     +-------------------------------------+----------------------------------+
     |:meth:`~snr`                         | Function to get the snr with the |
@@ -477,7 +478,7 @@ class LeR(LensGalaxyParameterDistribution):
             self.class_initialization(params=kwargs)
             # initialization self.snr and self.pdet from GWSNR class
             if not snr_finder and not pdet_finder:
-                self.gwsnr_intialization(params=kwargs)
+                self.gwsnr_initialization(params=kwargs)
                 self.gwsnr = True
                 self.pdet = pdet_finder
             else:
@@ -801,7 +802,7 @@ class LeR(LensGalaxyParameterDistribution):
         self.gw_param_sampler_dict["lens_functions"]=self.lens_functions.copy()
         self.gw_param_sampler_dict["lens_functions_params"]=self.lens_functions_params.copy()
 
-    def gwsnr_intialization(self, params=None):
+    def gwsnr_initialization(self, params=None):
         """
         Function to initialize the GWSNR class from the `gwsnr` package.
 
@@ -813,36 +814,54 @@ class LeR(LensGalaxyParameterDistribution):
         from gwsnr import GWSNR
 
         # initialization of GWSNR class
+        if 'mminbh' in self.gw_param_samplers_params['source_frame_masses']:
+            min_bh_mass = self.gw_param_samplers_params['source_frame_masses']['mminbh']
+        else:
+            min_bh_mass = 2.0
+
+        if 'mmaxbh' in self.gw_param_samplers_params['source_frame_masses']:
+            max_bh_mass = self.gw_param_samplers_params['source_frame_masses']['mmaxbh']
+        else:
+            max_bh_mass = 200.0
         input_params = dict(
+            # General settings
             npool=self.npool,
-            mtot_min=2*self.gw_param_samplers_params['source_frame_masses']['mminbh'],
-            mtot_max=2*self.gw_param_samplers_params['source_frame_masses']['mmaxbh']+10.0,
+            snr_method="interpolation_aligned_spins",
+            snr_type="optimal_snr",
+            gwsnr_verbose=True,
+            multiprocessing_verbose=True,
+            pdet_kwargs=None,
+            # Settings for interpolation grid
+            mtot_min=min_bh_mass*2,
+            mtot_max=max_bh_mass*2*(1+self.z_max) if max_bh_mass*2*(1+self.z_max)<500.0 else 500.0,
             ratio_min=0.1,
             ratio_max=1.0,
             spin_max=0.99,
[…]
