+++ requirements.txt
@@ -9,5 +9,5 @@ astropy>=5.1
 tqdm>=4.64.1
 pointpats>=2.3
 shapely>=2.0.1
-gwcosmo
+# gwcosmo
 #gwcosmo@git+https://git.ligo.org/lscsoft/gwcosmo.git@v1.0.0
