### 2025-10-30 #7420 @hemantaph
Hi @warrickball,

I've successfully re-submitted gwsnr (https://github.com/hemantaph/gwsnr/tree/main/gwsnr), which ler depends on.

I'm currently updating ler (as also can be seen in the recent updates) and expect to complete this within two weeks. As the reviewer requested, I've started on a test file for ler, and I'll be finishing that up as well.

Finally, I plan to update the associated paper with the reviewer's feedback in the coming days.
