+++ ler/utils/cosmological_coversions.py
@@ -1,0 +1,285 @@
+import numpy as np
+from numba import njit
+from .function_interpolation import FunctionConditioning
+# for redshift to luminosity distance conversion
+from astropy.cosmology import LambdaCDM
+
+def redshift_optimal_spacing(z_min, z_max, resolution):
+    n_geom = int(resolution * 0.6)  # 60% for geometric spacing (low z)
+    n_lin = resolution - n_geom  # 40% for linear spacing (high z)
+    # Transition point between geometric and linear
+    z_transition = z_min + (z_max - z_min) * 0.3
+    # Geometric spacing for low z
+    z_low = np.geomspace(z_min, z_transition, n_geom)
+    # Linear spacing for high z
+    z_high = np.linspace(z_transition, z_max, n_lin + 1)[
+        1:
+    ]  # Skip first point to avoid duplicate
+    # Combine arrays
+    zs = np.concatenate([z_low, z_high])
+    return zs
+
+def luminosity_distance(z=None, z_min=0.001, z_max=10., cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7), directory="./interpolator_json", create_new=False, resolution=500, get_attribute=True):
+    """
+    Function to create a lookup table for the luminosity distance wrt redshift.
+
+    Parameters
+    ----------
+    z : `numpy.ndarray` or `float`
+        Source redshifts
+    z_min : `float`
+        Minimum redshift of the source population
+    z_max : `float`
+        Maximum redshift of the source population
+
+    Attributes
+    ----------
+    z_to_luminosity_distance : `ler.utils.FunctionConditioning`
+        Object of FunctionConditioning class containing the luminosity distance wrt redshift
+    """
+
+    z_min = 0.001 if z_min == 0. else z_min
+
+    zs = redshift_optimal_spacing(z_min, z_max, resolution)
+    Dl = cosmo.luminosity_distance(zs).value
+    
+    luminosity_distance_object = FunctionConditioning(
+        function=Dl,
+        x_array=zs,
+        conditioned_y_array=None,
+        identifier_dict=dict(z_min=z_min, z_max=z_max, cosmology=cosmo, resolution=resolution, details="luminosity_distance from astropy.cosmology"),
+        directory=directory,
+        sub_directory="luminosity_distance",
+        name="luminosity_distance",
+        create_new=create_new,
+        create_function_inverse=True,
+        create_function=True,
+        create_pdf=False,
+        create_rvs=False,
+        callback='function',
+    )
+    luminosity_distance_object.__doc__ = """
+    Redshift to luminosity distance conversion.
+
+    Parameters
+    ----------
+    zs : `numpy.ndarray` or `float`
+        Source redshifts
+
+    Returns
+    ----------
+    luminosity_distance : `numpy.ndarray`
+        luminosity distance in Mpc
+
+    Examples
+    ----------
+    >>> from ler.gw_source_population import SourceGalaxyPopulationModel
+    >>> ler = SourceGalaxyPopulationModel()  # with default LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+    >>> luminosity_distance = ler.luminosity_distance(1.)
+    >>> luminosity_distance = ler.luminosity_distance.function(np.array([1., 2.]))
+    >>> redshift = ler.luminosity_distance.function_inverse(np.array([100., 200.]))
+    """
+
+    return luminosity_distance_object if get_attribute else luminosity_distance_object(z)
+
+
[…]
