+++ ler/lens_galaxy_population/__init__.py
@@ -2,5 +2,5 @@ from .lens_galaxy_parameter_distribution import *
 from .optical_depth import *
 from .mp import *
 from .jit_functions import *
-
-from .import lens_param_data
+from .cross_section_interpolator import *
+from .lens_parameter_sampler import *
