+++ ler/_version.py
@@ -1,0 +1,2 @@
+# Version information for ler package
+__version__ = "0.4.3"
