+++ ler/gw_source_population/cbc_source_parameter_distribution.py
@@ -62,7 +62,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         default: False
     directory : `str`
         Directory to store the interpolator pickle files
-        default: './interpolator_pickle'
+        default: './interpolator_json'
     create_new_interpolator : `dict`
         Dictionary of boolean values and resolution to create new interpolator.
         default: dict(redshift_distribution=dict(create_new=False, resolution=500), z_to_luminosity_distance=dict(create_new=False, resolution=500), differential_comoving_volume=dict(create_new=False, resolution=500))
@@ -247,9 +247,9 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         source_priors=None,
         source_priors_params=None,
         cosmology=None,
-        spin_zero=True,
+        spin_zero=False,
         spin_precession=False,
-        directory="./interpolator_pickle",
+        directory="./interpolator_json",
         create_new_interpolator=False,
     ):
         # set attributes
@@ -689,7 +689,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         mass_object = FunctionConditioning(
             function=None,
             x_array=None,
-            param_dict_given=identifier_dict,
+            identifier_dict=identifier_dict,
             directory=self.directory,
             sub_directory="source_frame_masses",
             name=identifier_dict['name'],
@@ -771,7 +771,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         mass_object = FunctionConditioning(
             function=None,
             x_array=None,
-            param_dict_given=identifier_dict,
+            identifier_dict=identifier_dict,
             directory=self.directory,
             sub_directory="source_frame_masses",
             name=identifier_dict['name'],
@@ -844,7 +844,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         mass_object = FunctionConditioning(
             function=None,
             x_array=None,
-            param_dict_given=identifier_dict,
+            identifier_dict=identifier_dict,
             directory=self.directory,
             sub_directory="source_frame_masses",
             name=identifier_dict['name'],
@@ -916,7 +916,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
     #     mass_object = FunctionConditioning(
     #         function=None,
     #         x_array=None,
-    #         param_dict_given=identifier_dict,
+    #         identifier_dict=identifier_dict,
     #         directory=self.directory,
     #         sub_directory="source_frame_masses",
     #         name=identifier_dict['name'],
@@ -1009,7 +1009,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         mass_object = FunctionConditioning(
             function=None,
             x_array=None,
-            param_dict_given=identifier_dict,
+            identifier_dict=identifier_dict,
             directory=self.directory,
             sub_directory="source_frame_masses",
             name=identifier_dict['name'],
@@ -1077,7 +1077,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         mass_object = FunctionConditioning(
             function=None,
             x_array=None,
-            param_dict_given=identifier_dict,
+            identifier_dict=identifier_dict,
             directory=self.directory,
             sub_directory="source_frame_masses",
             name=identifier_dict['name'],
@@ -1184,7 +1184,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
         mass_object = FunctionConditioning(
             function=model,
             x_array=mass,
-            param_dict_given=identifier_dict,
+            identifier_dict=identifier_dict,
             directory=self.directory,
             sub_directory="source_frame_masses",
[…]
