### 2026-01-06 6c3c3d1
docstrings update
