+++ docstring_template.py
@@ -1,0 +1,384 @@
+# -*- coding: utf-8 -*-
+"""
+Docstring Template based on LeR style conventions.
+
+This template follows NumPy/SciPy docstring conventions with reStructuredText 
+(RST) formatting for Sphinx documentation. It demonstrates the standard format 
+used throughout the ``ler`` package.
+
+Style Guidelines:
+    - Use double backticks (``type``) for type annotations in docstrings
+    - Use ``\n`` for explicit line breaks in RST output
+    - Use 4-column format for Instance Attributes tables (Attribute, Type, Unit, Description)
+    - Use 2-column format for Instance Methods tables (Method, Description)
+    - Private methods (prefixed with ``_``) should not appear in Instance Methods tables
+    - Properties should have docstrings with Returns section only (not Parameters)
+
+Copyright (C) 2026 Author Name. Distributed under MIT License.
+
+AI prompt to use this template: 
+1. Check for any inconsistencies and redundancies in the code (don't touch the docstrings). Please point them out without making any changes.
+2. Rewrite docstrings (including inline docstrings) of *.py using docstring_template.py. Convert some of the Instance Methods to private methods with underscore prefix and remove them from the class docstring. Use '\n' for newline and don't use '\\n'. Dont add unnecessary docstrings like the following:
+    # =============================================================================
+    # MODULE DOCSTRING TEMPLATE
+    # =============================================================================
+3. Create properties of the Instance Attributes if it is not there. Properties should be at the end.
+"""
+
+# =============================================================================
+# MODULE DOCSTRING TEMPLATE
+# =============================================================================
+"""
+Module for [brief description].
+
+[Extended description explaining the module's purpose, main classes/functions,
+and how it fits into the larger package structure.]
+
+Inheritance hierarchy (if applicable):
+
+- :class:`~package.ParentClass` \n
+  - :class:`~package.GrandparentClass` \n
+
+Usage:
+    Basic workflow example:
+
+    >>> from package import ClassName
+    >>> obj = ClassName()
+    >>> result = obj.some_method()
+
+Copyright (C) 2026 Author Name. Distributed under MIT License.
+"""
+
+
+# =============================================================================
+# CLASS DOCSTRING TEMPLATE
+# =============================================================================
+class TemplateClass:
+    """
+    One-line summary of the class.
+
+    Extended description of the class explaining its purpose, functionality,
+    and how it relates to other classes in the package. This should be 2-4
+    sentences that provide context.
+
+    Key Features: \n
+    - Feature one description \n
+    - Feature two description \n
+    - Feature three description \n
+
+    Parameters
+    ----------
+    param1 : ``int``
+        Description of param1. \n
+        default: 4
+    param2 : ``float``
+        Description of param2. \n
+        default: 0.0
+    param3 : ``str``
+        Description of param3 with options. \n
+        Options: \n
+        - 'option_a': Description of option A \n
+        - 'option_b': Description of option B \n
+        - 'option_c': Description of option C \n
+        default: 'option_a'
+    param4 : ``dict`` or ``None``
+        Description of dict parameter. \n
+        default: dict(key1="value1", key2="value2")
+    param5 : ``callable`` or ``None``
+        Description of callable parameter. \n
+        If None, uses default behavior. \n
+        The function should follow the signature: \n
+        ``def func(arg1, arg2): return result`` \n
+        default: None
+    **kwargs : ``dict``
+        Additional keyword arguments passed to parent classes: \n
+        :class:`~package.ParentClass1`, \n
+        :class:`~package.ParentClass2`.
+
+    Examples
+    --------
+    Basic usage:
+
+    >>> from package import TemplateClass
+    >>> obj = TemplateClass()
+    >>> result = obj.some_method()
+
+
+    Instance Methods
+    ----------
[…]
