+++ ler/lens_galaxy_population/optical_depth.py
@@ -155,9 +155,9 @@ class OpticalDepth:
         self.optical_depth = self.lens_functions["optical_depth"]
 
 
-    #-----------------
+    #-----------
     # Initialization 
-    #-----------------
+    #-----------
     def default_lens_samplers_and_functions(self, lens_type):
         """
         Function to categorize the lens priors/samplers
@@ -415,7 +415,7 @@ class OpticalDepth:
                 # check input sampler is string or function
                 if isinstance(sampler_name, str):
                     # available lens samplers for name e.g. 'axis_ratio'
-                    dict_ = self.available_lens_prior_list_and_its_params[
+                    dict_ = self.available_lens_samplers[
                         name
                     ]  # e.g. {'axis_ratio_padilla_strauss': {'q_min': 0.2, 'q_max': 1.0}, ....}
                     if sampler_name in dict_:  # e.g. 'axis_ratio_padilla_strauss'
@@ -448,7 +448,7 @@ class OpticalDepth:
                 function_name = lens_functions[name]
                 if isinstance(function_name, str):
                     # available lens functions for name e.g. 'optical_depth'
-                    dict_ = self.available_lens_functions_and_its_params[name]
+                    dict_ = self.available_lens_functions[name]
                     if function_name in dict_:
                         param_dict = dict_[function_name]
                         if (lens_functions_params is None) or (
@@ -512,7 +512,7 @@ class OpticalDepth:
         identifier_dict["sigma_resolution"] = self.create_new_interpolator["axis_ratio"][
             "sigma_resolution"
         ]
-        param_dict = self.available_lens_prior_list_and_its_params["axis_ratio"][
+        param_dict = self.available_lens_samplers["axis_ratio"][
             "axis_ratio_rayleigh"
         ]
         if param_dict:
@@ -586,7 +586,7 @@ class OpticalDepth:
         identifier_dict["resolution"] = self.create_new_interpolator["axis_ratio"][
             "resolution"
         ]
-        param_dict = self.available_lens_prior_list_and_its_params["axis_ratio"][
+        param_dict = self.available_lens_samplers["axis_ratio"][
             "axis_ratio_padilla_strauss"
         ]
         if param_dict:
@@ -769,7 +769,7 @@ class OpticalDepth:
                 if key not in name_list:
                     identifier_dict["external_shear"][key] = value
 
-        param_dict = self.available_lens_prior_list_and_its_params["lens_redshift"][
+        param_dict = self.available_lens_samplers["lens_redshift"][
             "lens_redshift_strongly_lensed_numerical"
         ]
         if param_dict:
@@ -888,12 +888,18 @@ class OpticalDepth:
         print("Checking if the function can be JIT compiled")
         _, use_njit_sampler = self._lens_redshift_multithreaded_njit(np.array([zl_scaled[0]]), np.array([zs[0]]))
 
-        # if use_njit_sampler:
-        print("Using multithreaded njit")
-        density_array, _ = self._lens_redshift_multithreaded_njit(zl_scaled, zs)
-        # else:
-        # print("Using multiprocessing")
-        # density_array = self._lens_redshift_multiprocessing(zl_scaled, zs)
+        cross_section_list = ['cross_section_epl_shear_interpolation', 'cross_section_sie_feixu', 'cross_section_sis']
+
+        if use_njit_sampler and (self.lens_functions['cross_section'] in cross_section_list):
+            print("Using multithreaded njit")
+            density_array, _ = self._lens_redshift_multithreaded_njit(zl_scaled, zs)
+        elif use_njit_sampler and (self.lens_functions['cross_section']=='cross_section_epl_shear_numerical'):
+            print("Using multiprocessing")
[…]

[…]
