### 2026-01-06 438fd40
docstrings update
