+++ ler/gw_source_population/sfr_with_time_delay.py
@@ -1,51 +1,65 @@
+# -*- coding: utf-8 -*-
+"""
+Module for computing star formation rates with time delays.
+
+This module provides functions for computing the star formation rate at a given
+redshift, accounting for time delays between formation and observation. The time
+delay distribution follows a 1/t power-law form, and the formation redshift is
+computed using the cosmological age-redshift relation.
+
+Key Features: \n
+- Time-delayed star formation rate computation \n
+- Integration with Madau & Fragos (2017) SFR model \n
+- Monte Carlo integration for time delay averaging \n
+- Cosmological calculations using Astropy \n
+
+Copyright (C) 2026 Hemanta Ph. Distributed under MIT License.
+"""
+
 import numpy as np
 from scipy.integrate import quad
 from scipy.optimize import fsolve
 from astropy.cosmology import LambdaCDM
-cosmo = LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
-# from scipy.interpolate import interp1d
+
 from .jit_functions import sfr_madau_fragos2017
+
 
 def sfr_with_time_delay(input_args):
     """
-    Compute the star formation rate at redshift z, given parameters a, b, c, and d, 
-    and cosmological parameters H0, Omega_M, and Omega_Lambda. 
-    The star formation rate is time-delayed relative to the observed redshift, 
-    with a time delay uniformly distributed between td_min and td_max. 
-    The time delay is computed using the cosmology provided by astropy.
+    Compute star formation rate at observed redshift with time delay.
+
+    The star formation rate is time-delayed relative to the observed redshift,
+    with a time delay uniformly distributed between td_min and td_max. The
+    formation redshift is computed using the cosmological age-redshift relation.
 
     Parameters
     ----------
-    input_args : list
-        z : float
-            observed redshift
-        idx : int
-            index of the galaxy
-        td_min : float
-            minimum time delay in Gyr
-        td_max : float
-            maximum time delay in Gyr
-        H0 : float
-            Hubble constant in km/s/Mpc
-        Omega_M : float
-            matter density parameter
-        Omega_Lambda : float
-            dark energy density parameter
-        a : float
-            parameter of the Madau-Fragos star formation rate
-        b : float
-            parameter of the Madau-Fragos star formation rate
-        c : float
-            parameter of the Madau-Fragos star formation rate
-        d : float
-            parameter of the Madau-Fragos star formation rate
+    input_args : ``list``
+        List containing the following elements in order: \n
+        - z (``float``): Observed redshift \n
+        - idx (``int``): Index identifier for the computation \n
+        - td_min (``float``): Minimum time delay (Gyr) \n
+        - td_max (``float``): Maximum time delay (Gyr) \n
+        - H0 (``float``): Hubble constant (km/s/Mpc) \n
+        - Omega_M (``float``): Matter density parameter \n
+        - Omega_Lambda (``float``): Dark energy density parameter \n
+        - a (``float``): Madau-Fragos SFR normalization parameter \n
+        - b (``float``): Madau-Fragos low-z power-law slope \n
+        - c (``float``): Madau-Fragos turnover parameter \n
+        - d (``float``): Madau-Fragos high-z power-law slope \n
 
     Returns
     -------
-    idx : int
-        index of the galaxy
-    result : float
-        star formation rate at observed redshift z
+    idx : ``int``
+        Index identifier (same as input).
+    result : ``float``
+        Time-averaged star formation rate at observed redshift z.
[…]
