+++ ler/rates/gwrates.py
@@ -87,76 +87,76 @@ class GWRATES(CBCSourceParameterDistribution):
     Instance Attributes
     ----------
     LeR class has the following attributes:\n
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     | Atrributes                          | Type                             |
     +=====================================+==================================+
     |:attr:`~npool`                       | `int`                            |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~z_min`                       | `float`                          |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~z_max`                       | `float`                          |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~event_type`                  | `str`                            |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~cosmo`                       | `astropy.cosmology`              |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~size`                        | `int`                            |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~batch_size`                  | `int`                            |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~json_file_names`             | `dict`                           |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~interpolator_directory`      | `str`                            |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~ler_directory`               | `str`                            |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~gwsnr`                       | `bool`                           |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~gw_param_sampler_dict`       | `dict`                           |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~snr_calculator_dict`         | `dict`                           |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~gw_param`                    | `dict`                           |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:attr:`~gw_param_detectable`         | `dict`                           |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
 
     Instance Methods
     ----------
     LeR class has the following methods:\n
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     | Methods                             | Description                      |
     +=====================================+==================================+
     |:meth:`~class_initialization`        | Function to initialize the       |
     |                                     | parent classes                   |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:meth:`~gwsnr_initialization`         | Function to initialize the       |
     |                                     | gwsnr class                      |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:meth:`~snr`                         | Function to get the snr with the |
     |                                     | given parameters.                |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:meth:`~store_gwrates_params`        | Function to store the all the    |
     |                                     | necessary parameters.            |
-    +-------------------------------------+----------------------------------+
+    +-------------------------+----------------------+
     |:meth:`~gw_cbc_statistics`           | Function to generate gw          |
     |                                     | GW source parameters.            |
[…]
