+++ docs/autoapi/ler/lens_galaxy_population/epl_shear_cross_section/index.rst
@@ -1,5 +1,3 @@
-:orphan:
-
 :py:mod:`ler.lens_galaxy_population.epl_shear_cross_section`
 ============================================================
 

[…]
