### 2026-01-08 95a8ce4
uadate docstrings and lerlogo
