+++ docs/autoapi/ler/image_properties/index.rst
@@ -22,7 +22,6 @@ Classes
 
 .. autoapisummary::
 
-   ler.image_properties.FunctionConditioning
    ler.image_properties.ImageProperties
 
 
@@ -34,7 +33,6 @@ Functions
 
    ler.image_properties.solve_lens_equation
    ler.image_properties.phi_q2_ellipticity_hemanta
-   ler.image_properties.luminosity_distance
    ler.image_properties.solve_lens_equation
 
 
@@ -45,6 +43,8 @@ Attributes
 .. autoapisummary::
 
    ler.image_properties.cosmo
+   ler.image_properties.MAX_RETRIES
+   ler.image_properties.MIN_MAGNIFICATION
 
 
 .. py:data:: cosmo
@@ -54,56 +54,71 @@ Attributes
 .. py:function:: solve_lens_equation(lens_parameters)
 
    
-   Function to solve the lens equation (min_image = 2)
+   Solve the lens equation to find image properties.
 
+   Uses the analytical solver from lenstronomy to find image positions,
+   magnifications, time delays, and hessian properties for strongly
+   lensed sources. Source positions are sampled from within the caustic
+   region to ensure multiple imaging.
 
    :Parameters:
 
-       **lens_parameters** : `list`
-           a list of parameters
-           lens_parameters[0] = min_images : minimum number of images
-           lens_parameters[1] = e1 : ellipticity
-           lens_parameters[2] = e2 : ellipticity
-           lens_parameters[3] = gamma : power-law index
-           lens_parameters[4] = gamma1 : shear
-           lens_parameters[5] = gamma2 : shear
-           lens_parameters[6] = zl : redshift of the lens
-           lens_parameters[7] = zs : redshift of the source
-           lens_parameters[8] = einstein_radius : Einstein radius
-           lens_parameters[9] = iteration : iteration number
-           lens_parameters[10:] = lens_model_list : numpy array of lens models
+       **lens_parameters** : ``numpy.ndarray``
+           Array of lens configuration parameters with the following structure:
+
+           - [0]: n_min_images - minimum number of images required
+
+           - [1]: e1 - ellipticity component 1
+
+           - [2]: e2 - ellipticity component 2
+
+           - [3]: gamma - power-law slope of mass density
+
+           - [4]: gamma1 - external shear component 1
+
+           - [5]: gamma2 - external shear component 2
+
+           - [6]: zl - lens redshift
+
+           - [7]: zs - source redshift
+
+           - [8]: einstein_radius - Einstein radius (units: arcsec)
+
+           - [9]: iteration - iteration index for tracking
+
+           - [10:]: lens_model_list - lens model names (e.g., 'EPL_NUMBA', 'SHEAR')
 
    :Returns:
 
-       **x_source** : `float`
-           x position of the source in the source plane, unit: arcsec
+       **x_source** : ``float``
+           Source x-position (units: arcsec).
 
-       **y_source** : `float`
-           y position of the source in the source plane
+       **y_source** : ``float``
+           Source y-position (units: arcsec).
 
-       **x0_image_position** : `float`
-           x position of the images in the source plane
+       **x0_image_position** : ``numpy.ndarray``
+           Image x-positions (units: arcsec).
 
-       **x1_image_position** : `float`
-           y position of the images in the source plane
+       **x1_image_position** : ``numpy.ndarray``
+           Image y-positions (units: arcsec).
 
-       **magnifications** : `float`
-           magnification of the images
+       **magnifications** : ``numpy.ndarray``
+           Magnification factors for each image.
 
-       **time_delays** : `float`
-           time-delay of the images
+       **time_delays** : ``numpy.ndarray``
+           Time delays for each image (units: seconds).
 
-       **nImages** : `int`
[…]
