+++ docs/autoapi/ler/image_properties/multiprocessing_routine/index.rst
@@ -5,7 +5,23 @@
 
 .. autoapi-nested-parse::
 
-   This sub-module contains the functions to solve the lens equation for a given set of lens parameters. The lens equation is solved using the analytical solver in lenstronomy. The functions in this sub-module are used in the multiprocessing routine to solve the lens equation for a given set of lens parameters.
+   Module for solving lens equations using multiprocessing.
+
+   This sub-module contains functions to solve the lens equation for a given set
+   of lens parameters. The lens equation is solved using the analytical solver in
+   lenstronomy. These functions are used in the multiprocessing routine within
+   the ImageProperties class.
+
+   Usage:
+       Basic workflow example:
+
+       >>> from ler.image_properties.multiprocessing_routine import solve_lens_equation
+       >>> import numpy as np
+       >>> from multiprocessing import Pool
+       >>> lens_params = np.array([2, 0.02, -0.01, 1.9, 0.1, 0.09, 0.25, 0.94, 1e-6, 0, 'EPL_NUMBA', 'SHEAR'], dtype=object)
+       >>> result = solve_lens_equation(lens_params)
+
+   Copyright (C) 2026 Phurailatpam Hemanta Kumar. Distributed under MIT License.
 
    ..
        !! processed by numpydoc !!
@@ -24,59 +40,93 @@ Functions
 
 
 
+Attributes
+~~~~~~~~~~
+
+.. autoapisummary::
+
+   ler.image_properties.multiprocessing_routine.MAX_RETRIES
+   ler.image_properties.multiprocessing_routine.MIN_MAGNIFICATION
+
+
+.. py:data:: MAX_RETRIES
+   :value: '100'
+
+   
+
+.. py:data:: MIN_MAGNIFICATION
+   :value: '0.01'
+
+   
+
 .. py:function:: solve_lens_equation(lens_parameters)
 
    
-   Function to solve the lens equation (min_image = 2)
+   Solve the lens equation to find image properties.
 
+   Uses the analytical solver from lenstronomy to find image positions,
+   magnifications, time delays, and hessian properties for strongly
+   lensed sources. Source positions are sampled from within the caustic
+   region to ensure multiple imaging.
 
    :Parameters:
 
-       **lens_parameters** : `list`
-           a list of parameters
-           lens_parameters[0] = min_images : minimum number of images
-           lens_parameters[1] = e1 : ellipticity
-           lens_parameters[2] = e2 : ellipticity
-           lens_parameters[3] = gamma : power-law index
-           lens_parameters[4] = gamma1 : shear
-           lens_parameters[5] = gamma2 : shear
-           lens_parameters[6] = zl : redshift of the lens
-           lens_parameters[7] = zs : redshift of the source
-           lens_parameters[8] = einstein_radius : Einstein radius
-           lens_parameters[9] = iteration : iteration number
-           lens_parameters[10:] = lens_model_list : numpy array of lens models
+       **lens_parameters** : ``numpy.ndarray``
+           Array of lens configuration parameters with the following structure:
+
+           - [0]: n_min_images - minimum number of images required
+
+           - [1]: e1 - ellipticity component 1
+
+           - [2]: e2 - ellipticity component 2
+
+           - [3]: gamma - power-law slope of mass density
+
+           - [4]: gamma1 - external shear component 1
+
+           - [5]: gamma2 - external shear component 2
+
+           - [6]: zl - lens redshift
+
+           - [7]: zs - source redshift
+
+           - [8]: einstein_radius - Einstein radius (units: arcsec)
+
+           - [9]: iteration - iteration index for tracking
+
[…]
