### 2026-01-08 4a717ff
uadate docstrings and lerlogo
