+++ docs/autoapi/ler/gw_source_population/index.rst
@@ -52,8 +52,6 @@ Functions
    ler.gw_source_population.sample_broken_powerlaw_nsbh_masses
    ler.gw_source_population.inverse_transform_sampler
    ler.gw_source_population.sample_from_powerlaw_distribution
-   ler.gw_source_population.cumulative_trapezoid
-   ler.gw_source_population.sample_source_redshift
    ler.gw_source_population.merger_rate_density_bbh_popI_II_oguri2018
    ler.gw_source_population.merger_rate_density_bbh_popIII_ken2022
    ler.gw_source_population.sfr_madau_fragos2017_with_bbh_td
@@ -65,24 +63,13 @@ Functions
    ler.gw_source_population.merger_rate_density_bbh_primordial_ken2022
    ler.gw_source_population.lognormal_distribution_2D
    ler.gw_source_population.inverse_transform_sampler_m1m2
-   ler.gw_source_population.erf
-   ler.gw_source_population.compute_normalization_factor
    ler.gw_source_population.bns_bimodal_pdf
-   ler.gw_source_population.smoothing_S
-   ler.gw_source_population.powerlaw_with_smoothing
-   ler.gw_source_population.broken_powerlaw_cdf
    ler.gw_source_population.sample_broken_powerlaw
    ler.gw_source_population.sample_broken_powerlaw_nsbh_masses
    ler.gw_source_population.broken_powerlaw_pdf
-   ler.gw_source_population.broken_powerlaw_unormalized
-   ler.gw_source_population.powerlaw_B
-   ler.gw_source_population.gaussian_G
    ler.gw_source_population.powerlaw_gaussian_pdf
-   ler.gw_source_population.powerlaw_gaussian_cdf
    ler.gw_source_population.sample_powerlaw_gaussian
    ler.gw_source_population.sample_powerlaw_gaussian_source_bbh_masses
-   ler.gw_source_population.sample_mass_ratio
-   ler.gw_source_population.powerlaw_gaussian_unnormalized
    ler.gw_source_population.sfr_madau_fragos2017
    ler.gw_source_population.sfr_with_time_delay
 
@@ -94,8 +81,6 @@ Attributes
 .. autoapisummary::
 
    ler.gw_source_population.chunk_size
-   ler.gw_source_population.cosmo
-   ler.gw_source_population.cosmo
 
 
 .. py:class:: FunctionConditioning(function=None, x_array=None, conditioned_y_array=None, y_array=None, non_zero_function=False, gaussian_kde=False, gaussian_kde_kwargs={}, identifier_dict={}, directory='./interpolator_json', sub_directory='default', name='default', create_new=False, create_function=False, create_function_inverse=False, create_pdf=False, create_rvs=False, multiprocessing_function=False, callback=None)
@@ -221,34 +206,40 @@ Attributes
 .. py:function:: merger_rate_density_bbh_popI_II_oguri2018(zs, R0=23.9 * 1e-09, b2=1.6, b3=2.1, b4=30)
 
    
-   Function to compute the merger rate density (PopI/PopII). Reference: Oguri et al. (2018). The output is in detector frame and is unnormalized.
+   Compute the merger rate density for PopI/II BBH.
 
+   Reference: Oguri et al. (2018). The output is in detector frame and is
+   unnormalized.
 
    :Parameters:
 
-       **zs** : `float` or `numpy.ndarray` (nD array of floats)
-           Source redshifts
+       **zs** : ``float`` or ``numpy.ndarray``
+           Source redshifts.
 
-       **R0** : `float`
-           local merger rate density at low redshift
-           default: 23.9*1e-9 Mpc^-3 yr^-1
+       **R0** : ``float``
+           Local merger rate density at low redshift (Mpc^-3 yr^-1).
 
[…]
