+++ docs/index.rst
@@ -4,7 +4,7 @@ Welcome to :red:`ler`'s documentation!
 .. raw:: html
 
    <div style="text-align: center; margin-top: 0px; margin-bottom: 10px; padding-top: 0px; padding-bottom: 10px;">
-      <img src="_static/logo.png" width="220rem" alt="ler logo">
+      <img src="_static/lerlogo.png" width="220rem" alt="ler logo">
    </div>
 
 :red:`ler` : :red_first:`LVK` (LIGO-Virgo-KAGRA) :red_first:`Event` (compact-binary mergers) :red_first:`Rate` calculator and simulator
@@ -124,7 +124,7 @@ Glossary
          <img src="_static/BBH_Merger_rate_density_and_PDF_of_redshift.png" width="480px" alt="BBH Merger rate density">
           <div style="text-align:left; max-width:480px; margin: 0 auto;">
             <p style="font-size: 0.9em; font-family: Arial, sans-serif; line-height: 1.5em;">
-               Figure showing the merger rate density of binary black hole (BBH) mergers as a function of redshift. <em>Source: <a href="https://ler.readthedocs.io/en/latest/GW_events.html">generated by the LeR package.</a>.</em>
+               Figure showing the merger rate density of binary black hole (BBH) mergers as a function of redshift. <em>Source: <a href="https://ler.hemantaph.comGW_events.html">generated by the LeR package.</a>.</em>
             </p>
          </div>
          </div>

[…]
