+++ ler/lens_galaxy_population/lens_galaxy_parameter_distribution.py
@@ -1,14 +1,26 @@
 """
-This module contains the ``LensGalaxyParameterDistribution`` class.
+Module for lens galaxy parameter distributions.
 
-The class is used to sample lens galaxy parameters and source parameters conditioned on
-the source being strongly lensed.
-It inherits from the ``ImageProperties`` class to calculate image properties (magnification,
-time delays, source position, image position, morse phase).
-It also inherits from ``CBCSourceParameterDistribution``.
+This module contains the ``LensGalaxyParameterDistribution`` class which samples
+lens galaxy parameters and source parameters conditioned on the source being
+strongly lensed. It handles the distribution of lens parameters such as velocity
+dispersion, axis ratio, rotation angle, shear, and density profile slope.
 
-Copyright (c) 2026 Phurailatpam Hemantakumar
-License: MIT
+Inheritance hierarchy:
+
+- :class:`~ler.gw_source_population.CBCSourceParameterDistribution` \n
+- :class:`~ler.image_properties.ImageProperties` \n
+- :class:`~ler.lens_galaxy_population.OpticalDepth` \n
+
+Usage:
+    Basic workflow example:
+
+    >>> from ler.lens_galaxy_population import LensGalaxyParameterDistribution
+    >>> lens = LensGalaxyParameterDistribution()
+    >>> lensed_params = lens.sample_lens_parameters(size=1000)
+    >>> print(lensed_params.keys())
+
+Copyright (C) 2026 Phurailatpam Hemantakumar. Distributed under MIT License.
 """
 
 import warnings
@@ -16,111 +28,134 @@ import numpy as np
 from scipy.integrate import quad
 from numba import njit
 
-# the following .py file will be called if they are not given in the class initialization
-from ..gw_source_population import CBCSourceParameterDistribution
 from .optical_depth import OpticalDepth
+from .sampler_functions import _njit_checks
+
+from ..gw_source_population import CBCSourceParameterDistribution
 from ..image_properties import ImageProperties
-from ..utils import is_njitted
 warnings.filterwarnings("ignore")
+
 
 class LensGalaxyParameterDistribution(CBCSourceParameterDistribution, ImageProperties, OpticalDepth):
     """
-    Class to sample lens galaxy parameters and source parameters conditioned on the source being strongly lensed.
+    Sample lens galaxy parameters conditioned on strong lensing.
 
-    This class deals with the distribution of lens galaxy parameters, such as velocity dispersion,
-    axis ratio, axis rotation angle, shear, and density profile slope. It also handles the
-    sampling of source parameters conditioned on the source being strongly lensed.
+    This class handles the distribution of lens galaxy parameters such as velocity
+    dispersion, axis ratio, axis rotation angle, shear, and density profile slope.
+    It samples source parameters conditioned on the source being strongly lensed
+    using cross-section based rejection or importance sampling.
+
+    Key Features: \n
+    - Samples lens parameters using EPL+shear galaxy model \n
+    - Supports rejection and importance sampling based on cross-section \n
+    - Computes optical depth weighted source redshift distributions \n
+    - Integrates with GW source population and image property calculations \n
 
     Parameters
     ----------
-    npool : int, optional
-        Number of processors to use.
-        Default is 4.
-    z_min : float, optional
-        Minimum redshift.
-        Default is 0.0.
-    z_max : float, optional
-        Maximum redshift.
-        Default is 10.0.
-    cosmology : astropy.cosmology, optional
-        Cosmology to use.
-        Default is None, which falls back to ``astropy.cosmology.FlatLambdaCDM(H0=70, Om0=0.3)``.
-    event_type : str, optional
[…]
