+++ ler/lens_galaxy_population/mp.py
@@ -1,18 +1,38 @@
-import numpy as np
-from scipy.integrate import quad
-from numba import njit, prange
-from lenstronomy.LensModel.Solver.epl_shear_solver import caustics_epl_shear
-from shapely.geometry import Polygon
+# -*- coding: utf-8 -*-
+"""
+Multiprocessing helper functions for lens galaxy population calculations.
 
-from .jit_functions import phi_cut_SIE, phi_q2_ellipticity_hemanta
-from ..utils import inverse_transform_sampler, cubic_spline_interpolator, cubic_spline_interpolator2d_array, inverse_transform_sampler2d, load_json
+This module provides parallelized functions for computing optical depth 
+and cross-sections for strong gravitational lensing. These functions are 
+designed to be used with Python's multiprocessing module for efficient 
+Monte Carlo integration over lens parameters.
+
+Key Features: \n
+- JIT-compiled parallel function for lens redshift sampling \n
+- Multiprocessing-compatible wrapper functions for optical depth \n
+- Cross-section calculation using EPL+Shear lens models \n
+
+Copyright (C) 2026 Phurailatpam Hemantakumar. Distributed under MIT License.
+"""
+
+import numpy as np
+from numba import njit, prange
+
+from .lens_functions import phi_cut_SIE, phi_q2_ellipticity, cross_section
+
+from ..utils import inverse_transform_sampler, load_pickle
 
 from .cross_section_interpolator import make_cross_section_reinit
 
+# Constants for cross-section unit scaling (1/pi and numerical offset)
+CS_UNIT_SLOPE = 0.31830988618379075
+CS_UNIT_INTERCEPT = -3.2311742677852644e-27
+
+
 @njit(parallel=True)
 def lens_redshift_strongly_lensed_njit(
-        zs_array, # 1D
-        zl_scaled, # 2D
+        zs_array,
+        zl_scaled,
         sigma_min,
         sigma_max,
         q_rvs,
@@ -24,542 +44,219 @@ def lens_redshift_strongly_lensed_njit(
         dVcdz_function,
         integration_size,
     ):
+    """
+    JIT-compiled parallel computation of lens redshift optical depth.
 
+    Computes the differential optical depth for strong lensing as a function 
+    of source and lens redshifts using Monte Carlo integration with parallel 
+    execution over source redshifts.
+
+    Parameters
+    ----------
+    zs_array : ``numpy.ndarray``
+        1D array of source redshifts.
+    zl_scaled : ``numpy.ndarray``
+        2D array of scaled lens redshifts (zl/zs).
+    sigma_min : ``float``
+        Minimum velocity dispersion (km/s).
+    sigma_max : ``float``
+        Maximum velocity dispersion (km/s).
+    q_rvs : ``callable``
+        Function to sample axis ratios given size and sigma.
+    phi_rvs : ``callable``
+        Function to sample axis rotation angles.
+    gamma_rvs : ``callable``
+        Function to sample density profile slopes.
+    shear_rvs : ``callable``
+        Function to sample external shear components (gamma1, gamma2).
+    number_density : ``callable``
+        Function to compute velocity dispersion number density.
+    cross_section : ``callable``
+        Function to compute lensing cross-section.
+    dVcdz_function : ``callable``
+        Function to compute differential comoving volume.
+    integration_size : ``int``
+        Number of Monte Carlo samples per (zs, zl) pair.
+
+    Returns
+    -------
+    result_array : ``numpy.ndarray``
+        2D array of optical depth values with shape (len(zs_array), len(zl_scaled[0])).
+    """
     size_zs = zs_array.shape[0]
     size_zl = zl_scaled.shape[1]
     result_array = np.zeros((size_zs, size_zl))
 
     for i in prange(size_zs):
-
[…]
