### 2026-01-15 7e957df
exmpale notebooks
