+++ paper/paper.md
@@ -68,7 +68,7 @@ bibliography: paper.bib
 
 Gravitational waves (GWs) are ripples in the fabric of space and time caused by the acceleration of unevenly distributed mass or masses. Observable GWs are created especially during the violent cosmic events of merging compact binaries, such as 'binary black holes' (BBHs), 'binary neutron stars' (BNSs), and 'neutron star and black hole pair' (NSBHs). GWs emitted by these events can be distorted or magnified by the gravitational fields of massive objects such as galaxies or galaxy clusters, a phenomenon known as gravitational lensing. Profound comprehension of gravitational lensing's impact on GW signals is imperative to their accurate interpretation and the extraction of astrophysical insights therein. For this purpose, statistical modelling of GWs lensing can provide valuable insights into the properties of the lensing objects and GW sources. Such statistics require accurate and efficient means to calculate the detectable lensing rates, which depend on up-to-date modeling and implementation of lens and source properties and their distribution. The outcomes of these computational analyses not only contribute to generating dependable forecasts but also play an important role in validating forthcoming lensing events [@Janquart2023] [@ligolensing2023] [@Abbott2021].
 
-Obtaining precise outcomes in statistical analyses of this nature necessitates the utilization of large-scale sampling, often numbering in the millions. However, this process is computationally demanding. The $ler$ framework addresses this by employing innovative techniques to optimize the workflow and computation efficiency required for handling large-scale statistical analyses, essential for modeling detectable events and calculating rates. Its integration of modular statistical components enhances the framework's adaptability and extendability, thus proving to be an invaluable asset in the evolving field of gravitational wave research. Detailed description, source code, and examples are available in $ler$ [documentation](https://ler.readthedocs.io/en/latest/).
+Obtaining precise outcomes in statistical analyses of this nature necessitates the utilization of large-scale sampling, often numbering in the millions. However, this process is computationally demanding. The $ler$ framework addresses this by employing innovative techniques to optimize the workflow and computation efficiency required for handling large-scale statistical analyses, essential for modeling detectable events and calculating rates. Its integration of modular statistical components enhances the framework's adaptability and extendability, thus proving to be an invaluable asset in the evolving field of gravitational wave research. Detailed description, source code, and examples are available in $ler$ [documentation](https://ler.hemantaph.com).
 
 # Statement of need
 
