+++ ler/lens_galaxy_population/sampler_functions.py
@@ -1,0 +1,1713 @@
+# -*- coding: utf-8 -*-
+"""
+Module for lens galaxy parameter sampling functions.
+
+This module provides probability density functions (PDFs) and random variable
+samplers (RVS) for lens galaxy parameters including redshift, velocity dispersion,
+and axis ratio. It also includes rejection and importance sampling algorithms
+for sampling lens parameters weighted by gravitational lensing cross sections.
+
+Key Components: \n
+- Lens redshift samplers (SIS model from Haris et al. 2018) \n
+- Velocity dispersion samplers (generalized gamma distribution) \n
+- Axis ratio samplers (Rayleigh and Padilla-Strauss distributions) \n
+- Rejection and importance sampling for cross-section weighted parameters \n
+
+Copyright (C) 2026 Hemantakumar Phurailatpam. Distributed under MIT License.
+"""
+
+import numpy as np
+from numba import njit, prange
+from scipy.interpolate import CubicSpline
+from astropy.cosmology import LambdaCDM
+from ..utils import is_njitted
+from multiprocessing import Pool
+from tqdm import tqdm
+from ..utils import (
+    save_pickle,
+    load_pickle,
+    inverse_transform_sampler,
+    redshift_optimal_spacing,
+)
+import os
+
+
+def available_sampler_list():
+    """
+    Return list of available lens parameter samplers.
+
+    Returns
+    -------
+    sampler_list : ``list``
+        List of available sampler function names.
+
+    Examples
+    --------
+    >>> samplers = available_sampler_list()
+    >>> print(samplers)
+    ['lens_redshift_strongly_lensed_sis_haris', 'velocity_dispersion_gengamma', 'axis_ratio_rayleigh', 'axis_ratio_padilla_strauss']
+    """
+    return [
+        "lens_redshift_strongly_lensed_sis_haris_pdf",
+        "lens_redshift_strongly_lensed_sis_haris_rvs",
+        "velocity_dispersion_ewoud_denisty_function",
+        "velocity_dispersion_bernardi_denisty_function",
+        "velocity_dispersion_gengamma_function",
+        "velocity_dispersion_gengamma_pdf",
+        "velocity_dispersion_gengamma_rvs",
+        "axis_ratio_rayleigh_rvs",
+        "axis_ratio_rayleigh_pdf",
+        "axis_ratio_padilla_strauss_rvs",
+        "axis_ratio_padilla_strauss_pdf",
+        "bounded_normal_sample",
+        "rejection_sampler",
+        "importance_sampler",
+        "importance_sampler_mp",
+    ]
+
+
+# ---------------------------------
+# Lens redshift sampler functions
+# ---------------------------------
+@njit
+def lens_redshift_strongly_lensed_sis_haris_pdf(
+    zl, zs, cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+):
+    """
+    Compute lens redshift PDF for SIS model (Haris et al. 2018).
+
+    Computes the probability density function for lens redshift between
+    zl=0 and zl=zs using the analytical form from Haris et al. (2018)
+    equation A7, based on the SIS (Singular Isothermal Sphere) lens model.
+
+    Parameters
+    ----------
+    zl : ``float``
+        Redshift of the lens galaxy.
+    zs : ``float``
+        Redshift of the source.
+    cosmo : ``astropy.cosmology``
+        Cosmology object for distance calculations. \n
+        default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+
+    Returns
+    -------
+    pdf : ``float``
[…]
