+++ ler/lens_galaxy_population/__init__.py
@@ -1,6 +1,6 @@
 from .lens_galaxy_parameter_distribution import *
 from .optical_depth import *
 from .mp import *
-from .jit_functions import *
+from .sampler_functions import *
+from .lens_functions import *
 from .cross_section_interpolator import *
-from .lens_parameter_sampler import *
