+++ ler/lens_galaxy_population/lens_parameter_sampler.py
@@ -1,310 +1,0 @@
-from numba import njit, prange
-import numpy as np
-C_LIGHT = 299792.458  # km/s
-
-# ----------------------
-# Rejection sampler
-# ----------------------
-def rejection_sampler(
-    zs, 
-    zl,
-    sigma_max,
-    sigma_rvs,
-    q_rvs,
-    phi_rvs,
-    gamma_rvs,
-    shear_rvs,
-    cross_section,
-    saftey_factor = 1.2
-):
-    size = zl.size
-
-    sigma_array = np.zeros(size)
-    q_array = np.zeros(size)
-    phi_array = np.zeros(size)
-    gamma_array = np.zeros(size)
-    gamma1_array = np.zeros(size)
-    gamma2_array = np.zeros(size)
-    idx_remaining = np.arange(size)
-
-    # Compute the maximum cross section
-    # to be used as the upper bound for the rejection sampler
-    cs_max = cross_section(
-                zs=zs,
-                zl=zl,
-                sigma=sigma_max*np.ones(size),
-                q=0.9*np.ones(size),
-                phi=np.zeros(size),
-                gamma=2.645*np.ones(size),
-                gamma1=np.zeros(size),
-                gamma2=np.zeros(size)
-              ) * saftey_factor
-
-    while len(idx_remaining) > 0:
-        n_remaining = len(idx_remaining)
-        sigma_samples = sigma_rvs(n_remaining, zl[idx_remaining])
-        q_samples = q_rvs(n_remaining, sigma_samples)
-        phi_samples = phi_rvs(n_remaining)
-        gamma_samples = gamma_rvs(n_remaining)
-        gamma1_samples, gamma2_samples = shear_rvs(n_remaining)
-
-        cs = cross_section(
-            zs=zs[idx_remaining],
-            zl=zl[idx_remaining],
-            sigma=sigma_samples,
-            q=q_samples,
-            phi=phi_samples,
-            gamma=gamma_samples,
-            gamma1=gamma1_samples,
-            gamma2=gamma2_samples
-        )
-
-        accept = np.random.random(n_remaining) < (cs / cs_max[idx_remaining])
-
-        accepted_indices = idx_remaining[accept]
-        sigma_array[accepted_indices] = sigma_samples[accept]
-        q_array[accepted_indices] = q_samples[accept]
-        phi_array[accepted_indices] = phi_samples[accept]
-        gamma_array[accepted_indices] = gamma_samples[accept]
-        gamma1_array[accepted_indices] = gamma1_samples[accept]
-        gamma2_array[accepted_indices] = gamma2_samples[accept]
-
-        idx_remaining = idx_remaining[~accept]
-
-    return sigma_array, q_array, phi_array, gamma_array, gamma1_array, gamma2_array
-
-def create_rejection_sampler(
-    sigma_max,
-    sigma_rvs,
-    q_rvs,
-    phi_rvs,
-    gamma_rvs,
-    shear_rvs,
-    cross_section,
-    saftey_factor=1.2,
-    use_njit_sampler=True
-):
-    
-    if use_njit_sampler:
-        _base_sampler = njit(rejection_sampler)
-        print("Faster, njitted and rejection sampling based lens parameter sampler will be used.")
-
-        @njit
-        def rejection_sampler_wrapper(
-            zs, zl,
-        ):
-            return _base_sampler(
-                zs=zs,
-                zl=zl,
-                sigma_max=sigma_max,
-                sigma_rvs=sigma_rvs,
-                q_rvs=q_rvs,
-                phi_rvs=phi_rvs,
-                gamma_rvs=gamma_rvs,
-                shear_rvs=shear_rvs,
-                cross_section=cross_section,
[…]
