+++ ler/gw_source_population/__init__.py
@@ -1,4 +1,4 @@
 from .cbc_source_redshift_distribution import *
 from .cbc_source_parameter_distribution import *
-from .jit_functions import *
+from .prior_functions import *
 from .sfr_with_time_delay import *

[…]
