### 2026-01-19 a050fb6
major updates: SL sampling, docs
