+++ ler/rates/gwrates.py
@@ -30,14 +30,22 @@ import warnings
 import pathlib
 import zipfile
 from importlib_resources import files as resources_files
+
 warnings.filterwarnings("ignore")
 import logging
-logging.getLogger('numexpr.utils').setLevel(logging.ERROR)
+
+logging.getLogger("numexpr.utils").setLevel(logging.ERROR)
 import contextlib
 import numpy as np
 from astropy.cosmology import LambdaCDM
 from ..gw_source_population import CBCSourceParameterDistribution
-from ..utils import load_json, append_json, get_param_from_json, batch_handler
+from ..utils import (
+    load_json,
+    append_json,
+    get_param_from_json,
+    batch_handler,
+    remove_file,
+)
 
 
 class GWRATES(CBCSourceParameterDistribution):
@@ -183,21 +191,25 @@ class GWRATES(CBCSourceParameterDistribution):
         # first check if the interpolator directory './interpolator_json' exists
         if not pathlib.Path(interpolator_directory).exists():
             # Get the path to the zip resource using importlib_resources
-            zip_resource = resources_files('ler.rates').joinpath('ler_data', 'interpolator_json.zip')
-            with zip_resource.open('rb') as zip_file:
-                print("Extracting interpolator data from package to the current working directory.")
-                
+            zip_resource = resources_files("ler.rates").joinpath(
+                "ler_data", "interpolator_json.zip"
+            )
+            with zip_resource.open("rb") as zip_file:
+                print(
+                    "Extracting interpolator data from package to the current working directory."
+                )
+
                 # Define destination path (current working directory)
                 dest_path = pathlib.Path.cwd()
 
                 # Extract the zip file, skipping __MACOSX metadata
-                with zipfile.ZipFile(zip_file, 'r') as zip_ref:
+                with zipfile.ZipFile(zip_file, "r") as zip_ref:
                     for member in zip_ref.namelist():
                         # Skip __MACOSX directory and its contents
-                        if member.startswith('__MACOSX'):
+                        if member.startswith("__MACOSX"):
                             continue
                         zip_ref.extract(member, dest_path)
-        
+
         print("\nInitializing GWRATES class...\n")
         # init gwrates attributes
         self.npool = npool
@@ -208,9 +220,9 @@ class GWRATES(CBCSourceParameterDistribution):
 
         # init json file names where datas will be stored
         self.json_file_names = dict(
-            gwrates_params="gwrates_params.json", 
-            gw_param="gw_param.json", 
-            gw_param_detectable="gw_param_detectable.json"
+            gwrates_params="gwrates_params.json",
+            gw_param="gw_param.json",
+            gw_param_detectable="gw_param_detectable.json",
         )
         if json_file_names:
             self.json_file_names.update(json_file_names)
@@ -225,7 +237,9 @@ class GWRATES(CBCSourceParameterDistribution):
             os.makedirs(ler_directory)
 
         # parent class initialization
-        self._parent_class_initialization(params=kwargs, pdet_finder=pdet_finder, verbose=verbose)
+        self._parent_class_initialization(
+            params=kwargs, pdet_finder=pdet_finder, verbose=verbose
+        )
 
     def _parent_class_initialization(self, params=None, pdet_finder=None, verbose=True):
         """
@@ -250,9 +264,11 @@ class GWRATES(CBCSourceParameterDistribution):
                 self.pdet_finder = self._gwsnr_initialization(params=params)
             else:
                 self.pdet_finder = pdet_finder
-            
+
             # store all the gwrates input parameters
[…]
