+++ ler/gw_source_population/cbc_source_redshift_distribution.py
@@ -2,10 +2,10 @@
 """
 Module for compact binary coalescence (CBC) source redshift distribution.
 
-This module provides the :class:`~CBCSourceRedshiftDistribution` class for 
-generating redshift distributions of compact binary sources (BBH, BNS, NSBH) 
-based on various merger rate density models. It supports multiple astrophysical 
-merger rate density prescriptions including PopI/II, PopIII, and primordial 
+This module provides the :class:`~CBCSourceRedshiftDistribution` class for
+generating redshift distributions of compact binary sources (BBH, BNS, NSBH)
+based on various merger rate density models. It supports multiple astrophysical
+merger rate density prescriptions including PopI/II, PopIII, and primordial
 black hole models.
 
 Inheritance hierarchy:
@@ -24,6 +24,7 @@ Copyright (C) 2024 Hemantakumar Phurailatpam. Distributed under MIT License.
 """
 
 import warnings
+
 warnings.filterwarnings("ignore")
 
 import numpy as np
@@ -34,16 +35,21 @@ from scipy.interpolate import CubicSpline
 # for redshift to luminosity distance conversion
 from astropy.cosmology import LambdaCDM
 
-from ..utils import  FunctionConditioning, interpolator_json_path, luminosity_distance, differential_comoving_volume
+from ..utils import (
+    FunctionConditioning,
+    interpolator_json_path,
+    luminosity_distance,
+    differential_comoving_volume,
+)
 
 
 class CBCSourceRedshiftDistribution(object):
     """
     Class for generating compact binary coalescence source redshift distributions.
 
-    This class generates source redshift distributions for compact binary 
-    coalescence events (BBH, BNS, NSBH) using various astrophysical merger rate 
-    density models. It provides interpolated functions for efficient sampling of 
+    This class generates source redshift distributions for compact binary
+    coalescence events (BBH, BNS, NSBH) using various astrophysical merger rate
+    density models. It provides interpolated functions for efficient sampling of
     source redshifts weighted by the merger rate density in the detector frame.
 
     Key Features: \n
@@ -176,18 +182,52 @@ class CBCSourceRedshiftDistribution(object):
         self.cosmo = cosmology if cosmology else LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
 
         # setting up the interpolator creation parameters
-        self.create_new_interpolator = self._setup_decision_dictionary(create_new_interpolator, merger_rate_density)
+        self.create_new_interpolator = self._setup_decision_dictionary(
+            create_new_interpolator, merger_rate_density
+        )
 
         # Initialize cosmological functions
-        self.luminosity_distance = luminosity_distance(z_min=self.z_min, z_max=self.z_max, cosmo=self.cosmo, directory=self.directory, create_new=self.create_new_interpolator["luminosity_distance"]["create_new"], resolution=self.create_new_interpolator["luminosity_distance"]["resolution"], get_attribute=True)
-        self.differential_comoving_volume = differential_comoving_volume(z_min=self.z_min, z_max=self.z_max, cosmo=self.cosmo, directory=self.directory, create_new=self.create_new_interpolator["differential_comoving_volume"]["create_new"], resolution=self.create_new_interpolator["differential_comoving_volume"]["resolution"], get_attribute=True)
+        self.luminosity_distance = luminosity_distance(
+            z_min=self.z_min,
+            z_max=self.z_max,
+            cosmo=self.cosmo,
+            directory=self.directory,
+            create_new=self.create_new_interpolator["luminosity_distance"][
+                "create_new"
+            ],
+            resolution=self.create_new_interpolator["luminosity_distance"][
+                "resolution"
+            ],
+            get_attribute=True,
+        )
[…]
