### 2026-01-20 9606082
setup.py updates; doc example updates
