+++ docs/autoapi/ler/lens_galaxy_population/sampler_functions/index.rst
@@ -91,7 +91,7 @@ Functions
    ..
        !! processed by numpydoc !!
 
-.. py:function:: lens_redshift_strongly_lensed_sis_haris_pdf(zl, zs, cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7))
+.. py:function:: lens_redshift_strongly_lensed_sis_haris_pdf(zl, zs, cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0))
 
    
    Compute lens redshift PDF for SIS model (Haris et al. 2018).
@@ -111,7 +111,7 @@ Functions
        **cosmo** : ``astropy.cosmology``
            Cosmology object for distance calculations.
 
-           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
    :Returns:
 
@@ -130,7 +130,7 @@ Functions
    .. rubric:: Examples
 
    >>> from astropy.cosmology import LambdaCDM
-   >>> cosmo = LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+   >>> cosmo = LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
    >>> pdf = lens_redshift_strongly_lensed_sis_haris_pdf(zl=0.5, zs=1.0, cosmo=cosmo)
    >>> print(f"PDF at zl=0.5: {pdf:.4f}")
    PDF at zl=0.5: 1.8750
@@ -140,7 +140,7 @@ Functions
    ..
        !! processed by numpydoc !!
 
-.. py:function:: lens_redshift_strongly_lensed_sis_haris_rvs(size, zs, z_min=0.001, z_max=10.0, cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7))
+.. py:function:: lens_redshift_strongly_lensed_sis_haris_rvs(size, zs, z_min=0.001, z_max=10.0, cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0))
 
    
    Sample lens redshifts for SIS model (Haris et al. 2018).
@@ -169,7 +169,7 @@ Functions
        **cosmo** : ``astropy.cosmology``
            Cosmology object for distance calculations.
 
-           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
    :Returns:
 
@@ -188,7 +188,7 @@ Functions
    .. rubric:: Examples
 
    >>> from astropy.cosmology import LambdaCDM
-   >>> cosmo = LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+   >>> cosmo = LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
    >>> zl_samples = lens_redshift_strongly_lensed_sis_haris_rvs(
    ...     size=1000,
    ...     zs=1.5,
