+++ docs/autoapi/ler/rates/index.rst
@@ -894,7 +894,7 @@ Functions
        **cosmology** : ``astropy.cosmology``
            Cosmology to use for the calculation.
 
-           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
        **pdet_finder** : ``function`` or ``None``
            Custom detection probability finder function.
@@ -1231,7 +1231,7 @@ Functions
           **cosmo** : ``astropy.cosmology``
               Cosmology used for luminosity distance and comoving volume calculations.
 
-              default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+              default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
 
 
@@ -2523,7 +2523,7 @@ Functions
        **cosmology** : ``astropy.cosmology`` or ``None``
            Cosmology to use for distance calculations.
 
-           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
        **spin_zero** : ``bool``
            If True, spin parameters are set to zero (no spin sampling).
@@ -4210,7 +4210,7 @@ Functions
        **cosmology** : ``astropy.cosmology`` or ``None``
            Cosmology to use for the calculation.
 
-           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
        **pdet_finder** : ``callable`` or ``None``
            Custom detection probability finder function.
@@ -4469,7 +4469,7 @@ Functions
           **cosmo** : ``astropy.cosmology``
               Cosmology used for luminosity distance and comoving volume calculations.
 
-              default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+              default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
 
 
