+++ docs/autoapi/ler/gw_source_population/cbc_source_parameter_distribution/index.rst
@@ -124,7 +124,7 @@ Attributes
        **cosmology** : ``astropy.cosmology`` or ``None``
            Cosmology to use for distance calculations.
 
-           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7)
+           default: LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
        **spin_zero** : ``bool``
            If True, spin parameters are set to zero (no spin sampling).
