### 2026-01-20 b8c7162
setup.py updates; doc example updates
