### 2026-01-22 9aaeb1b
setup.py updates
