+++ docs/analytical_formulation_unlensed.md
@@ -1,0 +1,102 @@
+# Analytical Formulation for Gravitational Wave Event Rates
+
+The annual rate of detectable unlensed gravitational wave (GW) events, $\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}$, can be determined by integrating the intrinsic rate of compact binary mergers across cosmic time and volume, weighted by the probability of detecting these events. This relationship can be expressed by first considering the total intrinsic merger rate per year, $\frac{\Delta N_{\rm U}}{\Delta t}$, and the average detection probability, $P({\rm obs})$, and it reads,
+
+$$
+\begin{equation}
+\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t} = \frac{\Delta N_{\rm U}}{\Delta t} \times P({\rm obs}) \tag{1}
+\end{equation}
+$$
+
+The detection probability $P({\rm obs})$ is effectively marginalized over all relevant GW source parameters $\vec{\theta}$, i.e.\,
+
+$$\vec{\theta}\in \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{JL}, \theta_{\rm jn}, \phi, \psi, {\rm RA, Dec}, t_c\}$$
+
+Refer to [Table 1](#table-1-prior-distributions-for-gw-source-parameters-used-in-the-unlensed-event-rate-calculations) for the complete list of GW parameters and their prior distributions used in this analysis.
+
+To account explicitly for the parameters dependence, the observed event rate can be written as an integral over $\vec{\theta}$, as,
+
+$$
+\begin{equation}
+\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t} = \frac{\Delta N_{\rm U}}{\Delta t} \int_{\vec{\theta}} P({\rm obs}|\vec{\theta}) \, P(\vec{\theta}) \, d\vec{\theta} \tag{2}
+\end{equation}
+$$
+
+where $P({\rm obs}|\vec{\theta})$ is the conditional probability of detecting a source given its parameters $\vec{\theta}$, and $P(\vec{\theta})$ is the joint prior distribution of the source parameters. $P(\vec{\theta})$ can be expanded to it's constituent parts, starting with the redshift distribution of sources, and it reads,
+
+$$
+\begin{equation}
+\begin{split}
+P(\vec{\theta}) &= P(z_s) \, P(m_1, m_2) \, P(a_1) \, P(a_2) \, P(\theta_1, \theta_2) \, P(\phi_{12}) \, P(\phi_{JL}) \\
+& \quad \times P(\theta_{\rm jn}) \, P(\phi) \, P(\psi) \, P({\rm RA}) \, P({\rm Dec}) \, P(t_c)
+\end{split}
+\end{equation}
+$$
+
+`ler` uses redshift ($z_s$) instead of luminosity distance ($D_L$) as one of the source parameters, and we assume that it is uncorrelated with the other parameters. Then, it's probability density distribution $P(z_s)$ is the normalized redshift distribution of sources and it propotionally relates to the differential intrinsic merger rate at detector frame, i.e.,
+
+$$
+\begin{equation}
+\begin{split}
+P(z_s) &\propto \frac{d^2 N}{dt \, dV_c} \frac{dV_c}{dz_s} \\
+&\propto \frac{1}{(1+z_s)}\frac{d^2 N}{d\tau \, dV_c} \frac{dV_c}{dz_s} \\
+& \propto \frac{1}{(1+z_s)} R_{\rm U}(z_s) \frac{dV_c}{dz_s}
+\end{split} \tag{3}
+\end{equation}
+$$
+
[…]
