+++ ler/image_properties/image_properties.py
@@ -63,8 +63,8 @@ class ImageProperties():
         Maximum number of images to consider per event. \n
         default: 4
     time_window : ``float``
-        Time window for lensed events (units: seconds). \n
-        default: 365*24*3600*20 (20 years)
+        Time window for lensed events from min(geocent_time) (units: seconds). \n
+        default: 365*24*3600*2 (2 years)
     lens_model_list : ``list``
         List of lens models to use. \n
         default: ['EPL_NUMBA', 'SHEAR']
@@ -133,6 +133,11 @@ class ImageProperties():
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
     | :attr:`~spin_precession`                            | ``bool``                  |          | Flag for spin precession                       |
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
+    | :attr:`~pdet_finder`                                | ``callable``              |          | Probability of detection calculator            |
+    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
+    | :attr:`~pdet_finder_output_keys`                     | ``list``                  |          | Keys for probability of detection outputs      |
+    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
+
     """
 
     def __init__(self, 
@@ -141,9 +146,10 @@ class ImageProperties():
                  n_max_images=4,
                  lens_model_list=['EPL_NUMBA', 'SHEAR'],
                  cosmology=None,
-                 time_window=365*24*3600*20,
+                 time_window=365*24*3600*2, # 2 years
                  spin_zero=True,
                  spin_precession=False,
+                 pdet_finder=None,
         ):
 
         self.npool = npool
@@ -154,6 +160,8 @@ class ImageProperties():
         self.spin_precession = spin_precession
         self.time_window = time_window
         self.cosmo = cosmology if cosmology else cosmo
+        self.pdet_finder = pdet_finder
+        self.pdet_finder_output_keys = None
 
     def image_properties(self, lens_parameters):
         """
@@ -333,7 +341,7 @@ class ImageProperties():
 
         return lens_parameters
 
-    def get_lensed_snrs(self, lensed_param, pdet_calculator, list_of_detectors=None):
+    def get_lensed_snrs(self, lensed_param, pdet_finder=None):
         """
         Compute detection probability for each lensed image.
 
@@ -353,18 +361,15 @@ class ImageProperties():
             - 'magnifications': image magnifications (shape: size x n_max_images) \n
             - 'time_delays': image time delays (shape: size x n_max_images) \n
             - 'image_type': morse phase type (shape: size x n_max_images) \n
-        pdet_calculator : ``callable``
+        pdet_finder : ``callable``
             Function that computes detection probability given GW parameters.
-        list_of_detectors : ``list`` or ``None``
-            List of detector names (e.g., ['H1', 'L1', 'V1']) for per-detector results. \n
-            default: None
 
         Returns
         -------
         result_dict : ``dict``
             Dictionary containing: \n
             - 'pdet_net': network detection probability (shape: size x n_max_images) \n
-            - Individual detector probabilities if list_of_detectors provided \n
+            - Individual detector probabilities if pdet_finder outputs them \n
         lensed_param : ``dict``
             Updated dictionary with effective parameters: \n
             - 'effective_luminosity_distance': magnification-corrected distance \n
@@ -382,14 +387,16 @@ class ImageProperties():
         image_type[image_type==2.] = np.pi/2
         image_type[image_type==3.] = np.pi
[…]
