### 2026-01-23 5dda509
version=0.5.2, docs:analytical_unlensed
