+++ pyproject.toml
@@ -24,7 +24,6 @@ classifiers = [
 dependencies = [
     "setuptools>=65.5.0",
     "matplotlib>=3.4.2",
-    "numpy<=1.26.4",
     "numba>=0.60.0",
     "gwsnr>=0.4.0",
     "importlib_resources",
