### 2026-01-24 #7420 @hemantaph
@warrickball @kazewong I have been finalising the programming component of the package and making it available to others for testing, to ensure that it is scientifically correct, coherent, and robust across a range of real-world use cases. The programming work was wrapped up yesterday, and I am now focusing on implementing the revisions requested by the reviewer on the manuscript. This should be completed within the next few days, after which the full package and paper will be ready.

I apologise for the extended timeline. Because this is a scientific software package, it requires thorough validation and repeated testing to ensure consistency and scientific accuracy. At one stage, I also needed to revisit and re-derive parts of the underlying mathematical formulation, as some details in the published literature (e.g., Haris et al. 2018) did not remain consistent when implemented within this framework. I appreciate your patience and understanding.
