+++ docs/analytical_formulation_unlensed copy.md
@@ -1,96 +1,106 @@
 # Analytical Formulation for Gravitational Wave Event Rates
 
-The annual rate of detectable unlensed gravitational wave (GW) events, $\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}$, can be determined by integrating the intrinsic rate of compact binary mergers across cosmic time and volume, weighted by the probability of detecting these events. This relationship can be expressed by first considering the total intrinsic merger rate per year, $\frac{\Delta N_{\rm U}}{\Delta t}$, and the average detection probability, $P({\rm obs})$, and it reads,
+## Table of Contents
+
+1. [Introduction](#introduction)
+2. [Parameter-Marginalized Event Rate](#parameter-marginalized-event-rate)
+3. [Redshift Distribution and Intrinsic Merger Rates](#redshift-distribution-and-intrinsic-merger-rates)
+4. [Detection Criterion and SNR Modeling](#detection-criterion-and-snr-modeling)
+5. [GW Source Parameter Priors](#gw-source-parameter-priors)
+
+## Introduction
+
+The annual rate of detectable (unlensed) gravitational-wave (GW) events, $\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}$, gives the expected number of observed compact-binary coalescences (BBH, BNS, and NSBH) per year for a given detector or detector network. It is obtained by combining the total intrinsic merger rate in the detector frame, $\frac{\Delta N_{\rm U}}{\Delta t}$, with the population-averaged probability of detection, $P({\rm obs})$, such that
 
 $$
 \begin{equation}
 \begin{split}
-\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t} = \frac{\Delta N_{\rm U}}{\Delta t} \times P({\rm obs}) \,.
+\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}
+= \frac{\Delta N_{\rm U}}{\Delta t} \times P({\rm obs}) \,.
 \end{split}
 \end{equation}
 $$
 
-The detection probability $P({\rm obs})$ is effectively marginalized over all relevant GW source parameters $\vec{\theta}$, i.e.\,
+## Parameter-Marginalized Event Rate
 
-$$\vec{\theta}\in \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{JL}, \theta_{\rm jn}, \phi, \psi, {\rm RA, Dec}, t_c\}$$
+The detection probability $P({\rm obs})$ is obtained by marginalizing over the relevant GW source parameters $\vec{\theta}$, i.e.,
 
-Refer to [Table 1](#table-1-prior-distributions-for-gw-source-parameters-used-in-the-unlensed-event-rate-calculations) for the complete list of GW parameters and their prior distributions used in this analysis.
+$$
+\vec{\theta}\in \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{JL}, \iota, \phi, \psi, {\rm RA}, {\rm Dec}, t_c\},
+$$
 
-To account explicitly for the parameters dependence, the observed event rate can be written as an integral over $\vec{\theta}$, as,
[…]
