+++ docs/analytical_formulation_unlensed.md
@@ -1,6 +1,16 @@
 # Analytical Formulation for Gravitational Wave Event Rates
 
-The annual rate of detectable **unlensed** gravitational wave (GW) events, $\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}$, can be determined by integrating the intrinsic rate of compact binary mergers across cosmic time and comoving volume, weighted by the probability that these events are detected. If the total intrinsic merger rate per year in the detector frame is denoted by $\frac{\Delta N_{\rm U}}{\Delta t}$, and the average detection probability is $P({\rm obs})$, then the observed event rate can be written as
+## Table of Contents
+
+1. [Introduction](#introduction)
+2. [Parameter-Marginalized Event Rate](#parameter-marginalized-event-rate)
+3. [Redshift Distribution and Intrinsic Merger Rates](#redshift-distribution-and-intrinsic-merger-rates)
+4. [Detection Criterion and SNR Modeling](#detection-criterion-and-snr-modeling)
+5. [GW Source Parameter Priors](#gw-source-parameter-priors)
+
+## Introduction
+
+The annual rate of detectable (unlensed) gravitational-wave (GW) events, $\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}$, gives the expected number of observed compact-binary coalescences (BBH, BNS, and NSBH) per year for a given detector or detector network. It is obtained by combining the total intrinsic merger rate in the detector-frame, $\frac{\Delta N_{\rm U}}{\Delta t}$, with the population-averaged probability of detection, $P({\rm obs})$, such that
 
 $$
 \begin{equation}
@@ -11,85 +21,80 @@ $$
 \end{equation}
 $$
 
-The detection probability $P({\rm obs})$ is effectively marginalized over all relevant GW source parameters $\vec{\theta}$, i.e.,
+## Parameter-Marginalized Event Rate
+
+The detection probability $P({\rm obs})$ is obtained by marginalizing over the relevant GW source parameters $\vec{\theta}$, i.e.,
 
 $$
-\vec{\theta}\in \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{JL}, \iota, \phi, \psi, {\rm RA}, {\rm Dec}, t_c\}.
+\vec{\theta}\in \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{JL}, \iota, \phi, \psi, {\rm RA}, {\rm Dec}, t_c\},
 $$
 
-Refer to [Table 1](#table-1-prior-distributions-for-gw-source-parameters-used-in-the-unlensed-event-rate-calculations) for the list of GW parameters and their prior distributions used in this analysis.
+where $z_s$ is the source redshift, $(m_1,m_2)$ are component masses, $(a_1,a_2)$ are dimensionless spin magnitudes, $(\theta_1,\theta_2,\phi_{12},\phi_{JL})$ describe spin orientations, $\iota$ is the inclination angle, $(\phi,\psi)$ are the coalescence phase and polarization angle, $({\rm RA},{\rm Dec})$ specify sky position, and $t_c$ is the coalescence time. The prior distributions used in this analysis are summarized in [Table 1](#gw-source-parameter-priors).
 
-To account explicitly for the parameter dependence, the observed event rate can be written as an integral over $\vec{\theta}$:
-
-$$
[…]
