### 2026-01-26 2d4d71e
doc update, SFR_time_delay.ipynb
