+++ ler/gw_source_population/prior_functions.py
@@ -159,7 +159,7 @@ def merger_rate_density_madau_dickinson_belczynski_ng_function(zs, R0=19 * 1e-9,
 
     This model uses a Madau-Dickinson-like functional form to fit the 
     merger rate density of field BHs, accounting for time delays and 
-    metallicity effects.
+    metallicity effects. Coefficients from Madau & Dickinson (2014) are translated as: B-> alpha_F, D-> beta_F, C-> c_F.
 
     density(zs) ∝ (1 + zs) ** alpha_F / (1 + ((1 + zs) / c_F) ** beta_F)
 
