+++ examples/rates/SFR_time_delay.md
@@ -1,0 +1,480 @@
+# Tutorial: Computing the Source-Frame BBH Merger-Rate Density from SFR + Exponential Time Delay
+
+This tutorial explains how to compute the **source-frame volumetric merger-rate density** of binary black holes (BBHs),
+
+$$ 
+\begin{equation}
+\begin{split}
+R_{\rm m}(z)\equiv \frac{dN_{\rm merge}}{dV_c\,dt_s},
+\end{split}
+\end{equation} 
+$$
+
+starting from the **cosmic star-formation rate density** (SFR) and a **time-delay distribution** between binary formation and merger.
+
+The goal is that a Master's student can:
+1. understand the mathematics, and  
+2. implement it correctly in Python.
+
+We follow the analytical formulation used in *Vitale et al., "Measuring the star formation rate with gravitational waves from binary black holes"*.
+
+---
+
+## 0) What you will compute (the target quantity)
+
+### Source-frame merger-rate density
+
+The source-frame merger-rate density is defined as
+
+$$ 
+\begin{equation}
+\begin{split}
+R_{\rm m}(z)\equiv \frac{dN_{\rm merge}}{dV_c\,dt_s}
+\quad [{\rm Gpc^{-3}\,yr^{-1}}].
+\end{split}
+\end{equation} 
+$$
+
+- $dV_c$: comoving volume element  
+- $dt_s$: **source-frame** time interval  
+- $z$: merger redshift  
+
+> This is an **astrophysical merger-rate density**, not a detector yield.
+
+---
+
+## 1) Ingredients of the model (physical picture)
+
+BBHs in the field form from massive stellar binaries. The merger-rate density at redshift $z$ depends on:
+
+1. **how many stars formed at earlier cosmic times** (the SFR history),  
+2. **how efficiently that star formation produces BBH progenitors** (metallicity dependent),  
+3. **how long it takes from formation to merger** (time delay).  
+
+Mathematically, this becomes a **convolution** of the formation rate with a delay-time distribution.
+
+---
+
+## 2) Cosmology: mapping redshift $\leftrightarrow$ cosmic time
+
+To connect **formation redshift** $z_f$ to **merger redshift** $z_m$, we use the cosmic time $t(z)$, interpreted as the **age of the Universe at redshift $z$**.
+
+### 2.1 Expansion rate
+
+Assuming flat $\Lambda$CDM, the Hubble rate is
+
+$$ 
+\begin{equation}
+\begin{split}
+H(z)=H_0\,E(z),
+\qquad
+E(z)=\sqrt{\Omega_m(1+z)^3+\Omega_\Lambda}.
+\end{split}
+\end{equation} 
+$$
+
+### 2.2 Differential time–redshift relation
+
+Cosmic time decreases with redshift as
+
+$$ 
+\begin{equation}
+\begin{split}
+\frac{dt}{dz}=-\frac{1}{(1+z)H(z)}.
+\end{split}
+\end{equation} 
+$$
+
+We will often use the positive quantity
+
+$$ 
+\begin{equation}
+\begin{split}
+\left|\frac{dt}{dz}\right|=\frac{1}{(1+z)H(z)}.
[…]
