+++ docs/analytical_formulation_unlensed copy 2.md
@@ -1,141 +1,0 @@
-# Analytical Formulation for Gravitational Wave Event Rates
-
-The annual rate of detectable (unlensed) gravitational-wave (GW) events, $\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}$, gives the expected number of observed compact-binary coalescences (BBH, BNS, and NSBH) per year for a given detector or detector network. It is obtained by combining the total intrinsic merger rate in the detector frame, $\frac{\Delta N_{\rm U}}{\Delta t}$, with the population-averaged probability of detection, $P({\rm obs})$, such that
-
-$$
-\begin{equation}
-\begin{split}
-\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}
-= \frac{\Delta N_{\rm U}}{\Delta t} \times P({\rm obs}) \,,
-\end{split}
-\end{equation}
-$$
-
-where $P({\rm obs})$ is evaluated by marginalizing over the full set of GW parameters $\vec{\theta}$, i.e.,
-
-$$
-\vec{\theta}\in \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{JL}, \iota, \phi, \psi, {\rm RA}, {\rm Dec}, t_c\} \, .
-$$
-
-See [Table 1](#table-1-prior-distributions-for-gw-source-parameters-used-in-the-unlensed-event-rate-calculations) for parameter descriptions and adopted prior distributions used in this analysis.
-
-To make the parameter dependence explicit, the observed event rate can be written as an integral over $\vec{\theta}$, and it reads
-
-$$
-\begin{equation}
-\begin{split}
-\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}
-= \frac{\Delta N_{\rm U}}{\Delta t} \int_{\vec{\theta}} P({\rm obs}|\vec{\theta}) \, P(\vec{\theta}) \, d\vec{\theta} \, ,
-\end{split}
-\end{equation}
-$$
-
-where $P({\rm obs}|\vec{\theta})$ is the conditional probability of detecting a source given its parameters $\vec{\theta}$, and $P(\vec{\theta})$ is the joint prior distribution of source parameters. Assuming the priors factorize, $P(\vec{\theta})$ can be expanded as
-
-$$
-\begin{equation}
-\begin{split}
-P(\vec{\theta}) &= P(z_s) \, P(m_1, m_2) \, P(a_1) \, P(a_2) \, P(\theta_1, \theta_2) \, P(\phi_{12}) \, P(\phi_{JL}) \\
-& \quad \times P(\iota) \, P(\phi) \, P(\psi) \, P({\rm RA}) \, P({\rm Dec}) \, P(t_c) \, .
-\end{split}
-\end{equation}
-$$
-
-The `ler` package uses redshift, $z_s$, rather than luminosity distance, $D_L$, as a source parameter, and it's intrinsic distribution is assumed to be uncorrelated with the other parameters. The redshift probability density distribution, $P(z_s)$, is defined as the normalized distribution of sources in redshift, and it is proportional to the intrinsic merger rate density in the detector frame and the comoving volume element. This can be written as
-
-$$
-\begin{equation}
-\begin{split}
-P(z_s) &\propto \frac{d^2 N}{dt \, dV_c} \frac{dV_c}{dz_s} \\
-&\propto \frac{1}{(1+z_s)}\frac{d^2 N}{d\tau \, dV_c} \frac{dV_c}{dz_s} \\
[…]

[…]
