### 2026-02-01 7d647d0
bug fix in resume of event selection, ler.py
doc updates
