+++ ler/rates/ler.py
@@ -350,8 +350,9 @@ class LeR(LensGalaxyParameterDistribution):
             # ImageProperties class params
             n_min_images=2,
             n_max_images=4,
-            time_window=365 * 24 * 3600 * 20,
+            time_window=365 * 24 * 3600 * 2,
             lens_model_list=["EPL_NUMBA", "SHEAR"],
+            effective_params_in_output=True,
             # CBCSourceParameterDistribution class params
             event_type=self.event_type,
             source_priors=None,
@@ -385,6 +386,7 @@ class LeR(LensGalaxyParameterDistribution):
             n_min_images=input_params["n_min_images"],
             n_max_images=input_params["n_max_images"],
             time_window=input_params["time_window"],
+            effective_params_in_output=input_params["effective_params_in_output"],
             lens_model_list=input_params["lens_model_list"],
             # CBCSourceParameterDistribution class params
             event_type=input_params["event_type"],
@@ -648,6 +650,7 @@ class LeR(LensGalaxyParameterDistribution):
         print(f"    n_min_images = {self.ler_args['n_min_images']},")
         print(f"    n_max_images = {self.ler_args['n_max_images']},")
         print(f"    time_window = {self.ler_args['time_window']},")
+        print(f"    effective_params_in_output = {self.ler_args['effective_params_in_output']},")
         print(f"    lens_model_list = {self.ler_args['lens_model_list']},")
 
         if "pdet_args" in self.ler_args:
@@ -1129,7 +1132,7 @@ class LeR(LensGalaxyParameterDistribution):
             +------------------------------+-----------+-------------------------------------------------------+
             | q                            |           | axis ratio                                            |
             +------------------------------+-----------+-------------------------------------------------------+
-            | theta_E                      | arcsec    | Einstein radius                                       |
+            | theta_E                      | radian    | Einstein radius                                       |
             +------------------------------+-----------+-------------------------------------------------------+
             | phi                          | rad       | axis rotation angle. counter-clockwise from the       |
             |                              |           | positive x-axis (RA-like axis) to the major axis of   |
@@ -1159,6 +1162,8 @@ class LeR(LensGalaxyParameterDistribution):
             +------------------------------+-----------+-------------------------------------------------------+
             | a_2                          |           | spin of the secondary compact binary                  |
             +------------------------------+-----------+-------------------------------------------------------+
+            | luminosity_distance          | Mpc       | luminosity distance of the source                      |
+            +------------------------------+-----------+-------------------------------------------------------+
             | mass_1_source                | Msun      | mass of the primary compact binary (source frame)     |
             +------------------------------+-----------+-------------------------------------------------------+
             | mass_2_source                | Msun      | mass of the secondary compact binary (source frame)   |
@@ -1167,9 +1172,9 @@ class LeR(LensGalaxyParameterDistribution):
             +------------------------------+-----------+-------------------------------------------------------+
             | mass_2                       | Msun      | mass of the secondary compact binary (detector frame) |
             +------------------------------+-----------+-------------------------------------------------------+
-            | x0_image_positions           | arcsec    | x-coordinate (RA-like axis) of the images             |
+            | x0_image_positions           | radian    | x-coordinate (RA-like axis) of the images             |
             +------------------------------+-----------+-------------------------------------------------------+
-            | x1_image_positions           | arcsec    | y-coordinate (Dec-like axis) of the images            |
+            | x1_image_positions           | radian    | y-coordinate (Dec-like axis) of the images            |
             +------------------------------+-----------+-------------------------------------------------------+
             | magnifications               |           | magnifications                                        |
             +------------------------------+-----------+-------------------------------------------------------+
@@ -1179,11 +1184,24 @@ class LeR(LensGalaxyParameterDistribution):
[…]
