+++ ler/image_properties/image_properties.py
@@ -65,6 +65,9 @@ class ImageProperties():
     time_window : ``float``
         Time window for lensed events from min(geocent_time) (units: seconds). \n
         default: 365*24*3600*2 (2 years)
+    effective_params_in_output : ``bool``
+        Whether to include effective parameters (effective_phase, effective_ra, effective_dec) in the output. \n
+        default: True
     lens_model_list : ``list``
         List of lens models to use. \n
         default: ['EPL_NUMBA', 'SHEAR']
@@ -125,6 +128,8 @@ class ImageProperties():
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
     | :attr:`~time_window`                                | ``float``                 | s        | Time window for lensed events                  |
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
+    | :attr:`~effective_params_in_output`                 | ``bool``                  |          | To include effective parameters in output      |
+    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
     | :attr:`~lens_model_list`                            | ``list``                  |          | List of lens models                            |
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
     | :attr:`~cosmo`                                      | ``astropy.cosmology``     |          | Cosmology for calculations                     |
@@ -135,7 +140,7 @@ class ImageProperties():
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
     | :attr:`~pdet_finder`                                | ``callable``              |          | Probability of detection calculator            |
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-    | :attr:`~pdet_finder_output_keys`                     | ``list``                  |          | Keys for probability of detection outputs      |
+    | :attr:`~pdet_finder_output_keys`                    | ``list``                  |          | Keys for probability of detection outputs      |
     +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
 
     """
@@ -150,6 +155,7 @@ class ImageProperties():
                  spin_zero=True,
                  spin_precession=False,
                  pdet_finder=None,
+                 effective_params_in_output=True,
         ):
 
         self.npool = npool
@@ -162,6 +168,7 @@ class ImageProperties():
         self.cosmo = cosmology if cosmology else cosmo
         self.pdet_finder = pdet_finder
         self.pdet_finder_output_keys = None
+        self.effective_params_in_output = effective_params_in_output
 
     def image_properties(self, lens_parameters):
         """
@@ -173,28 +180,55 @@ class ImageProperties():
         Parameters
         ----------
         lens_parameters : ``dict``
-            Dictionary containing lens and source parameters with keys: \n
-            - 'zs': source redshift (array) \n
-            - 'zl': lens redshift (array) \n
-            - 'gamma1': external shear component 1 (array) \n
-            - 'gamma2': external shear component 2 (array) \n
-            - 'phi': position angle of lens ellipticity (array) \n
-            - 'q': axis ratio of lens (array) \n
-            - 'gamma': power-law slope of mass density (array) \n
-            - 'theta_E': Einstein radius in radians (array) \n
+            Dictionary containing lens and source parameters shown in the table: \n
+            +------------------------------+-----------+-------------------------------------------------------+
+            | Parameter                    | Units     | Description                                           |
+            +==============================+===========+=======================================================+
+            | zl                           |           | redshift of the lens                                  |
+            +------------------------------+-----------+-------------------------------------------------------+
+            | zs                           |           | redshift of the source                                |
+            +------------------------------+-----------+-------------------------------------------------------+
+            | sigma                        | km s^-1   | velocity dispersion                                   |
+            +------------------------------+-----------+-------------------------------------------------------+
[…]
