+++ docs/unlensed/analytical_formulation_unlensed.tex
@@ -1,0 +1,260 @@
+\documentclass[11pt]{article}
+
+% -------------------------------------------------
+% Packages
+% -------------------------------------------------
+\usepackage{amsmath,amssymb}
+\usepackage{physics}
+\usepackage{graphicx}
+\usepackage{booktabs}
+\usepackage{longtable}
+\usepackage{geometry}
+\usepackage{hyperref}
+\usepackage{caption}
+\usepackage{array}
+
+\geometry{margin=1in}
+
+% -------------------------------------------------
+% Metadata
+% -------------------------------------------------
+\title{Analytical Formulation for Gravitational Wave Event Rates}
+\author{Phurailatpam Hemantakumar\\
+\href{https://hemantaph.com}{https://hemantaph.com}}
+\date{Last updated: June 2024}
+
+% -------------------------------------------------
+\begin{document}
+\maketitle
+\tableofcontents
+\newpage
+
+% =================================================
+\section{Overview}
+
+This document presents the analytical framework for calculating gravitational-wave (GW) event rates from compact-binary coalescences (CBCs), encompassing binary black holes (BBH), binary neutron stars (BNS), and neutron star--black hole binaries (NSBH). The formulation combines cosmic merger-rate densities with detector sensitivity and selection effects to predict annual detection rates. Practical implementations using the \texttt{ler} package are referenced throughout.
+
+% =================================================
+\section{Introduction}
+
+The annual rate of detectable (unlensed) GW events,
+$\Delta N^{\rm obs}_{\rm U} / \Delta t$,
+represents the expected number of observed CBC mergers per year for a given detector network. It is obtained by combining the total intrinsic merger rate in the detector frame,
+$\Delta N_{\rm U} / \Delta t$,
+with the population-averaged probability of detection $P({\rm obs})$:
+%
+\begin{equation}
+\begin{split}
+\frac{\Delta N^{\rm obs}_{\rm U}}{\Delta t}
+=
+\frac{\Delta N_{\rm U}}{\Delta t}
+\times
+P({\rm obs}) \, .
+\end{split}
+\end{equation}
+
+% =================================================
+\section{Parameter-Marginalized Event Rate}
+
+Let the full parameter vector be
+$\vec{\theta} = \{\vec{\theta}_{\rm int}, \vec{\theta}_{\rm ext}\}$,
+where intrinsic parameters describe the source-frame properties of the binary, and extrinsic parameters describe its configuration relative to the detector network. Explicitly,
+%
+\begin{equation}
+\begin{split}
+\vec{\theta} =
+\{&
+z_s, m_1, m_2, a_1, a_2,
+\theta_1, \theta_2, \phi_{12}, \phi_{JL}, \\
+&\iota, \phi, \psi, {\rm RA}, {\rm Dec}, t_c
+\} .
+\end{split}
+\end{equation}
+
+Given a joint prior $P(\vec{\theta})$ and conditional detection probability
+$P({\rm obs} \mid \vec{\theta})$,
+the population-averaged detection probability is
+%
+\begin{equation}
+\begin{split}
+P({\rm obs})
+=
+\int
+P({\rm obs} \mid \vec{\theta})\,
+P(\vec{\theta})\,
+d\vec{\theta} .
+\end{split}
+\end{equation}
+
[…]
