+++ docs/analytical_formulation_unlensed_pdf.rst
@@ -1,22 +1,0 @@
-Analytical Formulation (PDF)
-==============================
-
-The detailed analytical formulation for unlensed gravitational wave event rates is available as a PDF document:
-
-:download:`Download Analytical Formulation PDF <_static/analytical_formulation_unlensed.pdf>`
-
-You can also `view the PDF directly <_static/analytical_formulation_unlensed.pdf>`_.
-
-Overview
---------
-
-This document presents the analytical framework for calculating gravitational-wave (GW) event rates from compact-binary coalescences (CBCs), encompassing binary black holes (BBH), binary neutron stars (BNS), and neutron star–black hole binaries (NSBH). The formulation combines cosmic merger-rate densities with detector sensitivity and selection effects to predict annual detection rates.
-
-Key Topics Covered
-------------------
-
-* Parameter-marginalized event rates
-* Redshift distribution and intrinsic merger rates
-* Detection criterion and SNR modeling
-* Simulation results and rate estimates
-* GW source parameter priors
