+++ docs/analytical_formulation_unlensed.md
@@ -1,6 +1,6 @@
 # Analytical Formulation for Gravitational Wave Event Rates
 
-Written by [Phurailatpam Hemantakumar](https://hemantaph.com). Last updated: June 2024.
+Written by [Phurailatpam Hemantakumar](https://hemantaph.com). Last updated: 2026-02-06.
 
 ## Overview
 
@@ -12,6 +12,7 @@ This document presents the analytical framework for calculating gravitational-wa
 - [Parameter-Marginalized Event Rate](#parameter-marginalized-event-rate)
 - [Redshift Distribution and Intrinsic Merger Rates](#redshift-distribution-and-intrinsic-merger-rates)
 - [Detection Criterion and SNR Modeling](#detection-criterion-and-snr-modeling)
+- [Complete Expression for GW (unlensed) Event Rate](#complete-expression-for-gw-unlensed-event-rate)
 - [Simulation Results](#simulation-results)
   - [Simulation Settings](#simulation-settings)
   - [GW Source Parameter Priors](#gw-source-parameter-priors)
@@ -34,17 +35,47 @@ $$
 
 ## Parameter-Marginalized Event Rate
 
-The observed event rate is obtained by averaging the detection probability over the compact-binary parameters that determine the emitted signal and its projection onto a detector network. Let $\vec{\theta}=\{\vec{\theta}_{\rm int},\vec{\theta}_{\rm ext}\}$, where the intrinsic parameters $\vec{\theta}_{\rm int}$ describe the binary’s source-frame properties and the extrinsic parameters $\vec{\theta}_{\rm ext}$ describe the source configuration relative to the observer. In this work,
+The observed event rate is obtained by averaging the detection probability over the compact-binary parameters that determine the emitted signal and its projection onto a detector network. 
+
+Typically GW parameters can be $\{\vec{\theta}_{\rm int},\vec{\theta}_{\rm ext}\}$, where the intrinsic parameters $\vec{\theta}_{\rm int}$ describe the binary’s source properties and the extrinsic parameters $\vec{\theta}_{\rm ext}$ describe the source configuration relative to the observer (in a geocentric frame). In this work,
 
 $$ 
 \begin{equation}
 \begin{split}
-\vec{\theta} = \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{JL}, \iota, \phi, \psi, {\rm RA}, {\rm Dec}, t_c\} \,,
+\vec{\theta} = \{z_s, m_1, m_2, a_1, a_2, \theta_1, \theta_2, \phi_{12}, \phi_{\rm JL}, \iota, \phi, \psi, {\rm RA}, {\rm Dec}, t_c\} \,,
 \end{split}
 \end{equation}
 $$
 
-where source-frame component masses $(m_1,m_2)$, the dimensionless spin magnitudes $(a_1,a_2)$, and the spin-orientation angles $(\theta_1,\theta_2,\phi_{12},\phi_{JL})$ are intrinsic parameters, while the source redshift $z_s$, the sky position $({\rm RA},{\rm Dec})$, the inclination $\iota$, the polarization angle $\psi$, the coalescence phase $\phi$, and the coalescence time $t_c$ are extrinsic parameters.
[…]
