+++ docs/autoapi/ler/lens_galaxy_population/lens_galaxy_parameter_distribution/index.rst
@@ -189,7 +189,7 @@ Classes
    | :meth:`~sample_all_routine_epl_shear_sl`            | Sample EPL+shear lens parameters with strong   |
    |                                                     | lensing condition                              |
    +-----------------------------------------------------+------------------------------------------------+
-   | :meth:`~strongly_lensed_source_redshifts`           | Sample source redshifts with lensing condition |
+   | :meth:`~strongly_lensed_source_redshift`           | Sample source redshifts with lensing condition |
    +-----------------------------------------------------+------------------------------------------------+
 
    Instance Attributes
@@ -224,6 +224,33 @@ Classes
 
    ..
        !! processed by numpydoc !!
+   .. py:property:: source_redshift_sl
+
+      
+      Function to sample source redshifts conditioned on strong lensing.
+
+
+
+      :Returns:
+
+          **source_redshift_sl** : ``ler.functions.FunctionConditioning``
+              Function for sampling source redshifts conditioned on strong lensing.
+
+
+
+
+
+
+
+
+
+
+
+
+
+      ..
+          !! processed by numpydoc !!
+
    .. py:property:: normalization_pdf_z_lensed
 
       
@@ -375,10 +402,6 @@ Classes
 
       ..
           !! processed by numpydoc !!
-
-   .. py:attribute:: sample_source_redshift_sl
-
-      
 
    .. py:attribute:: sample_lens_parameters_routine
 
@@ -545,20 +568,26 @@ Classes
       ..
           !! processed by numpydoc !!
 
-   .. py:method:: strongly_lensed_source_redshifts(size=1000)
+   .. py:method:: strongly_lensed_source_redshift(size, get_attribute=False, **kwargs)
 
       
       Sample source redshifts conditioned on strong lensing.
 
-      Uses rejection sampling to generate source redshifts from the CBC source
-      population weighted by the optical depth, which increases with redshift.
 
       :Parameters:
 
           **size** : ``int``
-              Number of redshifts to sample.
+              Number of samples to generate.
 
               default: 1000
+
+          **get_attribute** : ``bool``
+              If True, returns the sampler object instead of samples.
+
+              default: False
+
+          **\*\*kwargs** : ``dict``
+              Additional parameters
 
       :Returns:
 
@@ -578,8 +607,8 @@ Classes
 
       >>> from ler.lens_galaxy_population import LensGalaxyParameterDistribution
       >>> lens = LensGalaxyParameterDistribution()
-      >>> zs = lens.strongly_lensed_source_redshifts(size=1000)
-      >>> print(f"Mean source redshift: {zs.mean():.2f}")
+      >>> zs = lens.strongly_lensed_source_redshift(size=1000)
+      >>> print(f"strongly lensed source redshift: {zs.mean():.2f}")
 
 
 
