+++ docs/analytical_formulation_lensed.md
@@ -1,321 +1,388 @@
 # Analytical Formulation for Lensed Gravitational Wave Event Rates
 
-Written by [Phurailatpam Hemantakumar](https://hemantaph.com).
+Written by [Phurailatpam Hemantakumar](https://hemantaph.com). Last updated: 2026-02-06.
 
 ## Overview
 
-This documentation extends the [Analytical Formulation for Gravitational Wave Event Rates](https://ler.hemantaph.com/analytical_formulation_unlensed.html) to account for strong gravitational lensing effects. We present a comprehensive framework for computing the rates of detectable lensed gravitational-wave events from compact-binary mergers. The formulation addresses key aspects of lensed systems: the optical depth describing lensing probability, the distributions of source and lens properties conditioned on strong lensing, the multi-image lensing cross-sections, and the detection prospects for lensed signals. This approach enables accurate predictions of observable lensed merger rates in current and future gravitational-wave detector networks.
+This document extends the framework in [Analytical Formulation for Gravitational Wave Event Rates](https://ler.hemantaph.com/analytical_formulation_unlensed.html) to incorporate strong gravitational lensing. The goal is to compute the rate of detectable lensed gravitational-wave (GW) events from compact-binary coalescences (CBCs) by combining the intrinsic merger population with a lens population model, multi-image lensing geometry, and a detection criterion applied to the lensed images. The formulation introduces the optical depth for strong lensing, conditional distributions for source and lens properties given strong lensing, the multi-image caustic cross-section for an EPL+Shear lens model, and the probability of detecting at least two lensed images in a detector network.
 
 
 ## Table of Contents
 
-1. [Introduction](#introduction)
-2. [Parameter-Marginalized Event Rate](#parameter-marginalized-event-rate)
-3. [Decomposing the Joint Source–Lens Parameter Distribution](#decomposing-the-joint-source–lens-parameter-distribution)
-4. [Source Redshift Distributions of Lensed Events](#source-redshift-distributions-of-lensed-events)
-5. [Optical Depth](#optical-depth)
-6. [Lens Redshift Distribution of Lensed Events](#lens-redshift-distribution-of-lensed-events)
-7. [Multi-Image Lensing Cross-Section](#multi-image-lensing-cross-section)
-8. [Lens Parameter Distributions of Lensed Events](#lens-parameter-distributions-of-lensed-events)
-9. [Source Position Distribution and Image Properties](#source-position-distribution-and-image-properties)
-10. [Detection Probability of Lensed Events](#detection-probability-of-lensed-events)
-11. [Complete Expression for Lensed Event Rate](#complete-expression-for-lensed-event-rate)
-12. [Lens Parameter Priors](#lens-parameter-priors)
[…]
