+++ ler/lens_galaxy_population/sampler_functions.py
@@ -71,7 +71,9 @@ def available_sampler_list():
 # ---------------------------------
 @njit
 def lens_redshift_strongly_lensed_sis_haris_pdf(
-    zl, zs, cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
+    zl,
+    zs,
+    cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0),
 ):
     """
     Compute lens redshift PDF for SIS model (Haris et al. 2018).
@@ -110,7 +112,11 @@ def lens_redshift_strongly_lensed_sis_haris_pdf(
 
 
 def lens_redshift_strongly_lensed_sis_haris_rvs(
-    size, zs, z_min=0.001, z_max=10.0, cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
+    size,
+    zs,
+    z_min=0.001,
+    z_max=10.0,
+    cosmo=LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0),
 ):
     """
     Sample lens redshifts for SIS model (Haris et al. 2018).
@@ -1218,12 +1224,44 @@ def importance_sampler(
     phi_rvs,
     gamma_rvs,
     shear_rvs,
-    sigma_pdf,
+    number_density,
     cross_section,
     n_prop,
 ):
     """
     Core importance sampling algorithm for lens parameters.
+
+    This function samples lens galaxy parameters weighted by their lensing
+    cross sections using importance sampling with a uniform proposal distribution
+    for velocity dispersion.
+
+    Algorithm
+    ---------
+    For each lens-source pair (zl_i, zs_i):
+
+    1. **Draw proposal samples** (n_prop samples per lens):
+       - sigma_k ~ Uniform(sigma_min, sigma_max)  with proposal density q(sigma) = 1/(sigma_max - sigma_min)
+       - q_k ~ p(q|sigma_k)  (axis ratio conditioned on velocity dispersion)
+       - φ_k ~ p(φ)  (orientation angle prior)
+       - gamma_k ~ p(gamma)  (power-law index prior)
+       - (gamma1_k, gamma2_k) ~ p(gamma1, gamma2)  (external shear prior)
+
+    2. **Compute lensing cross sections**:
+       - cs_k = CrossSection(zs_i, zl_i, sigma_k, q_k, φ_k, gamma_k, gamma1_k, gamma2_k)
+       - Normalize: cs_k ← cs_k / Σ_k cs_k
+
+    3. **Compute importance weights**:
+       - The target distribution is: p(θ|zl) ∝ p(sigma|zl) x CrossSection(θ)
+       - The proposal distribution is: q(θ) ∝ Uniform(sigma) x p(q|sigma) x p(φ) x p(gamma) x p(shear)
+       - Importance weight: w_k = cs_k x [p(sigma_k|zl) / q(sigma)]
+       - Normalize: w_k ← w_k / Σ_k w_k
+
+    4. **Resample**:
+       - Draw one sample index from {1, ..., n_prop} with probabilities {w_1, ..., w_n_prop}
+       - Return the corresponding parameter values as the posterior sample
+
+    The algorithm produces samples from the posterior distribution of lens
+    parameters weighted by their contribution to the strong lensing cross section.
 
     Parameters
     ----------
@@ -1243,8 +1281,8 @@ def importance_sampler(
[…]
