### 2026-02-13 d6888b3
fix: pin setuptools<81 to fix pkg_resources import for sphinx_tabs
