+++ docs/analytical_formulation_unlensed.md
@@ -91,7 +91,8 @@ The sampling priors and parameter ranges are summarized in [Table 1](#table1), w
           style="border:1px solid #ccc; border-radius:10px;"></iframe>
 </div>
 
-*Figure 1: GW coordinate system visualization (interactive). The figure shows three coordinate frames: the equatorial (observer) frame (blue), the wave frame (green), and the source frame (orbital plane) (yellow). In the equatorial frame, the $z$-axis points along Earth’s rotation axis and the $x$-axis points toward the vernal equinox. In the wave frame, the $z$-axis points along the GW propagation direction toward the geocentre, and the $(x,y)$ axes span the wave plane. In the source frame, the $z$-axis is aligned with the orbital angular momentum and the $(x,y)$ axes lie in the orbital plane. The inclination $\iota$ is the angle between the line of sight and the orbital angular momentum. The polarization angle $\psi$ is measured in the wave plane from the ascending line of nodes (the intersection of the equatorial plane with the wave plane) to the wave-frame $x$-axis, with the rotation taken about the wave-frame $z$-axis. See the `LALSimulation` polarization convention for details [Ref](https://lscsoft.docs.ligo.org/lalsuite/lalsimulation/group___l_a_l_simulation__h.html). The sky location is specified by $({\rm RA},{\rm Dec})$ in the equatorial frame, where ${\rm RA}$ is measured in the equatorial plane from the vernal equinox and ${\rm Dec}$ is measured from the equatorial plane toward the north celestial pole. The ring of test points in the wave plane illustrates the action of the $h_{+,\times}$ polarizations on the observed strain, wrt the wave-frame axes. The sizes and distances in the figure are not to scale, and the visualization is meant for qualitative understanding of the coordinate systems and angles rather than a quantitative representation of any particular system.*
[…]
