+++ docs/analytical_formulation_unlensed.md
@@ -91,7 +91,7 @@ The sampling priors and parameter ranges are summarized in [Table 1](#table1), w
           style="border:1px solid #ccc; border-radius:10px;"></iframe>
 </div>
 
-*Figure 1: Interactive visualization of the gravitational-wave coordinate system following the LALSimulation conventions. The figure shows three frames: the Equatorial (Observer) frame (blue), the Wave frame (green), and the Source (Orbital) frame (yellow). In the equatorial frame, the $z$ axis is aligned with Earth’s rotation axis and the $x$ axis points toward the vernal equinox; the sky position is specified by $({\rm RA},{\rm Dec})$, with ${\rm RA}$ measured in the equatorial plane from the vernal equinox and ${\rm Dec}$ measured from the equatorial plane toward the north celestial pole. In the wave frame, the $z$ axis is along the GW propagation direction toward the geocentre and the $(x,y)$ axes span the wave plane, with the wave-frame $x$ axis defining the polarization basis. In the source frame, the $z$ axis is aligned with the orbital angular momentum and the $(x,y)$ axes lie in the orbital plane; the inclination $\iota$ is the angle between the line of sight and the orbital angular momentum (defined at $f_{\rm ref}$ for precessing systems). The polarization angle $\psi$ is measured in the wave plane from the ascending line of nodes, defined by the intersection of the equatorial plane with the wave plane, to the wave-frame $x$ axis, with rotation about the wave-frame $z$ axis; see the LALSimulation polarization convention for details [Ref](https://lscsoft.docs.ligo.org/lalsuite/lalsimulation/group___l_a_l_simulation__h.html). The ring of test points in the wave plane illustrates the action of the $h_{+,\times}$ polarizations on the observed strain, wrt the wave-frame axes. The sizes and distances in the figure are not to scale, and the visualization is meant for qualitative understanding of the coordinate systems and angles rather than a quantitative representation of any particular system.*
[…]
