+++ docs/index.rst
@@ -12,7 +12,7 @@ Welcome to :red:`ler`'s documentation!
 
 ``ler`` (/ˈɛlɚ/) is a statistical-based Python package whose core function is to calculate detectable rates of gravitational wave events (both lensed and unlensed). It provides a comprehensive framework for simulating compact binary coalescence events and their detection by ground-based gravitational wave detectors.
 
-``ler`` is closely integrated with the ``gwsnr`` package (`see gwsnr documentation <https://gwsnr.hemantaph.com>`_), which provides efficient gravitational-wave Signal-to-Noise Ratio calculation. This allows researchers to perform fast $P_{\rm det}$ calculations for population studies and hierarchical Bayesian inference with selection effects.
+``ler`` is closely integrated with the ``gwsnr`` package (`see gwsnr documentation <https://gwsnr.hemantaph.com>`_), which provides efficient gravitational-wave Signal-to-Noise Ratio calculation. This allows researchers to perform fast Probability of Detection (:math:`P_{\text{det}}`) calculations for population studies and hierarchical Bayesian inference with selection effects.
 
 For a detailed technical overview, see the :doc:`Summary` section, and browse the topics under :blue:`CONTENTS` in the sidebar (or top-left menu) of your browser.
 
