+++ docs/analytical_formulation_lensed.md
@@ -511,7 +511,7 @@ An illustrative example of a multi-image configuration for an EPL+Shear lens is 
   <img src="_static/lens_configuration.jpg" alt="Merger rate density and PDF of redshift for BBH mergers" width="600"/>
 </div>
 
->**Figure 2:** Multi-image configuration for an EPL+Shear lens, showing the outer double-image caustic (orange dashed) and the inner quad-image (diamond) caustic (green solid) in the source plane. The Einstein ring (blue, dotted) serves as the angular scale reference, with $\theta_E(z_s=3.0, z_L=0.8, \sigma=160\,{\rm km/s}) \approx 2 \times 10^{-6}$ rad. The lens geometry is defined by an axis ratio $q=0.6$, orientation angle $\phi=0.52$ rad, mass density slope $\gamma=1.84$, and external shear components $\gamma_1=\gamma_2=-0.05$. A source located at $(\beta_x, \beta_y) = (-0.25, 0.04)$ lies inside the diamond caustic near a fold boundary, generating four lensed images (stars). These are labeled by index $i$ (ordered by arrival time, where $i=1$ arrives first), image type, and absolute magnification $|\mu_i|$. The caustic structure is governed by the lens parameters where $q$ determines ellipticity and $\phi$ controls counter-clockwise rotation relative to the x-axis. Increasing in $\gamma$ cause the outer caustic to expand rapidly while the inner diamond caustic shrinks, whereas shear components induce skewness and rotation. An interactive demonstration of these dependencies is provided in the [example notebook](https://colab.research.google.com/drive/1X1XDID3riGJhMgVBUueq-W72wOiSG3r9?usp=sharing). Notably, for the chosen configuration, the proximity of the source to the caustic fold results in high-magnification images which can significantly enhance the detectability of such gravitational lensing systems.
+>**Figure 2:** Multi-image configuration for an EPL+Shear lens, showing the outer double-image caustic (orange dashed) and the inner quad-image (diamond) caustic (green solid) in the source plane. The Einstein ring (blue, dotted) serves as the angular scale reference, with $\theta_E(z_s=3.0, z_L=0.8, \sigma=160\,{\rm km/s}) \approx 2 \times 10^{-6}$ rad. The lens geometry is defined by an axis ratio $q=0.6$, orientation angle $\phi=0.52$ rad, mass density slope $\gamma=1.84$, and external shear components $\gamma_1=\gamma_2=-0.05$. A source located at $(\beta_x, \beta_y) = (-0.25, 0.04)$ lies inside the diamond caustic near a fold boundary, generating four lensed images (stars). These are labeled by index $i$ (ordered by arrival time, where $i=1$ arrives first), image type, time-delay wrt the first image $\Delta t_{1,i}$ , and absolute magnification $|\mu_i|$. The caustic structure is governed by the lens parameters where $q$ determines ellipticity and $\phi$ controls counter-clockwise rotation relative to the x-axis. Increasing in $\gamma$ cause the outer caustic to expand rapidly while the inner diamond caustic shrinks, whereas shear components induce skewness and rotation. An interactive demonstration of these dependencies is provided in the [example notebook](https://colab.research.google.com/drive/1X1XDID3riGJhMgVBUueq-W72wOiSG3r9?usp=sharing). Notably, for the chosen configuration, the proximity of the source to the caustic fold results in high-magnification images which can significantly enhance the detectability of such gravitational lensing systems.
 
 
 ## Detection Probability of Lensed Events
