### 2026-02-19 327450f
Fix sphinx build: ensure setuptools and sphinx are properly installed
