+++ requirements_sphinx.txt
@@ -1,9 +1,9 @@
+setuptools>=65.0.0
+sphinx>=7.0.0
 html5lib==1.0b8
-setuptools<81.0.0
 nbsphinx==0.9.3
 sphinx-autoapi==3.0.0
 autoapi==2.0.1
-nbsphinx==0.9.3
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme==3.0.2
 numpydoc==1.6.0
