### 2026-02-19 9f3bac0
Fix: install setuptools before sphinx deps to provide pkg_resources
