+++ .github/workflows/documentation.yml
@@ -36,9 +36,12 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y pandoc
+      - name: Install setuptools
+        run: |
+          pip install "setuptools>=65.0.0"
       - name: Install Python dependencies
         run: |
-          pip install -r requirements_sphinx.txt  # Or pip install sphinx sphinx-rtd-theme
+          pip install -r requirements_sphinx.txt
       - name: Install project dependencies
         run: |
           pip install -e .
