### 2026-02-19 e78c567
Fix: upgrade sphinx-tabs to drop pkg_resources dependency
