+++ requirements_sphinx.txt
@@ -7,7 +7,7 @@ autoapi==2.0.1
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme==3.0.2
 numpydoc==1.6.0
-sphinx-tabs==3.4.1
+sphinx-tabs>=3.4.7
 sphinx-jupyterbook-latex==1.0.0
 myst-parser==3.0.1
 linkify-it-py==2.0.3
