### 2026-02-19 77f6c32
Use uv for faster dependency installation in docs workflow
