+++ .github/workflows/documentation.yml
@@ -36,15 +36,17 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y pandoc
+      - name: Install uv
+        uses: astral-sh/setup-uv@v4
       - name: Install setuptools
         run: |
-          pip install "setuptools>=65.0.0"
+          uv pip install --system "setuptools>=65.0.0"
       - name: Install Python dependencies
         run: |
-          pip install -r requirements_sphinx.txt
+          uv pip install --system -r requirements_sphinx.txt
       - name: Install project dependencies
         run: |
-          pip install -e .
+          uv pip install --system -e .
       - name: Build Sphinx documentation
         run: |
           cd docs
