+++ requirements_sphinx.txt
@@ -1,13 +1,13 @@
-setuptools>=65.0.0
-sphinx>=7.0.0
 html5lib==1.0b8
+setuptools
 nbsphinx==0.9.3
-sphinx-autoapi==3.0.0
+sphinx-autoapi==3.6.1
 autoapi==2.0.1
+nbsphinx==0.9.3
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme==3.0.2
 numpydoc==1.6.0
-sphinx-tabs>=3.4.7
+sphinx-tabs==3.4.7
 sphinx-jupyterbook-latex==1.0.0
 myst-parser==3.0.1
 linkify-it-py==2.0.3
