+++ docs/index.rst
@@ -77,8 +77,8 @@ Glossary
       .. raw:: html
 
          <div style="text-align:center;">
-         <img src="_static/gw.gif" width="480px" alt="Animation of GW propagation">
-          <div style="text-align:left; max-width:480px; margin: 0 auto;">
+         <img src="_static/gw.gif" width="90%" alt="Animation of GW propagation">
+          <div style="text-align:left; max-width:90%; margin: 0 auto;">
             <p style="font-size: 0.9em; font-family: Arial, sans-serif; line-height: 1.5em;">
                Animation showing the propagation of gravitational waves from inspiraling binary black holes. As the waves travel, they stretch and squeeze spacetime in their path. <em>Source: <a href="https://community.wolfram.com/groups/-/m/t/790989">Jeffrey Bryant, Wolfram | Alpha, LLC.</a>.</em>
             </p>
@@ -106,8 +106,8 @@ Glossary
       .. raw:: html
 
          <div style="text-align:center;">
-         <img src="_static/Light_Bending.gif" width="480px" alt="Animation of light bending">
-          <div style="text-align:left; max-width:480px; margin: 0 auto;">
+         <img src="_static/Light_Bending.gif" width="90%" alt="Animation of light bending">
+          <div style="text-align:left; max-width:90%; margin: 0 auto;">
             <p style="font-size: 0.9em; font-family: Arial, sans-serif; line-height: 1.5em;">
                Animation showing the bending of light by a massive object. <em>Source: <a href="https://science.nasa.gov/universe/how-gravity-warps-light/">NASA, ESA, and Goddard Space Flight Center/K. Jackson.</a>.</em>
             </p>
@@ -121,8 +121,8 @@ Glossary
       .. raw:: html
 
          <div style="text-align:center;">
-         <img src="_static/BBH_Merger_rate_density_and_PDF_of_redshift.png" width="480px" alt="BBH Merger rate density">
-          <div style="text-align:left; max-width:480px; margin: 0 auto;">
+         <img src="_static/Merger_rate_density_and_PDF_of_redshift.png" width="90%" alt="BBH Merger rate density">
+          <div style="text-align:left; max-width:90%; margin: 0 auto;">
             <p style="font-size: 0.9em; font-family: Arial, sans-serif; line-height: 1.5em;">
                Figure showing the merger rate density of binary black hole (BBH) mergers as a function of redshift. <em>Source: <a href="https://ler.hemantaph.comGW_events.html">generated by the LeR package.</a>.</em>
             </p>
