+++ .github/workflows/documentation.yml
@@ -36,17 +36,12 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y pandoc
-      - name: Install uv
-        uses: astral-sh/setup-uv@v4
-      - name: Install setuptools
-        run: |
-          uv pip install --system "setuptools>=65.0.0"
       - name: Install Python dependencies
         run: |
-          uv pip install --system -r requirements_sphinx.txt
+          pip install -r requirements_sphinx.txt  # Or pip install sphinx sphinx-rtd-theme
       - name: Install project dependencies
         run: |
-          uv pip install --system -e .
+          pip install -e .
       - name: Build Sphinx documentation
         run: |
           cd docs
