+++ ler/image_properties/image_properties.py
@@ -19,25 +19,28 @@ Copyright (C) 2026 Phurailatpam Hemanta Kumar. Distributed under MIT License.
 """
 
 import warnings
+
 warnings.filterwarnings("ignore")
 import logging
-logging.getLogger('numexpr.utils').setLevel(logging.ERROR)
+
+logging.getLogger("numexpr.utils").setLevel(logging.ERROR)
 # for multiprocessing
-from multiprocessing import Pool
+import multiprocessing as mp
 from tqdm import tqdm
 
 import numpy as np
 from numba import njit
 
 from astropy.cosmology import LambdaCDM
+
 cosmo = LambdaCDM(H0=70, Om0=0.3, Ode0=0.7, Tcmb0=0.0, Neff=3.04, m_nu=None, Ob0=0.0)
 
 # the following .py file will be called if they are not given in the class initialization
-from .multiprocessing_routine import solve_lens_equation
+from .multiprocessing_routine import solve_lens_equation, _init_worker_multiprocessing
 from ..lens_galaxy_population.lens_functions import phi_q2_ellipticity
 
 
-class ImageProperties():
+class ImageProperties:
     """
     Class to compute image properties of strongly lensed gravitational wave events.
 
@@ -65,7 +68,7 @@ class ImageProperties():
     time_window : ``float``
         Time window for lensed events from min(geocent_time) (units: seconds). \n
         default: 365*24*3600*2 (2 years)
-    effective_params_in_output : ``bool``
+    include_effective_parameters : ``bool``
         Whether to include effective parameters (effective_phase, effective_ra, effective_dec) in the output. \n
         default: True
     lens_model_list : ``list``
@@ -82,6 +85,12 @@ class ImageProperties():
         If True (and spin_zero=False), sample precessing spin parameters. \n
         If False (and spin_zero=False), sample aligned/anti-aligned spins. \n
         default: False
+    multiprocessing_verbose : ``bool``
+        If True, shows a progress bar for multiprocessing tasks. \n
+        default: True
+    include_redundant_parameters : ``bool``
+        If True, removes redundant parameters (e.g., theta_E, n_images, mass_1, mass_2, luminosity_distance) from output to save memory. \n
+    
 
     Examples
     --------
@@ -101,7 +110,7 @@ class ImageProperties():
     ... )
     >>> result = ip.image_properties(lens_parameters)
     >>> print(result.keys())
-    
+
 
     Instance Methods
     ----------
@@ -117,46 +126,53 @@ class ImageProperties():
     Instance Attributes
     ----------
     ImageProperties has the following attributes: \n
-    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-    | Attribute                                           | Type                      | Unit     | Description                                    |
-    +=====================================================+===========================+==========+================================================+
-    | :attr:`~npool`                                      | ``int``                   |          | Number of multiprocessing workers              |
-    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-    | :attr:`~n_min_images`                               | ``int``                   |          | Minimum number of images required              |
-    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-    | :attr:`~n_max_images`                               | ``int``                   |          | Maximum number of images per event             |
-    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-    | :attr:`~time_window`                                | ``float``                 | s        | Time window for lensed events                  |
-    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-    | :attr:`~effective_params_in_output`                 | ``bool``                  |          | To include effective parameters in output      |
-    +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
[…]
