+++ ler/rates/ler.py
@@ -293,18 +293,23 @@ class LeR(LensGalaxyParameterDistribution):
 
     def _parent_class_initialization(self, params=None, pdet_finder=None, verbose=True):
         """
-        Function to initialize the parent classes.
+        Helper function to initialize the parent classes.
 
         Parameters
         ----------
-        params : ``dict``
-            dictionary of parameters to initialize the parent classes
+        params : ``dict`` or ``None``
+            Dictionary of parameters to initialize the parent classes.
+        pdet_finder : ``callable`` or ``None``
+            Custom detection probability finder function.
+        verbose : ``bool``
+            If True, print initialization parameters. \n
+            default: True
         """
 
         def initialization():
             # initialization of parent class
             self._parent_initialization_helper(params=params)
-            # initialization self.snr and self.pdet_finder from GWSNR class
+            # initialization of self.pdet_finder from gwsnr or from custom function
             if not pdet_finder:
                 self.pdet_finder = self._gwsnr_initialization(params=params)
             else:
@@ -327,8 +332,8 @@ class LeR(LensGalaxyParameterDistribution):
 
         Parameters
         ----------
-        params : ``dict``
-            dictionary of parameters to initialize the parent classes
+        params : ``dict`` or ``None``
+            Dictionary of parameters to initialize the parent classes.
         """
 
         # initialization of LensGalaxyParameterDistribution class
@@ -352,7 +357,9 @@ class LeR(LensGalaxyParameterDistribution):
             n_max_images=4,
             time_window=365 * 24 * 3600 * 2,
             lens_model_list=["EPL_NUMBA", "SHEAR"],
-            effective_params_in_output=True,
+            include_effective_parameters=False,
+            multiprocessing_verbose=True,
+            include_redundant_parameters=False,
             # CBCSourceParameterDistribution class params
             event_type=self.event_type,
             source_priors=None,
@@ -386,8 +393,10 @@ class LeR(LensGalaxyParameterDistribution):
             n_min_images=input_params["n_min_images"],
             n_max_images=input_params["n_max_images"],
             time_window=input_params["time_window"],
-            effective_params_in_output=input_params["effective_params_in_output"],
+            include_effective_parameters=input_params["include_effective_parameters"],
             lens_model_list=input_params["lens_model_list"],
+            multiprocessing_verbose=input_params["multiprocessing_verbose"],
+            include_redundant_parameters=input_params["include_redundant_parameters"],
             # CBCSourceParameterDistribution class params
             event_type=input_params["event_type"],
             source_priors=input_params["source_priors"],
@@ -443,8 +452,8 @@ class LeR(LensGalaxyParameterDistribution):
                 snr_th_net=10.0,
                 pdet_type="boolean",
                 distribution_type="noncentral_chi2",
-                include_optimal_snr=True,
-                include_observed_snr=True,
+                include_optimal_snr=False,
+                include_observed_snr=False,
             ),
             # Settings for interpolation grid
             mtot_min=min_bh_mass * 2,
@@ -559,7 +568,7 @@ class LeR(LensGalaxyParameterDistribution):
 
     def _print_all_init_args(self):
         """
-        Function to print all the parameters.
+        Helper function to print all initialization parameters.
         """
 
         # print all relevant functions and sampler priors
@@ -657,8 +666,10 @@ class LeR(LensGalaxyParameterDistribution):
         print(f"    n_min_images = {self.ler_args['n_min_images']},")
         print(f"    n_max_images = {self.ler_args['n_max_images']},")
         print(f"    time_window = {self.ler_args['time_window']},")
-        print(f"    effective_params_in_output = {self.ler_args['effective_params_in_output']},")
[…]
