+++ ler/_version.py
@@ -1,2 +1,2 @@
 # Version information for ler package
-__version__ = "0.5.4"
+__version__ = "0.5.9"

[…]
