+++ docs/autoapi/ler/lens_galaxy_population/index.rst
@@ -2821,7 +2821,7 @@ Attributes
               +--------------------+--------------+--------------------------------------+
               | a_1                |              | spin_1 of the compact binary         |
               +--------------------+--------------+--------------------------------------+
-              | a_2                |              | spin_2 of the compact binary         |
+              | a_2                |              | spin of the secondary compact binary |
               +--------------------+--------------+--------------------------------------+
               | luminosity_distance| Mpc          | luminosity distance                  |
               +--------------------+--------------+--------------------------------------+
@@ -3375,7 +3375,7 @@ Attributes
           !! processed by numpydoc !!
 
 
-.. py:class:: ImageProperties(npool=4, n_min_images=2, n_max_images=4, lens_model_list=['EPL_NUMBA', 'SHEAR'], cosmology=None, time_window=365 * 24 * 3600 * 2, spin_zero=True, spin_precession=False, pdet_finder=None, effective_params_in_output=True)
+.. py:class:: ImageProperties(npool=4, n_min_images=2, n_max_images=4, lens_model_list=['EPL_NUMBA', 'SHEAR'], cosmology=None, time_window=365 * 24 * 3600 * 2, spin_zero=True, spin_precession=False, pdet_finder=None, include_effective_parameters=False, multiprocessing_verbose=True, include_redundant_parameters=False)
 
 
    
@@ -3417,7 +3417,7 @@ Attributes
 
            default: 365*24*3600*2 (2 years)
 
-       **effective_params_in_output** : ``bool``
+       **include_effective_parameters** : ``bool``
            Whether to include effective parameters (effective_phase, effective_ra, effective_dec) in the output.
 
            default: True
@@ -3445,6 +3445,14 @@ Attributes
            If False (and spin_zero=False), sample aligned/anti-aligned spins.
 
            default: False
+
+       **multiprocessing_verbose** : ``bool``
+           If True, shows a progress bar for multiprocessing tasks.
+
+           default: True
+
+       **include_redundant_parameters** : ``bool``
+           If True, removes redundant parameters (e.g., theta_E, n_images, mass_1, mass_2, luminosity_distance) from output to save memory.
 
 
 
@@ -3491,31 +3499,35 @@ Attributes
    ----------
    ImageProperties has the following attributes:
 
-   +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-   | Attribute                                           | Type                      | Unit     | Description                                    |
-   +=====================================================+===========================+==========+================================================+
-   | :attr:`~npool`                                      | ``int``                   |          | Number of multiprocessing workers              |
-   +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-   | :attr:`~n_min_images`                               | ``int``                   |          | Minimum number of images required              |
-   +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-   | :attr:`~n_max_images`                               | ``int``                   |          | Maximum number of images per event             |
-   +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-   | :attr:`~time_window`                                | ``float``                 | s        | Time window for lensed events                  |
-   +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-   | :attr:`~effective_params_in_output`                 | ``bool``                  |          | To include effective parameters in output      |
-   +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-   | :attr:`~lens_model_list`                            | ``list``                  |          | List of lens models                            |
-   +-----------------------------------------------------+---------------------------+----------+------------------------------------------------+
-   | :attr:`~cosmo`                                      | ``astropy.cosmology``     |          | Cosmology for calculations                     |
[…]
