+++ docs/analytical_formulation_lensed.md
@@ -23,7 +23,7 @@ This document extends the framework in [Analytical Formulation for Gravitational
 - [Simulation Results](#simulation-results)
   - [Simulation Settings](#simulation-settings)
   - [Lens and source Parameter Priors](#lens-and-source-parameter-priors)
-  - [Plot Comparison for Lensed+Detectable, Lensed and Intrinsic Populations](#plot-comparison-for-lenseddetectable-lensed-and-intrinsic-populations)
+  - [Plot Comparison for Lensed+Detectable, Lensed and Intrinsic Populations](#plot-comparison-for-lensed-detectable-lensed-and-intrinsic-populations)
   - [Biases in the inferred GW parameters](#biases-in-the-inferred-gw-parameters)
   - [Time Delays and Magnifications of Detected Lensed Events](#time-delays-and-magnifications-of-detected-lensed-events)
   - [Rate Estimates and Comparisons](#rate-estimates-and-comparisons)
@@ -669,7 +669,7 @@ When a strongly lensed gravitational wave is detected without being identified a
   <img src="_static/mc_vs_zs.png" alt="Corner plot comparing inferred source redshift and chirp mass for lensed and unlensed events" width="450"/>
 </div>
 
->**Figure 5:** Joint kernel density estimate (KDE) with 10, 40, 68, and 95 percent contour levels and marginal distributions comparing the inferred source redshift $z^*_s$ and inferred source-frame chirp mass ${\cal M}_c^{* \, {\rm src}}$ for the detectable unlensed (blue) and lensed (orange) populations. While the unlensed parameters reflect the true source properties, the inferred parameters for lensed events are systematically biased by magnification. The inferred redshift $z^*_s$ is derived from the effective luminosity distance $d_{L}^{\rm eff} = d_{L} / \sqrt{|\mu|}$, which subsequently affects the source-frame mass reconstruction via ${\cal M}_c^{* \, {\rm src}} = {\cal M}_c / (1+z^*_s)$, where ${\cal M}_c$ is the detector-frame chirp mass. Consequently, the lensed population exhibits a distinct shift toward higher inferred redshifts—though these still underestimate the true source redshifts of lensed events—and significantly higher inferred chirp masses relative to the unlensed population. This magnification bias demonstrates that unrecognized lensed events can contaminate the observed catalog by mimicking heavy binary black hole mergers and appearing as high-mass outliers; however, given the low expected rate of lensed events relative to unlensed ones ($\approx$ 1:3000 at O4 design sensitivity), this contamination is unlikely to significantly skew population studies or cosmological inferences even if unaccounted for.
[…]
