+++ docs/autoapi/ler/gw_source_population/cbc_source_parameter_distribution/index.rst
@@ -1024,7 +1024,7 @@ Attributes
               +--------------------+--------------+--------------------------------------+
               | a_1                |              | spin_1 of the compact binary         |
               +--------------------+--------------+--------------------------------------+
-              | a_2                |              | spin_2 of the compact binary         |
+              | a_2                |              | spin of the secondary compact binary |
               +--------------------+--------------+--------------------------------------+
               | luminosity_distance| Mpc          | luminosity distance                  |
               +--------------------+--------------+--------------------------------------+
