+++ docs/autoapi/ler/gw_source_population/index.rst
@@ -2967,7 +2967,7 @@ Attributes
               +--------------------+--------------+--------------------------------------+
               | a_1                |              | spin_1 of the compact binary         |
               +--------------------+--------------+--------------------------------------+
-              | a_2                |              | spin_2 of the compact binary         |
+              | a_2                |              | spin of the secondary compact binary |
               +--------------------+--------------+--------------------------------------+
               | luminosity_distance| Mpc          | luminosity distance                  |
               +--------------------+--------------+--------------------------------------+

[…]
