### 2026-02-24 daf3c33
github workflow with uv
