+++ paper/paper.bib
@@ -1,3 +1,71 @@
+@misc{Haris2018,
+      title={Identifying strongly lensed gravitational wave signals from binary black hole mergers}, 
+      author={K. Haris and Ajit Kumar Mehta and Sumit Kumar and Tejaswi Venumadhav and Parameswaran Ajith},
+      year={2018},
+      eprint={1807.07062},
+      archivePrefix={arXiv},
+      primaryClass={gr-qc},
+      url={https://arxiv.org/abs/1807.07062}, 
+}
+
+@misc{Phurailatpam2025,
+      title={gwsnr: A python package for efficient signal-to-noise calculation of gravitational-waves}, 
+      author={Hemantakumar Phurailatpam and Otto Akseli Hannuksela},
+      year={2025},
+      eprint={2412.09888},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.IM},
+      url={https://arxiv.org/abs/2412.09888}, 
+}
+
+@article{Ng2025,
+   title={Uncovering faint lensed gravitational-wave signals and reprioritizing their follow-up analysis using galaxy lensing forecasts with detected counterparts},
+   volume={540},
+   ISSN={1365-2966},
+   url={http://dx.doi.org/10.1093/mnras/staf874},
+   DOI={10.1093/mnras/staf874},
+   number={4},
+   journal={Monthly Notices of the Royal Astronomical Society},
+   publisher={Oxford University Press (OUP)},
+   author={Ng, Leo C Y and Janquart, Justin and Phurailatpam, Hemantakumar and Narola, Harsh and Poon, Jason S C and Van Den Broeck, Chris and Hannuksela, Otto A},
+   year={2025},
+   month=may, pages={2937–2951} 
+}
+
+@misc{More&Phurailatpam2025,
+      title={Gravitational lensing: towards combining the multi-messengers}, 
+      author={Anupreeta More and Hemantakumar Phurailatpam},
+      year={2025},
+      eprint={2502.02536},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.CO},
+      url={https://arxiv.org/abs/2502.02536}, 
+}
+
+@misc{Hannuksela2025,
+      title={Strong gravitational-wave lensing posterior odds}, 
+      author={Otto A. Hannuksela and K. Haris and Justin Janquart and Harsh Narola and Hemantakumar Phurailatpam and Jolien D. E. Creighton and Chris Van Den Broeck},
+      year={2025},
+      eprint={2510.15463},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.HE},
+      url={https://arxiv.org/abs/2510.15463}, 
+}
+
+
+@article{Ng2018,
+   title={Precise LIGO lensing rate predictions for binary black holes},
+   volume={97},
+   ISSN={2470-0029},
+   url={http://dx.doi.org/10.1103/PhysRevD.97.023012},
+   DOI={10.1103/physrevd.97.023012},
+   number={2},
+   journal={Physical Review D},
+   publisher={American Physical Society (APS)},
+   author={Ng, Ken K. Y. and Wong, Kaze W. K. and Broadhurst, Tom and Li, Tjonnie G. F.},
+   year={2018},
+   month=jan }
+
 @article{Janquart2023,
[…]
