+++ paper/paper.md
@@ -66,63 +66,136 @@ bibliography: paper.bib
 
 # Summary
 
-Gravitational waves (GWs) are ripples in the fabric of space and time caused by the acceleration of unevenly distributed mass or masses. Observable GWs are created especially during the violent cosmic events of merging compact binaries, such as 'binary black holes' (BBHs), 'binary neutron stars' (BNSs), and 'neutron star and black hole pair' (NSBHs). GWs emitted by these events can be distorted or magnified by the gravitational fields of massive objects such as galaxies or galaxy clusters, a phenomenon known as gravitational lensing. Profound comprehension of gravitational lensing's impact on GW signals is imperative to their accurate interpretation and the extraction of astrophysical insights therein. For this purpose, statistical modelling of GWs lensing can provide valuable insights into the properties of the lensing objects and GW sources. Such statistics require accurate and efficient means to calculate the detectable lensing rates, which depend on up-to-date modeling and implementation of lens and source properties and their distribution. The outcomes of these computational analyses not only contribute to generating dependable forecasts but also play an important role in validating forthcoming lensing events [@Janquart2023] [@ligolensing2023] [@Abbott2021].
+<!-- Summary should include: A description of the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
 
-Obtaining precise outcomes in statistical analyses of this nature necessitates the utilization of large-scale sampling, often numbering in the millions. However, this process is computationally demanding. The $ler$ framework addresses this by employing innovative techniques to optimize the workflow and computation efficiency required for handling large-scale statistical analyses, essential for modeling detectable events and calculating rates. Its integration of modular statistical components enhances the framework's adaptability and extendability, thus proving to be an invaluable asset in the evolving field of gravitational wave research. Detailed description, source code, and examples are available in $ler$ [documentation](https://ler.hemantaph.com).
+<!-- domain knowledge 1 -->
+<!-- Gravitational waves (GWs) are ripples in the fabric of space and time caused by the acceleration of unevenly distributed mass or masses. Observable GWs are created especially during the violent cosmic events of merging compact binaries, such as 'binary black holes' (BBHs), 'binary neutron stars' (BNSs), and 'neutron star and black hole pair' (NSBHs). -->
+<!-- domain knowledge 2 -->
+<!-- As these signals travel across the universe, the vast majority reach our detectors unobstructed, forming the baseline (unlensed) GW population.
+However, a fraction (<0.1%) of these signals encounter massive objects (lenses) like galaxies or galaxy clusters along their path, which can split them into multiple images, that are magnified (de-magnified), time and phase shifted, through strong gravitational lensing [Wierda2021] [@Abbott2021]. -->
+<!-- narrower problem. What's the current state. Purpose of the work -->
+<!-- A profound comprehension and accurate modelling of both the unlensed population and lens population and the lensing effects is imperative for accurate astrophysical interpretation, inferring merger rate histories, and validating forthcoming lensing events [@Janquart2023] [@ligolensing2023] [@Abbott2021] [@Wierda2021]. -->
+
+Gravitational waves are spacetime ripples caused by accelerating massive objects, primarily observable from the violent cosmic mergers of compact binaries like 'binary black holes' (BBHs), 'binary neutron stars' (BNSs), and 'neutron star and black hole pairs' (NSBHs). While most of these signals reach our detectors unobstructed to form the baseline unlensed population, a small fraction encounter massive intervening objects such as galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott2021] [@Wierda2021].
+
+<!-- what does software do? high level functionality. -->
[…]
