+++ .github/workflows/documentation.yml
@@ -34,7 +34,6 @@ jobs:
           python-version: '3.11'
       - name: Install system dependencies
         run: |
-          sudo apt-get update
           sudo apt-get install -y pandoc
       - name: Install uv
         run: |
