+++ requirements_sphinx.txt
@@ -1,3 +1,4 @@
+sphinx==8.1.3
 html5lib==1.0b8
 setuptools==80.9.0
 nbsphinx==0.9.3
