+++ .github/workflows/documentation.yml
@@ -2,69 +2,63 @@
 name: Deploy Sphinx Docs to Pages
 
 on:
-  # Runs on pushes targeting the default branch
   push:
-    branches: ["main"] # Or "master", depending on your repo
-
-  # Allows you to run this workflow manually from the Actions tab
+    branches: ["main"] 
   workflow_dispatch:
 
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
 permissions:
   contents: read
   pages: write
   id-token: write
 
-# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
-# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
 concurrency:
   group: "pages"
   cancel-in-progress: false
 
 jobs:
-  # Build job
   build:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout
+      - name: Checkout repository
         uses: actions/checkout@v4
+
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
           python-version: '3.11'
-      - name: Install system dependencies
+
+      - name: Install system dependencies (Pandoc for nbsphinx)
         run: |
+          sudo apt-get update
           sudo apt-get install -y pandoc
-      - name: Install uv
+
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v5
+        with:
+          enable-cache: true
+
+      - name: Install package and documentation dependencies
         run: |
-          pip install --upgrade pipx
-          pipx install uv
-      - name: Install Python dependencies with uv
-        run: |
-          uv pip install --upgrade sphinx
-          uv pip install -r requirements_sphinx.txt
-      - name: Install project dependencies with uv
-        run: |
-          uv add ler
+          uv pip install --system -e ".[docs]"
+
       - name: Build Sphinx documentation
         run: |
           cd docs
-          rm -r autoapi/*
+          rm -rf autoapi/ 
           make clean
           make html
-      - name: List build directory contents
+
+      - name: Verify build directory contents
         run: |
           ls -la docs/
           ls -la docs/_build/ || echo "_build directory not found"
           ls -la docs/_build/html/ || echo "html directory not found"
-      - name: Upload artifact
+
+      - name: Upload GitHub Pages artifact
         uses: actions/upload-pages-artifact@v3
         with:
-          # Upload entire directory
-          path: './docs/_build/html' # Correct path based on Makefile
-  
+          path: './docs/_build/html'
 
-  # Deployment job
   deploy:
     environment:
       name: github-pages
