+++ pyproject.toml
@@ -54,7 +54,6 @@ docs = [
     "nbsphinx==0.9.3",
     "sphinx-autoapi==3.0.0",
     "autoapi==2.0.1",
-    "nbsphinx==0.9.3",
     "sphinx-copybutton==0.5.2",
     "sphinx-rtd-theme==3.0.2",
     "numpydoc==1.6.0",
