+++ pyproject.toml
@@ -47,10 +47,10 @@ dev = [
     "ruff>=0.1.0",
 ]
 docs = [
-    "sphinx>=7.0",
+    "sphinx==8.1.3",
     "sphinx-autodoc-typehints>=1.23",
     "html5lib==1.0b8",
-    "setuptools",
+    "setuptools==80.9.0",
     "nbsphinx==0.9.3",
     "sphinx-autoapi==3.0.0",
     "autoapi==2.0.1",
