+++ pyproject.toml
@@ -47,21 +47,21 @@ dev = [
     "ruff>=0.1.0",
 ]
 docs = [
-    "sphinx==8.1.3",
+    "sphinx>=8.1.3",
     "sphinx-autodoc-typehints>=1.23",
-    "html5lib==1.0b8",
-    "setuptools==80.9.0",
-    "nbsphinx==0.9.3",
-    "sphinx-autoapi==3.0.0",
-    "autoapi==2.0.1",
-    "sphinx-copybutton==0.5.2",
-    "sphinx-rtd-theme==3.0.2",
-    "numpydoc==1.6.0",
-    "sphinx-tabs==3.4.1",
-    "sphinx-jupyterbook-latex==1.0.0",
-    "myst-parser==3.0.1",
-    "linkify-it-py==2.0.3",
-    "sphinx-rtd-dark-mode==1.3.0",
+    "html5lib>=1.0b8",
+    "setuptools>=80.9.0",
+    "nbsphinx>=0.9.3",
+    "sphinx-autoapi>=3.0.0",
+    "autoapi>=2.0.1",
+    "sphinx-copybutton>=0.5.2",
+    "sphinx-rtd-theme>=3.0.2",
+    "numpydoc>=1.6.0",
+    "sphinx-tabs>=3.4.7", # Updated to fix docutils conflict
+    "sphinx-jupyterbook-latex>=1.0.0",
+    "myst-parser>=3.0.1",
+    "linkify-it-py>=2.0.3",
+    "sphinx-rtd-dark-mode>=1.3.0",
 ]
 
 [tool.setuptools]
