+++ ler/gw_source_population/cbc_source_parameter_distribution.py
@@ -114,7 +114,7 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
     +=====================================================+================================================+
     | :meth:`~sample_gw_parameters`                       | Sample all GW parameters for compact binaries  |
     +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~binary_masses_BBH_powerlaw_gaussian`| Sample BBH masses with PowerLaw+PEAK model     |
+    | :meth:`~binary_masses_BBH_powerlaw_gaussian`.       | Sample BBH masses with PowerLaw+PEAK model     |
     +-----------------------------------------------------+------------------------------------------------+
     | :meth:`~binary_masses_BBH_popIII_lognormal`         | Sample pop III BBH masses from lognormal       |
     +-----------------------------------------------------+------------------------------------------------+
