+++ requirements_sphinx.txt
@@ -8,7 +8,7 @@ nbsphinx==0.9.3
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme==3.0.2
 numpydoc==1.6.0
-sphinx-tabs==3.4.7
+sphinx-design==0.6.1
 sphinx-jupyterbook-latex==1.0.0
 myst-parser==3.0.1
 linkify-it-py==2.0.3
@@ -23,7 +23,7 @@ sphinx-rtd-dark-mode==1.3.0
 # sphinx-copybutton
 # sphinx-rtd-theme
 # numpydoc
-# sphinx-tabs
+# sphinx-design
 # sphinx-jupyterbook-latex
 # myst-parser
 # linkify-it-py

[…]
