### 2026-03-02 313c213
Fix link formatting in README.md
