### 2026-03-03 5cc2f9a
njiited epl shear solver
