+++ paper/paper.md
@@ -126,7 +126,7 @@ Modular pipeline specialized for population statistics in the GW context, allowi
 Modularity is a key design principle, where each component is developed as an independent module, allowing for easy maintenance, testing, and future enhancements. This includes modules that deals with GW source properties, lens galaxy attributes, image properties, and rate calculations. SNR computation and Pdet calculation is handle seperately by `gwsnr` backend. 
 
 <!-- efficiency -->
-Most functions in $ler$ is a class object, which encompases cosmology related (e.g., luminosity distance, comoving volume [@astropy]), sampling related (e.g., sampling of source and lens properties), selection effect related (e.g., cross-section calculation). This class object leverages njitted interpolation of any smooth fuctions (single or multi-dimensional) or PDFs or CDFs (for inverse transform sampling) to optimize the workflow. Furthermore, for sampling strongly lensed events (with cross-section based strong lensing selection with importance sampling), parallelization is implemented through *prange* to further speed up the process. The major bottleneck of the entire workflow is lens equation solving, which is handled by *lenstronomy* [@Birrer2021], but $ler$ tackles this through parallelization with multiprocessing. SNR calculations are optimized using [*gwsnr*](https://gwsnr.readthedocs.io/en/latest/) python package backend, which again uses njitted interpolation and multiprocessing. This combination of techniques allows $ler$ to achieve significant speed-ups (1000x) compared to traditional methods, enabling the generation of large simulated populations (e.g., 10 million events) and convergence of rate integrals in a reasonable time frame (5 min for unlensed and 1 hour for lensed events on 6 CPU cores), which is crucial for robust forecasting studies and building statistics for validation of lensing events.
+Most functions in $ler$ is a class object, which encompases cosmology related (e.g., luminosity distance, comoving volume [@astropy]), sampling related (e.g., sampling of source and lens properties), selection effect related (e.g., cross-section calculation). This class object leverages njitted interpolation of any smooth fuctions (single or multi-dimensional) or PDFs or CDFs (for inverse transform sampling) to optimize the workflow. Furthermore, for sampling strongly lensed events (with cross-section based strong lensing selection with importance sampling), parallelization is implemented through *prange* to further speed up the process. The major bottleneck of the entire workflow is lens equation solving, which is handled by *lenstronomy* [@Birrer2021], but $ler$ tackles this through parallelization with multiprocessing. SNR calculations are optimized using [*gwsnr*](https://gwsnr.hemantaph.com) python package backend, which again uses njitted interpolation and multiprocessing. This combination of techniques allows $ler$ to achieve significant speed-ups (1000x) compared to traditional methods, enabling the generation of large simulated populations (e.g., 10 million events) and convergence of rate integrals in a reasonable time frame (5 min for unlensed and 1 hour for lensed events on 6 CPU cores), which is crucial for robust forecasting studies and building statistics for validation of lensing events.
 <!-- adaptability -->
 
 The modular design also facilitates the integration of additional statistics and models as research evolves, ensuring that $ler$ remains adaptable to new scientific insights and requirements. $ler$ inherently supports customization, allowing users to input their own distributions for source and lens properties, as well as custom detection criteria, making it a versatile tool for a wide range of astrophysical studies.
