### 2026-03-11 e22eabf
cross-section calculation
