+++ ler/image_properties/sample_caustic_points_njit copy.py
@@ -1,276 +1,0 @@
-# sample_caustic_points_njit.py
-from .cross_section_njit import caustic_double_points_epl_shear, polygon_area
-import numpy as np
-from numba import njit, prange
-
-TWO_PI = 2.0 * np.pi
-
-
-# ---------------------------------------------------------------------------
-# Geometry helpers
-# ---------------------------------------------------------------------------
-
-
-@njit(cache=True)
-def _poly_bbox(xv, yv):
-    """Compute polygon bounding box (manual reduction; fast in njit)."""
-    n = xv.size
-    xmin = xv[0]
-    xmax = xv[0]
-    ymin = yv[0]
-    ymax = yv[0]
-    for i in range(1, n):
-        x = xv[i]
-        y = yv[i]
-        if x < xmin:
-            xmin = x
-        elif x > xmax:
-            xmax = x
-        if y < ymin:
-            ymin = y
-        elif y > ymax:
-            ymax = y
-    return xmin, xmax, ymin, ymax
-
-
-@njit(cache=True)
-def _build_edge_coeffs(xv, yv):
-    """
-    Precompute edge coefficients for the even–odd (ray crossing) point-in-polygon test.
-
-    For each edge (i -> j), store:
-      xi, yi, yj, slope = (xj-xi)/(yj-yi)
-
-    Horizontal edges (yj == yi) are marked with slope=0 and will be skipped by the test.
-    """
-    n = xv.size
-    xi = np.empty(n, dtype=np.float64)
-    yi = np.empty(n, dtype=np.float64)
-    yj = np.empty(n, dtype=np.float64)
-    slope = np.empty(n, dtype=np.float64)
-
-    # edge from i to j=i+1 (last to 0)
-    for i in range(n - 1):
-        x0 = xv[i]
-        y0 = yv[i]
-        x1 = xv[i + 1]
-        y1 = yv[i + 1]
-        xi[i] = x0
-        yi[i] = y0
-        yj[i] = y1
-        dy = y1 - y0
-        if abs(dy) < 1e-300:
-            slope[i] = 0.0
-        else:
-            slope[i] = (x1 - x0) / dy
-
-    # last edge
-    x0 = xv[n - 1]
-    y0 = yv[n - 1]
-    x1 = xv[0]
-    y1 = yv[0]
-    xi[n - 1] = x0
-    yi[n - 1] = y0
-    yj[n - 1] = y1
-    dy = y1 - y0
-    if abs(dy) < 1e-300:
-        slope[n - 1] = 0.0
-    else:
-        slope[n - 1] = (x1 - x0) / dy
-
-    return xi, yi, yj, slope
-
-
-@njit(cache=True, inline="always")
-def _point_in_poly_precomp(x, y, xi, yi, yj, slope):
-    """
-    Even–odd rule with precomputed edge coefficients.
-
-    Skips horizontal edges via slope==0 sentinel.
-    """
-    inside = False
-    n = xi.size
-    for k in range(n):
-        s = slope[k]
-        if s == 0.0:
-            continue
-        y0 = yi[k]
-        y1 = yj[k]
-        # (y0 > y) != (y1 > y) is the standard crossing condition
-        c0 = y0 > y
-        c1 = y1 > y
-        if c0 != c1:
-            xint = xi[k] + (y - y0) * s
-            if x < xint:
-                inside = not inside
-    return 1 if inside else 0
-
-
-@njit(parallel=True, cache=True)
-def _points_in_poly_precomp(xs, ys, xi, yi, yj, slope):
-    """Parallel PIP for many points, using precomputed edge coefficients."""
-    n_points = xs.size
-    out = np.empty(n_points, dtype=np.uint8)
[…]
