+++ ler/utils/cosmological_coversions.py
@@ -4,20 +4,138 @@ from .function_interpolation import FunctionConditioning
 # for redshift to luminosity distance conversion
 from astropy.cosmology import LambdaCDM
 
-def redshift_optimal_spacing(z_min, z_max, resolution):
-    n_geom = int(resolution * 0.6)  # 60% for geometric spacing (low z)
-    n_lin = resolution - n_geom  # 40% for linear spacing (high z)
-    # Transition point between geometric and linear
-    z_transition = z_min + (z_max - z_min) * 0.3
-    # Geometric spacing for low z
-    z_low = np.geomspace(z_min, z_transition, n_geom)
-    # Linear spacing for high z
-    z_high = np.linspace(z_transition, z_max, n_lin + 1)[
-        1:
-    ]  # Skip first point to avoid duplicate
-    # Combine arrays
-    zs = np.concatenate([z_low, z_high])
-    return zs
+
+def generate_mixed_grid(
+    x_min,
+    x_max,
+    resolution,
+    power_law_part='lower',
+    spacing_trend='increasing',
+    power=2.3,
+    value_transition_fraction=0.6,
+    num_transition_fraction=0.8,
+    auto_match_slope=True,
+):
+    """
+    Generalized mixed spacing grid generator. Safely handles negative ranges.
+
+    Parameters
+    ----------
+    x_min, x_max : float
+        The physical domain boundaries (can be negative, crossing zero, etc.).
+    power_law_part : str
+        'lower' to apply power-law to the first segment.
+        'upper' to apply power-law to the second segment.
+    spacing_trend : str
+        'increasing': Step sizes grow as x increases.
+        'decreasing': Step sizes shrink as x increases.
+    power : float
+        Exponent (>= 1.0) to control curvature. Overridden if auto_match_slope=True.
+    auto_match_slope : bool
+        If True, analytically calculates the required power AND trend to guarantee 
+        a smooth transition without a kink.
+
+    Returns
+    -------
+    numpy.ndarray
+        The generated grid points.
+
+    Examples
+    --------
+    from ler.utils.cosmological_conversions import generate_mixed_grid
+
+    resolution=20
+    # linear+power-law with power-law in the upper segment and decreasing step sizes
+    x = generate_mixed_grid(
+        x_min=0.0, x_max=10.0, resolution=resolution,
+        power_law_part='upper',
+        spacing_trend='decreasing',  # Forces largest steps near z_trans
+        power=2.5,
+        value_transition_fraction=0.6,
+        num_transition_fraction=0.3,
+        auto_match_slope=True       # We accept the kink to control the exact power
+    )
+    # powerlaw+linear with power-law in the lower segment and increasing step sizes
+    x = generate_mixed_grid(
+        x_min=0.0, x_max=10.0, resolution=resolution,
+        power_law_part='lower',
+        spacing_trend='increasing',  # Forces largest steps near z_trans
+        power=2.5,
+        value_transition_fraction=0.3,
+        num_transition_fraction=0.6,
+        auto_match_slope=True       # We accept the kink to control the exact power
+    )
+    """
+    if x_max <= x_min:
+        return np.linspace(x_min, x_max, resolution)
+
+    # 1. Normalized transition parameter
[…]
