### 2026-03-22 #7420 @warrickball
Hi @hemantaph, how are things going? I can see there has been work in the repo though I'll flag that it's become quite messy. [This commit](https://github.com/hemantaph/ler/commit/e22eabf84efc74053b751269fbd2ad1cefb2212d) seems to have included a huge amount of churn in the repo, including a bunch of scripts (`fix_inline{,2,3}.py` and `THREAD_MANAGEMENT_*.md`), that looks like it needs tidying up.

It also looks like you're including the built docs in the repo, which shouldn't be necessary. They should be built as part of the documentation pipeline. But this may have already been the case, since I haven't explored the repo history deeply.

I'm going to be on vacation for essentially the next four weeks, and I'd like to be able to return to this and start progressing with the review again. I appreciate that the reviewers @kazewong and @michellegurevich haven't commented in many months, I expect because this has become so drawn out that it isn't clear what's required and they're potentially no longer available. I'm happy to try to reboot this review, including finding new reviewers if necessary, on my return on April 20 but I expect that the submission will be ready for it. That means addressing https://github.com/hemantaph/ler/issues/12, https://github.com/hemantaph/ler/issues/13 and https://github.com/hemantaph/ler/issues/14 (or showing that they've been addressed), even if we change reviewers, and a statement from you here that the review is ready to continue.
