### 2026-03-23 #7420 @hemantaph
@warrickball I will clean up the main repo. Also, I am rewriting the paper draft with the new JOSS requirements. I will make a bigger update in the coming days, with all the issues suggested.
