### 2026-03-26 67eb772
updates on image property calculation
