+++ ler/image_properties/cross_section_njit test1.py
@@ -1,16 +1,16 @@
 """
-Module for analytical caustic computation and cross-section evaluation.
+Module for analytical caustic computation and cross-section evaluation
+of EPL (Elliptical Power-Law) + external shear lens models.
 
 Provides numba-accelerated routines for computing critical curves, caustic
-boundaries, polygon areas, and lensing cross sections for EPL (Elliptical
-Power-Law) + external shear lens models. All core functions are decorated
-with ``@njit`` for high performance.
+boundaries, polygon areas, and lensing cross sections. All core functions
+are decorated with ``@njit`` for high performance.
 
 Usage:
-    Basic workflow example:
+    Compute the double-caustic boundary for a single lens:
 
-    >>> from ler.image_properties.cross_section_njit import caustic_points_epl_shear
-    >>> pts = caustic_points_epl_shear(theta_E=1.0, q=0.8, phi=0.0, gamma=2.0, gamma1=0.03, gamma2=-0.01)
+    >>> from ler.image_properties.cross_section_njit import caustics_epl_shear
+    >>> pts = caustics_epl_shear(q=0.8, phi=0.0, gamma=2.0, gamma1=0.03, gamma2=-0.01)
 
 Copyright (C) 2026 Phurailatpam Hemantakumar. Distributed under MIT License.
 """
@@ -21,7 +21,6 @@ from numba import njit, prange
 C_LIGHT = 299792458.0  # m/s
 PI = np.pi
 TWO_PI = 2.0 * PI
-EPS = 1e-16
 
 @njit(cache=True, fastmath=True)
 def phi_q2_ellipticity(phi, q):
@@ -31,16 +30,16 @@ def phi_q2_ellipticity(phi, q):
     Parameters
     ----------
     phi : ``float``
-        Position angle of the lens major axis in radians. \n
+        Position angle of the lens major axis in radians.
     q : ``float``
-        Axis ratio (minor/major), where ``0 < q <= 1``. \n
+        Axis ratio (minor/major), where ``0 < q <= 1``.
 
     Returns
     -------
     e1 : ``float``
-        First ellipticity component. \n
+        First ellipticity component.
     e2 : ``float``
-        Second ellipticity component. \n
+        Second ellipticity component.
 
     Examples
     --------
@@ -57,16 +56,16 @@ def _shear_cartesian2polar(gamma1, gamma2):
     Parameters
     ----------
     gamma1 : ``float``
-        First Cartesian shear component. \n
+        First Cartesian shear component.
     gamma2 : ``float``
-        Second Cartesian shear component. \n
+        Second Cartesian shear component.
 
     Returns
     -------
     phi : ``float``
-        Shear position angle in radians. \n
+        Shear position angle in radians.
     gamma : ``float``
-        Shear magnitude. \n
+        Shear magnitude.
     """
     phi = np.arctan2(gamma2, gamma1) / 2
     gamma = np.sqrt(gamma1**2 + gamma2**2)
@@ -77,24 +76,6 @@ def _shear_cartesian2polar(gamma1, gamma2):
 def ellipticity2phi_q(e1, e2):
     """
     Convert complex ellipticity moduli to orientation angle and axis ratio.
-
-    Parameters
-    ----------
-    e1 : ``float``
-        First ellipticity component. \n
-    e2 : ``float``
-        Second ellipticity component. \n
-
-    Returns
-    -------
-    phi : ``float``
-        Orientation angle in radians. \n
-    q : ``float``
[…]
