### 2026-05-04 e79e33f
major update; v0.6.0
