+++ ler/image_properties/epl_shear_njit copy 2.py
@@ -1,1497 +1,0 @@
-"""
-Module for semi-analytical image finding in EPL + external shear lens models.
-
-Provides a numba-accelerated solver that locates multiple lensed images,
-computes their magnifications, arrival times, and image types for an
-elliptical power-law (EPL) mass profile with external shear.
-
-Usage:
-    Solve for image positions of a single source:
-
-    >>> from ler.image_properties.epl_shear_njit import image_position_analytical_njit
-    >>> x_img, y_img, tau, mu, itype, n = image_position_analytical_njit(
-    ...     x=0.1, y=0.05, q=0.8, phi=0.3, gamma=2.0,
-    ...     gamma1=0.03, gamma2=-0.01
-    ... )
-
-Copyright (C) 2026 Phurailatpam Hemantakumar. Distributed under MIT License.
-"""
-
-import numpy as np
-from numba import njit, prange
-from .sample_caustic_points_njit import sample_source_from_double_caustic
-from .cross_section_njit import cdot
-
-
-EPS = 1e-16
-MAX_ROOTS = 16
-MAX_IMGS = 16
-C_LIGHT = 299792458.0  # m/s
-
-# @njit(cache=True, fastmath=True)
-def pol_to_ell(r, theta, q):
-    """
-    Convert polar coordinates to elliptical coordinates.
-    """
-
-    phi = np.arctan2(np.sin(theta), np.cos(theta) * q)
-    rell = r * np.sqrt(q**2 * np.cos(theta) ** 2 + np.sin(theta) ** 2)
-    return rell, phi
-
-# @njit(cache=True, fastmath=True)
-def omega_scalar(phi, t, q, niter_max=200, tol=1e-16):
-    """
-    Scalar version of ``omega`` that avoids temporary array allocation.
-    """
-    f = (1.0 - q) / (1.0 + q)
-
-    # Floating-point safeguard: if q is extremely close to 1, f approaches 0.
-    # This prevents np.log(f) from evaluating to -inf and crashing the loop counter.
-    if f <= EPS:
-        return np.exp(1j * phi)
-
-    omega_sum = 0.0 + 0.0j
-    niter = min(niter_max, int(np.log(tol) / np.log(f)) + 2)
-    Omega = np.exp(1j * phi)
-    fact = -f * np.exp(2j * phi)
-
-    for n in range(1, niter):
-        omega_sum += Omega
-        Omega *= (2.0 * n - (2.0 - t)) / (2.0 * n + (2.0 - t)) * fact
-    omega_sum += Omega
-    return omega_sum
-
-# @njit(cache=True, fastmath=True)
-def _alpha_epl_shear_scalar(x, y, b, q, t=1, gamma1=0, gamma2=0):
-    """
-    Scalar complex deflection for EPL + external shear. Same as _alpha_epl_shear but avoids temporary array allocation by calling the scalar omega function.
-
-    alpha_EPL = (2 * b) / (1 + q) * (b / R)^t * R / b * Omega
-    alpha_Shear = gamma1 * x + gamma2 * y + i * (gamma2 * x - gamma1 * y)
-    """
-    zz = x * q + 1j * y
-    R = np.abs(zz)
-    phi = np.angle(zz)
-    Omega = omega_scalar(phi, t, q)
-    # alph = (2 * b) / (1 + q) * np.nan_to_num((b / R) ** t * R / b) * Omega
-    alph = (2* b**t)/(1 + q) * R**(1-t) * Omega
-    
-    return (
-        alph
-        + (gamma1 * x + gamma2 * y)
-        + 1j * (gamma2 * x - gamma1 * y)
-    )
-
-# @njit(cache=True, fastmath=True)
-def _image_type_name(code):
-    """
[…]
