+++ ler/_version.py
@@ -1,2 +1,2 @@
 # Version information for ler package
-__version__ = "0.5.9"
+__version__ = "0.6.0"

[…]
