+++ tests/conftest.py
@@ -1,0 +1,150 @@
+"""
+Pytest configuration and shared fixtures for the ler test suite.
+
+This module sets up session-scoped fixtures that are shared across all tests.
+Most importantly, it ensures that the bundled interpolator data is extracted
+to a session cache directory once per test session, avoiding repeated costly
+interpolator construction. The cache directory is intentionally persistent so
+that `interpolator_json` is not deleted after the test run.
+
+Wall-clock / JIT comparison tests, heavy cross-section samplers, and the LeR
+speed benchmark subprocess checks are marked ``@pytest.mark.slow`` and are
+deselected from the default collection.
+"""
+
+import os
+import re
+from pathlib import Path
+import zipfile
+import pytest
+from importlib_resources import files as resources_files
+
+
+_TEST_CACHE_DIR = Path(__file__).resolve().parent / ".cache"
+_NUMBA_CACHE_DIR = _TEST_CACHE_DIR / "numba"
+_MPL_CACHE_DIR = _TEST_CACHE_DIR / "matplotlib"
+_NUMBA_CACHE_DIR.mkdir(parents=True, exist_ok=True)
+_MPL_CACHE_DIR.mkdir(parents=True, exist_ok=True)
+
+# Some optional backends imported during collection enable Numba caching at
+# import time. Provide cache locations before test modules import those backends.
+os.environ.setdefault("NUMBA_CACHE_DIR", str(_NUMBA_CACHE_DIR))
+os.environ.setdefault("MPLCONFIGDIR", str(_MPL_CACHE_DIR))
+
+
+def pytest_addoption(parser):
+    """Register ``--run-slow`` so default runs can omit ``slow`` tests."""
+    parser.addoption(
+        "--run-slow",
+        action="store_true",
+        default=False,
+        help=(
+            "Include @pytest.mark.slow tests in the default suite (wall-clock / heavy benchmarks). "
+            "Not needed when selecting them explicitly (e.g. ``pytest -m slow``)."
+        ),
+    )
+
+
+def _mark_expression_requests_slow(markexpr):
+    """
+    Return True if ``-m markexpr`` is intended to select ``slow`` tests.
+
+    Default (empty expression) returns False so we deselect ``slow`` tests.
+    Expressions containing ``not slow`` return False.
+    """
+    m = (markexpr or "").strip()
+    if not m:
+        return False
+    if re.search(r"\bnot\s+slow\b", m):
+        return False
+    return bool(re.search(r"\bslow\b", m))
+
+
+@pytest.fixture(scope="session")
+def interpolator_directory(tmp_path_factory):
+    """
+    Session-scoped fixture that extracts the bundled interpolator data to a
+    cache directory. The directory is reused for the entire test session and
+    kept on disk after the tests finish.
+
+    Returns
+    -------
+    interpolator_directory : str
+        Absolute path to the extracted interpolator_json directory.
+    """
+    # Use a persistent cache directory by default so interpolator_json is not
+    # deleted when the pytest temporary directory is cleaned up.
+    #
+    # Users can override the location if needed (e.g. on shared runners).
+    cache_root = os.environ.get("LER_TEST_INTERPOLATOR_CACHE_DIR", "").strip()
+    if cache_root:
+        cache_dir = Path(cache_root).expanduser().resolve()
+    else:
+        cache_dir = Path(__file__).resolve().parent / ".cache" / "interpolators"
+    cache_dir.mkdir(parents=True, exist_ok=True)
+
+    interpolator_json_dir = cache_dir / "interpolator_json"
+    if interpolator_json_dir.exists():
+        return str(interpolator_json_dir)
+
+    # NOTE: use the top-level package as the anchor to avoid importing
[…]
