+++ tests/unit/test_lens_galaxy_parameter_distribution.py
@@ -1,0 +1,355 @@
+"""
+Unit tests for LensGalaxyParameterDistribution.
+
+This module verifies lens-galaxy parameter sampling and lensed-source redshift
+sampling for ``ler.lens_galaxy_population.LensGalaxyParameterDistribution``.
+
+Test Coverage:
+--------------
+- Class initialization sanity checks (attributes, cached interpolators)
+- Strongly-lensed source redshift sampler:
+  - ``strongly_lensed_source_redshift`` returns finite samples within [z_min, z_max]
+  - ``get_attribute=True`` returns a ``FunctionConditioning`` object
+- Cross-section-based EPL+shear lens-parameter sampling (``epl_shear_sl_parameters_rvs``):
+  - ``test_cross_section_based_sampler_variants``: ``importance_sampler_partial``
+    on default runs (slow tests omitted); uses ``n_prop`` from
+    ``N_PROP_CROSS_SECTION`` for a fast CI-friendly check.
+  - ``test_cross_section_based_sampler_variants_heavy`` *(slow)*: the three
+    slower samplers — ``rejection_sampler_full``, ``importance_sampler_full``,
+    ``rejection_sampler_partial`` — same output assertions as the fast test.
+  - ``test_njit_speed`` *(slow)*: no-JIT subprocess vs in-process njit for
+    ``importance_sampler_partial`` with ``npool=1`` and ``npool=6``.
+- Intrinsic lens-parameter sampling (no strong-lensing weighting):
+  - ``sample_all_routine_epl_shear_intrinsic`` returns finite samples with expected keys.
+"""
+
+import numpy as np
+import pytest
+from tests_utils import CommonTestUtils, median_call_time
+from ler.lens_galaxy_population import LensGalaxyParameterDistribution
+from ler.utils import FunctionConditioning
+
+
+# ---------------------------------------------------------------------------
+# Test configuration
+# ---------------------------------------------------------------------------
+
+N_SAMPLES = 10
+
+DEFAULT_CONFIG = dict(
+    npool=6,
+    z_min=0.0,
+    z_max=10.0,
+    create_new_interpolator=False,
+    verbose=False,
+    buffer_size=200,
+)
+
+# Fast sampler: exercised on every default run (slow-marked tests deselected).
+CROSS_SECTION_SAMPLER_FAST = "importance_sampler_partial"
+
+# Heavier cross-section samplers: only under ``@pytest.mark.slow`` (slow rejection /
+# importance paths with larger proposal counts).
+CROSS_SECTION_SAMPLERS_SLOW_ONLY = [
+    "rejection_sampler_full",
+    "importance_sampler_full",
+    "rejection_sampler_partial",
+]
+
+# Proposal counts per sampler (interface/sanity coverage, not production efficiency).
+N_PROP_CROSS_SECTION = dict(
+    rejection_sampler_full=500,
+    importance_sampler_full=100,
+    rejection_sampler_partial=100,
+    importance_sampler_partial=50,
+)
+
+
+# ---------------------------------------------------------------------------
+# Small helpers to keep tests readable
+# ---------------------------------------------------------------------------
+
+def _make_config(interpolator_dir, **overrides):
+    # Start from defaults and allow per-test overrides.
+    config = DEFAULT_CONFIG.copy()
+    config["directory"] = interpolator_dir
+    config.update(overrides)
+    return config
+
+
+def _run_cross_section_sampler_variant(lens, sampler_type):
+    """
+    Reconfigure ``lens`` for the given cross-section sampler, sample EPL+shear
+    lens parameters, and return the parameter dict.
+
+    Mutates ``lens.lens_functions``, ``lens.lens_functions_params``, and
[…]

[…]
