### 2026-05-04 aa33b92
buf fixs tests; npool<=4
