+++ .github/workflows/documentation.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: '3.11'
+          python-version-file: ".python-version"
 
       - name: Install system dependencies (Pandoc for nbsphinx)
         run: |

[…]
