+++ tests/unit/test_gwrates.py
@@ -1,499 +1,0 @@
-"""
-Unit tests for the GWRATES class.
-
-Test Coverage:
---------------
-- ``test_init``: core attributes, default ``json_file_names``, ``ler_directory``
-  exists on disk, inheritance from ``CBCSourceParameterDistribution``,
-  custom ``pdet_finder`` stored, custom astropy cosmology stored.
-- ``test_gw_cbc_statistics``: output keys and finite values; ``batch_size`` and
-  ``resume`` (first run 20 samples, second run resumes from batch 3 → 40 total,
-  file content matches returned dict); ``output_jsonfile`` modes (``True``:
-  default name, content verified; ``False``: no file written, verified).
-- ``test_gw_rate``: returns ``(rate, gw_param_detectable)``; rate equals
-  ``normalization_pdf_z * detectable / total``; ``pdet_type='probability_distribution'``
-  sums pdet_net directly; ``pdet_threshold`` boundary cases: 0.0 (all detectable),
-  2.0 (none, rate=0); ``gw_param=None`` loads from the default JSON file on disk;
-  invalid ``pdet_type`` raises ``ValueError``.
-- ``test_selecting_n_detectable_events``: collects at least ``size`` detectable
-  events when ``stopping_criteria=None``; ``trim_to_size=True`` trims randomly to
-  exactly ``size`` and values are a subset of the original; ``resume=True`` from
-  an already-complete file skips resampling; ``stopping_criteria`` stops when the
-  rate converges and the collected size exceeds ``size``.
-- ``test_utilities``: ``rate_function`` math; ``_load_param`` returns a copy.
-
-"""
-
-import os
-import numpy as np
-import pytest
-from astropy.cosmology import LambdaCDM
-from tests_utils import (
-    CommonTestUtils,
-    EXPECTED_UNLENSED_GW_KEYS_ALIGNED_SPIN,
-    EXPECTED_UNLENSED_PARAM_KEYS,
-    clamp_npool_for_numba,
-)
-
-from ler.rates import GWRATES
-from ler.gw_source_population import CBCSourceParameterDistribution
-from ler.utils import get_param_from_json
-
-
-# ---------------------------------------------------------------------------
-# Test configuration
-# ---------------------------------------------------------------------------
-
-# Default keys created by GWRATES.__init__ for parameter storage.
-DEFAULT_JSON_KEYS = [
-    "gwrates_params",
-    "gw_param",
-    "gw_param_detectable",
-]
-
-# Minimum keys expected in gw_cbc_statistics output (see ``tests_utils``).
-EXPECTED_PARAM_KEYS = EXPECTED_UNLENSED_PARAM_KEYS
-
-N_SAMPLES = 20
-
-DEFAULT_CONFIG = dict(
-    # Preferred high npool on big machines (6); clamp so Numba succeeds on slim CI VMs.
-    npool=clamp_npool_for_numba(6),
-    z_min=0.0,
-    z_max=10.0,
-    event_type="BBH",
-    create_new_interpolator=False,
-    verbose=False,
-)
-
-def mock_pdet_finder(gw_param_dict):
-    """Lightweight mock: detection probability = 1 for every event."""
-    size = len(gw_param_dict["zs"])
-    return dict(pdet_net=np.ones(size))
-
-
-def _make_gwrates(interpolator_directory, ler_directory, **overrides):
-    # Always provide a mock pdet_finder so gwsnr is not initialized.
-    cfg = DEFAULT_CONFIG.copy()
-    cfg.update(overrides)
-    return GWRATES(
-        pdet_finder=mock_pdet_finder,
-        interpolator_directory=interpolator_directory,
-        ler_directory=ler_directory,
-        **cfg,
-    )
-
-
[…]
