+++ tests/integration/test_gwrates_populations.py
@@ -1,313 +1,0 @@
-"""
-Integration tests for ``GWRATES`` (population sampling and BBH hyperparameter stress).
-
-Test Coverage:
---------------
-- ``test_gwrates_population_rates_and_outputs``: BBH / BNS / NSBH;
-  ``gw_cbc_statistics`` at ``SIZE``; parameter dict keys and finite arrays;
-  ``gw_rate`` above a loose floor and **below** ``normalization_pdf_z``
-  (intrinsic merger-rate normalization from the CBC model; ``gw_rate`` is that
-  rate times the simulated detectable fraction).
-- ``test_gwrates_hyper_uncertainty_rates``: BBH only; vary one of ``H0``,
-  merger ``R0``, or ``mass_1_source`` ``alpha_1`` between two values; compares
-  rates so the endpoint with larger astrophysical rate gives a larger number
-  (``normalization_pdf_z`` for the ``H0`` case, ``gw_rate`` for ``R0`` /
-  ``alpha_1``).
-
-They use large Monte Carlo samples (``SIZE``, ``BATCH_SIZE`` below) and real
-``gwsnr`` detection, and run in default CI (``.github/workflows/tests.yml``).
-
-"""
-
-import os
-
-import numpy as np
-import pytest
-from astropy.cosmology import LambdaCDM
-
-from ler.rates import GWRATES
-from tests_utils import (
-    CommonTestUtils,
-    EXPECTED_UNLENSED_GW_KEYS_NO_SPIN,
-    EXPECTED_UNLENSED_GW_KEYS_PRECESSING_SPIN,
-    clamp_npool_for_numba,
-)
-
-# Prefer six workers where Numba permits; clamps on four-vCPU GitHub-hosted runners.
-NPOOL = clamp_npool_for_numba(6)
-
-SIZE = 20000
-BATCH_SIZE = 10000
-
-Z_MIN = 0.0
-Z_MAX = 10.0
-
-COSMO_OM0 = 0.315
-COSMO_ODE0 = 0.685
-COSMO_TCMB = 2.725
-
-# Planck 2018-like mean with a ±0.42 km/s/Mpc bracket for the H0-only check.
-H0_LOW = 67.66 - 0.42
-H0_HIGH = 67.66 + 0.42
-
-# Merger-rate amplitude R0 (Gpc^-3 yr^-1): low vs high vs BBH default ~19e-9.
-R0_LOW = 14.0 * 1e-9
-R0_HIGH = 26.0 * 1e-9
-
-# Broken-power-law slope below break for ``mass_1_source`` (BBH only in these tests).
-ALPHA_1_LOW = -0.05
-ALPHA_1_HIGH = 2.86
-
-
-def _make_cosmo(h0):
-    """Flat ``LambdaCDM`` with fixed densities; only ``H0`` varies."""
-    return LambdaCDM(
-        H0=h0,
-        Om0=COSMO_OM0,
-        Ode0=COSMO_ODE0,
-        Tcmb0=COSMO_TCMB,
-    )
-
-
-class TestGWRATESPopulationsUncertainty(CommonTestUtils):
-    """Integration tests for ``GWRATES`` with real ``gwsnr`` detection."""
-
-    @pytest.mark.parametrize(
-        "event_type, rate_low, expected_gw_keys",
-        [
-            ("BBH", 1.0e2, EXPECTED_UNLENSED_GW_KEYS_PRECESSING_SPIN),
-            ("BNS", 1e5, EXPECTED_UNLENSED_GW_KEYS_NO_SPIN),
-            ("NSBH", 1e4, EXPECTED_UNLENSED_GW_KEYS_NO_SPIN),
-        ],
-    )
-    def test_gwrates_population_rates_and_outputs(
[…]
