+++ tests/unit/test_ler.py
@@ -1,0 +1,894 @@
+"""
+Unit tests for the LeR class.
+
+Test Coverage:
+--------------
+- ``test_init``: core attributes, default ``json_file_names``, ``ler_directory``
+  exists on disk, inheritance from ``LensGalaxyParameterDistribution``,
+  custom ``pdet_finder`` stored, custom astropy cosmology stored.
+- ``test_unlensed_cbc_statistics``: output keys and finite values; ``batch_size``
+  and ``resume`` (first run 20 samples, second run resumes from batch 3 → 40 total,
+  file content matches returned dict); ``output_jsonfile`` modes (``True``:
+  default name, content verified; ``False``: no file written, verified).
+- ``test_lensed_cbc_statistics``: output keys and finite values; ``batch_size``
+  and ``resume`` (first run 10 samples, second run resumes → 20 total, file
+  content matches returned dict); ``output_jsonfile`` modes; image columns and
+  ``pdet_net`` shapes; optional ``include_redundant_parameters`` /
+  ``include_effective_parameters`` key checks.
+- ``test_lensed_cbc_statistics_redundant_and_effective``: small run with both
+  flags true; asserts :data:`EXPECTED_REDUNDANT_KEYS` and
+  :data:`EXPECTED_EFFECTIVE_KEYS`.
+- ``test_unlensed_rate``: returns ``(rate, unlensed_param_detectable)``; rate
+  equals ``normalization_pdf_z * detectable / total``; ``pdet_type=
+  'probability_distribution'`` sums pdet_net directly; ``pdet_threshold``
+  boundary cases: 0.0 (all detectable), 2.0 (none, rate=0); ``unlensed_param=None``
+  loads from the default JSON file on disk; invalid ``pdet_type`` raises
+  ``ValueError``.
+- ``test_lensed_rate``: returns ``(rate, lensed_param_detectable)``; rate equals
+  ``normalization_pdf_z_lensed * detectable / total``; ``pdet_threshold=2.0``
+  (none detectable, rate=0); ``lensed_param=None`` loads from the default JSON
+  file on disk.
+- ``test_utilities``: ``rate_function`` math for both ``param_type='unlensed'``
+  and ``param_type='lensed'``; ``_load_param`` returns a copy.
+- ``test_selecting_n_unlensed_detectable_events``: ``stopping_criteria=None``
+  accumulates batches until detectable events exceed ``size``; meta-data file has
+  the expected keys and the final rate matches the return value; ``trim_to_size=False``
+  gives final_size >= size; ``resume=True`` + ``trim_to_size=True`` randomly
+  subsamples to ``new_size`` (every value present in the original collection);
+  ``stopping_criteria=dict(...)`` stops when cumulative rate has converged and
+  collected size exceeds ``size``.
+- ``test_selecting_n_lensed_detectable_events``: same three test cases as the
+  unlensed variant; uses smaller ``batch_size`` and ``size`` because lensed
+  sampling is slower; ``pdet_threshold=[0.5, 0.5]`` (default 2-image condition).
+
+"""
+
+import os
+import numpy as np
+import pytest
+from astropy.cosmology import LambdaCDM
+from ler.rates import LeR
+from ler.lens_galaxy_population import LensGalaxyParameterDistribution
+from ler.utils import get_param_from_json
+from tests_utils import (
+    CommonTestUtils,
+    EXPECTED_EFFECTIVE_KEYS,
+    EXPECTED_IMAGE_KEYS,
+    EXPECTED_SOURCE_POS_KEYS,
+    EXPECTED_LENSED_PARAM_KEYS,
+    EXPECTED_REDUNDANT_KEYS,
+    EXPECTED_UNLENSED_GW_KEYS_ALIGNED_SPIN,
[…]
