+++ paper/paper.md
@@ -66,135 +66,58 @@ bibliography: paper.bib
 
 # Summary
 
-<!-- Summary should include: A description of the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
+Gravitational waves (GWs) are ripples in spacetime caused by accelerating massive objects, primarily produced by violent cosmic mergers of compact binaries such as binary black holes (BBHs), binary neutron stars (BNSs), and neutron star-black hole pairs (NSBHs). Since 2015, the LIGO-Virgo-KAGRA (LVK) collaboration has routinely detected such GW transients during observing runs [@LIGO2015] [@VIRGO2015] [@KAGRA2021]. Most of these signals reach detectors unobstructed, forming a baseline unlensed population, but a small fraction encounter massive intervening objects like galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott2021] [@Wierda2021].
 
-<!-- domain knowledge 1 -->
-<!-- Gravitational waves (GWs) are ripples in the fabric of space and time caused by the acceleration of unevenly distributed mass or masses. Observable GWs are created especially during the violent cosmic events of merging compact binaries, such as 'binary black holes' (BBHs), 'binary neutron stars' (BNSs), and 'neutron star and black hole pair' (NSBHs). -->
-<!-- domain knowledge 2 -->
-<!-- As these signals travel across the universe, the vast majority reach our detectors unobstructed, forming the baseline (unlensed) GW population.
-However, a fraction (<0.1%) of these signals encounter massive objects (lenses) like galaxies or galaxy clusters along their path, which can split them into multiple images, that are magnified (de-magnified), time and phase shifted, through strong gravitational lensing [Wierda2021] [@Abbott2021]. -->
-<!-- narrower problem. What's the current state. Purpose of the work -->
-<!-- A profound comprehension and accurate modelling of both the unlensed population and lens population and the lensing effects is imperative for accurate astrophysical interpretation, inferring merger rate histories, and validating forthcoming lensing events [@Janquart2023] [@ligolensing2023] [@Abbott2021] [@Wierda2021]. -->
-
-Gravitational waves are spacetime ripples caused by accelerating massive objects, primarily observable from the violent cosmic mergers of compact binaries like 'binary black holes' (BBHs), 'binary neutron stars' (BNSs), and 'neutron star and black hole pairs' (NSBHs). While most of these signals reach our detectors unobstructed to form the baseline unlensed population, a small fraction encounter massive intervening objects such as galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott2021] [@Wierda2021].
-
-<!-- what does software do? high level functionality. -->
-<!-- $ler$ is a Python framework designed for the comprehensive statistical modeling of GW populations, providing a unified pipeline to calculate detectable rates for both unlensed and lensed events. Obtaining precise outcomes in these statistical analyses necessitates large-scale sampling of source properties, lens galaxy attributes, and lensing-induced image properties calculation (lens equation solves), along with the computation of detectable rates based on simulated detector responses, often numbering in the millions. This process is traditionally highly demanding. The $ler$ framework addresses this by employing ... techniques to optimize the workflow and computation efficiency required for handling large-scale statistical analyses, which leads to significant reductions in computational time (1000x speed up than traditional methods). Detailed description, source code, and examples are available in $ler$ [documentation](https://ler.hemantaph.com). -->
-
[…]
