+++ paper/paper.bib
@@ -1,3 +1,37 @@
+@article{Thrane2019,
+   title={An introduction to Bayesian inference in gravitational-wave astronomy: Parameter estimation, model selection, and hierarchical models},
+   volume={36},
+   ISSN={1448-6083},
+   url={http://dx.doi.org/10.1017/pasa.2019.2},
+   DOI={10.1017/pasa.2019.2},
+   journal={Publications of the Astronomical Society of Australia},
+   publisher={Cambridge University Press (CUP)},
+   author={Thrane, Eric and Talbot, Colm},
+   year={2019} 
+}
+
+@misc{lerdoc,
+  author = {Phurailatpam, Hemantakumar and More, Anupreeta and Narola, Harsh and Ng, Chung Yin and Janquart, Justin and Van Den Broeck, Chris and Hannuksela, Otto Akseli and Singh, Neha and Keitel, David},
+  title = {ler: A Python package for gravitational-wave lensed and unlensed event rate calculations},
+  year = {2026},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://ler.hemantaph.com}},
+}
+
+@article{Tessore2015,
+   title={The elliptical power law profile lens},
+   volume={580},
+   ISSN={1432-0746},
+   url={http://dx.doi.org/10.1051/0004-6361/201526773},
+   DOI={10.1051/0004-6361/201526773},
+   journal={Astronomy &amp; Astrophysics},
+   publisher={EDP Sciences},
+   author={Tessore, Nicolas and Benton Metcalf, R.},
+   year={2015},
+   month=Aug, pages={A79} 
+}
+
 @misc{Haris2018,
       title={Identifying strongly lensed gravitational wave signals from binary black hole mergers}, 
       author={K. Haris and Ajit Kumar Mehta and Sumit Kumar and Tejaswi Venumadhav and Parameswaran Ajith},
