+++ tests/unit/test_ler.py
@@ -699,14 +699,19 @@ class TestLeR(CommonTestUtils):
         meta_data = get_param_from_json(
             os.path.join(ler_instance.ler_directory, meta_data_file)
         )
-        assert isinstance(meta_data, dict) and len(meta_data) == 3, \
-            f"meta_data: expected dict with 3 keys, got {len(meta_data)}"
-        for key in ("events_total", "detectable_events", "total_rate"):
+        expected_meta_keys = ("events_total", "detectable_events", "total_rate", "batch_rate")
+        assert isinstance(meta_data, dict) and len(meta_data) == len(expected_meta_keys), \
+            f"meta_data: expected dict with {len(expected_meta_keys)} keys, got {len(meta_data)}"
+        for key in expected_meta_keys:
             assert key in meta_data, f"meta_data: missing key '{key}'"
         assert isinstance(meta_data["total_rate"][-1], float) \
             and np.isfinite(meta_data["total_rate"][-1]) \
             and meta_data["total_rate"][-1] > 0, \
             f"meta_data: expected finite positive float total_rate, got {meta_data['total_rate'][-1]}"
+        assert isinstance(meta_data["batch_rate"][-1], float) \
+            and np.isfinite(meta_data["batch_rate"][-1]) \
+            and meta_data["batch_rate"][-1] > 0, \
+            f"meta_data: expected finite positive float batch_rate, got {meta_data['batch_rate'][-1]}"
         assert meta_data["total_rate"][-1] == rate, \
             f"meta_data: final total_rate {meta_data['total_rate'][-1]} != returned rate {rate}"
 
@@ -784,7 +789,7 @@ class TestLeR(CommonTestUtils):
         -----
         - ``stopping_criteria=None``: accumulates batches until detectable events
           exceed ``size``; meta-data file has keys ``events_total``,
-          ``detectable_events``, ``total_rate``; final rate matches the return value.
+          ``detectable_events``, ``total_rate``, ``batch_rate``; final rate matches the return value.
         - ``trim_to_size=False``: returned dict has final_size >= size; stored JSON
           file matches the returned dict.
         - ``resume=True`` + ``trim_to_size=True``: no new sampling (existing file
@@ -818,14 +823,19 @@ class TestLeR(CommonTestUtils):
         meta_data = get_param_from_json(
             os.path.join(ler_instance.ler_directory, meta_data_file)
         )
-        assert isinstance(meta_data, dict) and len(meta_data) == 3, \
-            f"meta_data: expected dict with 3 keys, got {len(meta_data)}"
-        for key in ("events_total", "detectable_events", "total_rate"):
+        expected_meta_keys = ("events_total", "detectable_events", "total_rate", "batch_rate")
+        assert isinstance(meta_data, dict) and len(meta_data) == len(expected_meta_keys), \
+            f"meta_data: expected dict with {len(expected_meta_keys)} keys, got {len(meta_data)}"
+        for key in expected_meta_keys:
             assert key in meta_data, f"meta_data: missing key '{key}'"
         assert isinstance(meta_data["total_rate"][-1], float) \
             and np.isfinite(meta_data["total_rate"][-1]) \
             and meta_data["total_rate"][-1] > 0, \
             f"meta_data: expected finite positive float total_rate, got {meta_data['total_rate'][-1]}"
+        assert isinstance(meta_data["batch_rate"][-1], float) \
+            and np.isfinite(meta_data["batch_rate"][-1]) \
+            and meta_data["batch_rate"][-1] > 0, \
+            f"meta_data: expected finite positive float batch_rate, got {meta_data['batch_rate'][-1]}"
         assert meta_data["total_rate"][-1] == rate, \
             f"meta_data: final total_rate {meta_data['total_rate'][-1]} != returned rate {rate}"
 
