### 2026-05-06 39e5431
paper draft debugging
