+++ paper/paper.md
@@ -21,22 +21,22 @@ authors:
     # equal-contrib: False
     affiliation: "4,5"
   - name: Ng Chung Yin (Leo)
-    # equal-contrib: False 
+    # equal-contrib: False
     affiliation: "1"
   - name: Justin Janquart
     orcid: 0000-0003-2888-7152
-    # equal-contrib: False 
+    # equal-contrib: False
     affiliation: "4,5,6,7"
   - name: Chris Van Den Broeck
     orcid: 0000-0001-6800-4006
-    # equal-contrib: False 
+    # equal-contrib: False
     affiliation: "4,5"
   - name: Otto Akseli Hannuksela
     orcid: 0000-0002-3887-7137
-    # equal-contrib: False 
+    # equal-contrib: False
     affiliation: "1"
   - name: Neha Singh
-    orcid: 0000-0002-1135-3456 
+    orcid: 0000-0002-1135-3456
     # equal-contrib: False
     affiliation: "8"
   - name: David Keitel
@@ -46,7 +46,7 @@ authors:
 affiliations:
   - name: Department of Physics, The Chinese University of Hong Kong, Shatin, New Territories, Hong Kong
     index: 1
-  - name:  The Inter-University Centre for Astronomy and Astrophysics (IUCAA), Post Bag 4, Ganeshkhind, Pune 411007, India
+  - name: The Inter-University Centre for Astronomy and Astrophysics (IUCAA), Post Bag 4, Ganeshkhind, Pune 411007, India
     index: 2
   - name: Kavli Institute for the Physics and Mathematics of the Universe (IPMU), 5-1-5 Kashiwanoha, Kashiwa-shi, Chiba 277-8583, Japan
     index: 3
@@ -60,35 +60,35 @@ affiliations:
     index: 7
   - name: Departament de Física, Universitat de les Illes Balears, IAC3-IEEC, Crta. Valldemossa km 7.5, E-07122 Palma, Spain
     index: 8
-date: 19 June 2024
+date: 6 May 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-Gravitational waves (GWs) are ripples in spacetime caused by accelerating massive objects, primarily produced by violent cosmic mergers of compact binaries such as binary black holes (BBHs), binary neutron stars (BNSs), and neutron star-black hole pairs (NSBHs). Since 2015, the LIGO-Virgo-KAGRA (LVK) collaboration has routinely detected such GW transients during observing runs [@LIGO2015] [@VIRGO2015] [@KAGRA2021]. Most of these signals reach detectors unobstructed, forming a baseline unlensed population, but a small fraction encounter massive intervening objects like galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott2021] [@Wierda2021].
+Gravitational waves (GWs) are ripples in spacetime caused by accelerating massive objects, primarily produced by violent cosmic mergers of compact binaries such as binary black holes (BBHs), binary neutron stars (BNSs), and neutron star-black hole pairs (NSBHs). Since 2015, the LIGO-Virgo-KAGRA (LVK) collaboration [@LIGO:2015] [@VIRGO:2015] [@KAGRA:2021] has routinely detected such GW transients during observing runs [@GWTC4:2026] [@GWTC3:2023] [@GWTC2:2021] [@GWTC1:2019]. Most of these signals reach detectors unobstructed, forming a baseline unlensed population, but a small fraction can encounter massive intervening objects like galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott:2021] [@Wierda:2021].
 
[…]
