### 2026-05-06 8101b6b
paper draft debugging
