+++ paper/paper.bib
@@ -52,7 +52,7 @@
 
 @misc{GWTC4:2026,
       title={GWTC-4.0: Updating the Gravitational-Wave Transient Catalog with Observations from the First Part of the Fourth LIGO-Virgo-KAGRA Observing Run},
-      author={The LIGO Scientific Collaboration and the Virgo Collaboration and the KAGRA Collaboration and A. G. Abac and I. Abouelfettouh and F. Acernese and K. Ackley and C. Adamcewicz and S. Adhicary and D. Adhikari and N. Adhikari and R. X. Adhikari and V. K. Adkins and S. Afroz and A. Agapito and D. Agarwal and M. Agathos and N. Aggarwal and S. Aggarwal and O. D. Aguiar and I. -L. Ahrend and L. Aiello and A. Ain and P. Ajith and T. Akutsu and S. Albanesi and W. Ali and S. Al-Kershi and C. All{\'e}n{\'e} and A. Allocca and S. Al-Shammari and P. A. Altin and S. Alvarez-Lopez and W. Amar and O. Amarasinghe and A. Amato and F. Amicucci and C. Amra and A. Ananyeva and S. B. Anderson and W. G. Anderson and M. Andia and M. Ando and M. Andr{\'e}s-Carcasona and T. Andri{\'c} and J. Anglin and S. Ansoldi and J. M. Antelis and S. Antier and M. Aoumi and E. Z. Appavuravther and S. Appert and S. K. Apple and K. Arai and A. Araya and M. C. Araya and M. Arca Sedda and J. S. Areeda and N. Aritomi and F. Armato and S. Armstrong and N. Arnaud and M. Arogeti and S. M. Aronson and K. G. Arun and G. Ashton and Y. Aso and L. Asprea and M. Assiduo and S. Assis de Souza Melo and S. M. Aston and P. Astone and F. Attadio and F. Aubin and K. AultONeal and G. Avallone and E. A. Avila and S. Babak and C. Badger and S. Bae and S. Bagnasco and L. Baiotti and R. Bajpai and T. Baka and A. M. Baker and K. A. Baker and T. Baker and G. Baldi and N. Baldicchi and M. Ball and G. Ballardin and S. W. Ballmer and S. Banagiri and B. Banerjee and D. Bankar and T. M. Baptiste and P. Baral and M. Baratti and J. C. Barayoga and B. C. Barish and D. Barker and N. Barman and P. Barneo and F. Barone and B. Barr and L. Barsotti and M. Barsuglia and D. Barta and A. M. Bartoletti and M. A. Barton and I. Bartos and A. Basalaev and R. Bassiri and A. Basti and M. Bawaj and P. Baxi and J. C. Bayley and A. C. Baylor and P. A. Baynard II and M. Bazzan and V. M. Bedakihale and F. Beirnaert and M. Bejger and D. Belardinelli and A. S. Bell and D. S. Bellie and L. Bellizzi and W. Benoit and I. Bentara and J. D. Bentley and M. Ben Yaala and S. Bera and F. Bergamin and B. K. Berger and S. Bernuzzi and M. Beroiz and C. P. L. Berry and D. Bersanetti and T. Bertheas and A. Bertolini and J. Betzwieser and D. Beveridge and G. Bevilacqua and N. Bevins and R. Bhandare and R. Bhatt and D. Bhattacharjee and S. Bhattacharyya and S. Bhaumik and V. Biancalana and A. Bianchi and I. A. Bilenko and G. Billingsley and A. Binetti and S. Bini and C. Binu and S. Biot and O. Birnholtz and S. Biscoveanu and A. Bisht and M. Bitossi and M. -A. Bizouard an
[…]
