+++ paper/paper.md
@@ -66,9 +66,9 @@ bibliography: paper.bib
 
 # Summary
 
-Gravitational waves (GWs) are ripples in spacetime caused by accelerating massive objects, primarily produced by violent cosmic mergers of compact binaries such as binary black holes (BBHs), binary neutron stars (BNSs), and neutron star-black hole pairs (NSBHs). Since 2015, the LIGO-Virgo-KAGRA (LVK) collaboration [@LIGO:2015] [@VIRGO:2015] [@KAGRA:2021] has routinely detected such GW transients during observing runs [@GWTC4:2026] [@GWTC3:2023] [@GWTC2:2021] [@GWTC1:2019]. Most of these signals reach detectors unobstructed, forming a baseline unlensed population, but a small fraction can encounter massive intervening objects like galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott:2021] [@Wierda:2021].
+Gravitational waves (GWs) are ripples in spacetime caused by accelerating massive objects, primarily produced by violent cosmic mergers of compact binaries such as binary black holes (BBHs), binary neutron stars (BNSs), and neutron star-black hole pairs (NSBHs). Since 2015, the LIGO-Virgo-KAGRA (LVK) collaboration [@LIGO:2015; @VIRGO:2015; @KAGRA:2021] has routinely detected such GW transients during observing runs [@GWTC4:2026; @GWTC3:2023; @GWTC2:2021; @GWTC1:2019]. Most of these signals reach detectors unobstructed, forming a baseline unlensed population, but a small fraction can encounter massive intervening objects like galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott:2021; @Wierda:2021].
 
-Accurate modeling of unlensed and lensed gravitational-wave populations is essential for astrophysical interpretation, inferring merger rate histories, and validating future lensing events [@Abbott:2021] [@Janquart:2023] [@ligolensing:2023] [@Wierda:2021]. The $ler$ software is a Python framework that provides a unified pipeline to model these populations statistically and calculate their detectable rates. Producing precise results requires computationally demanding tasks such as large-scale sampling of source and lens attributes, solving lens equations, and simulating detector responses. $ler$ overcomes these traditional bottlenecks through optimization and parallelization, reducing computational time significantly and achieving up to a thousandfold speedup [@lertestgithub:2026] over conventional methods. Source code, validation examples, and comprehensive documentation are available through the project repository [@lergithub:2026], supporting examples [@lertestgithub:2026], and documentation site [@lerdoc:2026], respectively.
+Accurate modeling of unlensed and lensed gravitational-wave populations is essential for astrophysical interpretation, inferring merger rate histories, and validating future lensing events [@Abbott:2021; @Janquart:2023; @ligolensing:2023; @Wierda:2021]. The $ler$ software is a Python framework that provides a unified pipeline to model these populations statistically and calculate their detectable rates. Producing precise results requires computationally demanding tasks such as large-scale sampling of source and lens attributes, solving lens equations, and simulating detector responses. $ler$ overcomes these traditional bottlenecks through optimization and parallelization, reducing computational time significantly and achieving up to a thousandfold speedup [@lertestgithub:2026] over conventional methods. Source code, validation examples, and comprehensive documentation are available through the project repository [@lergithub:2026], supporting examples [@lertestgithub:2026], and documentation site [@lerdoc:2026], respectively.
 
 # Statement of need
 
@@ -80,7 +80,7 @@ $ler$ produces parameter distributions for intrinsic populations and for populat
 
[…]
