### 2026-05-06 cd6f536
paper draft debugging
