+++ paper/paper.bib
@@ -52,7 +52,7 @@
 
 @misc{GWTC4:2026,
       title={GWTC-4.0: Updating the Gravitational-Wave Transient Catalog with Observations from the First Part of the Fourth LIGO-Virgo-KAGRA Observing Run},
-      author={{LIGO Scientific Collaboration} and {Virgo Collaboration} and {KAGRA Collaboration} and others},
+      author={{LIGO Scientific Collaboration et al.}},
       year={2025},
       eprint={2508.18082},
       archivePrefix={arXiv},
@@ -69,7 +69,7 @@
    number={4},
    journal={Physical Review X},
    publisher={American Physical Society (APS)},
-   author={Abbott, R. and others},
+   author={{R. Abbott et al.}},
    year={2023},
    month=dec, pages={041039}
 }
@@ -83,7 +83,7 @@
    number={2},
    journal={Physical Review X},
    publisher={American Physical Society (APS)},
-   author={Abbott, R. and others},
+   author={{R. Abbott et al.}},
    year={2021},
    month=jun, pages={021053}
 }
@@ -97,7 +97,7 @@
    number={3},
    journal={Physical Review X},
    publisher={American Physical Society (APS)},
-   author={Abbott, B.\,P. and Abbott, R. and Abbott, T.\,D. and Abraham, S. and Acernese, F. and Ackley, K. and Adams, C. and Adhikari, R.\,X. and Adya, V.\,B. and Affeldt, C. and Agathos, M. and Agatsuma, K. and Aggarwal, N. and Aguiar, O.\,D. and Aiello, L. and Ain, A. and Ajith, P. and Allen, G. and Allocca, A. and Aloy, M.\,A. and Altin, P.\,A. and Amato, A. and Ananyeva, A. and Anderson, S.\,B. and Anderson, W.\,G. and Angelova, S.\,V. and Antier, S. and Appert, S. and Arai, K. and Araya, M.\,C. and Areeda, J.\,S. and Ar{\`e}ne, M. and Arnaud, N. and Arun, K.\,G. and Ascenzi, S. and Ashton, G. and Aston, S.\,M. and Astone, P. and Aubin, F. and Aufmuth, P. and AultONeal, K. and Austin, C. and Avendano, V. and Avila-Alvarez, A. and Babak, S. and Bacon, P. and Badaracco, F. and Bader, M.\,K.\,M. and Bae, S. and Baker, P.\,T. and Baldaccini, F. and Ballardin, G. and Ballmer, S.\,W. and Banagiri, S. and Barayoga, J.\,C. and Barclay, S.\,E. and Barish, B.\,C. and Barker, D. and Barkett, K. and Barnum, S. and Barone, F. and Barr, B. and Barsotti, L. and Barsuglia, M. and Barta, D. and Bartlett, J. and Bartos, I. and Bassiri, R. and Basti, A. and Bawaj, M. and Bayley, J.\,C. and Bazzan, M. and B{\'e}csy, B. and Bejger, M. and Belahcene, I. and Bell, A.\,S. and Beniwal, D. and Berger, B.\,K. and Bergmann, G. and Bernuzzi, S. and Bero, J.\,J. and Berry, C.\,P.\,L. and Bersanetti, D. and Bertolini, A. and Betzwieser, J. and Bhandare, R. and Bidler, J. and Bilenko, I.\,A. and Bilgili, S.\,A. and Billingsley, G. and Birch, J. and Bi
[…]
