### 2026-05-06 a9ca690
paper draft debugging
