+++ paper/paper.bib
@@ -129,7 +129,7 @@
    ISSN={1432-0746},
    url={http://dx.doi.org/10.1051/0004-6361/201526773},
    DOI={10.1051/0004-6361/201526773},
-   journal={Astronomy &amp; Astrophysics},
+   journal={Astronomy and Astrophysics},
    publisher={EDP Sciences},
    author={Tessore, Nicolas and Benton Metcalf, R.},
    year={2015},
@@ -345,7 +345,7 @@
 
 @misc{ligolensing:2023,
       title={Search for gravitational-lensing signatures in the full third observing run of the LIGO-Virgo network},
-      author={The LIGO Scientific Collaboration and the Virgo Collaboration and the KAGRA Collaboration and R. Abbott and H. Abe and F. Acernese and K. Ackley and S. Adhicary and N. Adhikari and R. X. Adhikari and V. K. Adkins and V. B. Adya and C. Affeldt and D. Agarwal and M. Agathos and O. D. Aguiar and L. Aiello and A. Ain and P. Ajith and T. Akutsu and S. Albanesi and R. A. Alfaidi and C. All{\'e}n{\'e} and A. Allocca and P. A. Altin and A. Amato and S. Anand and A. Ananyeva and S. B. Anderson and W. G. Anderson and M. Ando and T. Andrade and N. Andres and M. Andr{\'e}s-Carcasona and T. Andri{\'c} and S. Ansoldi and J. M. Antelis and S. Antier and T. Apostolatos and E. Z. Appavuravther and S. Appert and S. K. Apple and K. Arai and A. Araya and M. C. Araya and J. S. Areeda and M. Ar{\`e}ne and N. Aritomi and N. Arnaud and M. Arogeti and S. M. Aronson and H. Asada and G. Ashton and Y. Aso and M. Assiduo and S. Assis de Souza Melo and S. M. Aston and P. Astone and F. Aubin and K. AultONeal and S. Babak and F. Badaracco and C. Badger and S. Bae and Y. Bae and S. Bagnasco and Y. Bai and J. G. Baier and J. Baird and R. Bajpai and T. Baka and M. Ball and G. Ballardin and S. W. Ballmer and G. Baltus and S. Banagiri and B. Banerjee and D. Bankar and J. C. Barayoga and B. C. Barish and D. Barker and P. Barneo and F. Barone and B. Barr and L. Barsotti and M. Barsuglia and D. Barta and J. Bartlett and M. A. Barton and I. Bartos and S. Basak and R. Bassiri and A. Basti and M. Bawaj and J. C. Bayley and M. Bazzan and B. B{\'e}csy and V. M. Bedakihale and F. Beirnaert and M. Bejger and I. Belahcene and A. S. Bell and V. Benedetto and D. Beniwal and W. Benoit and J. D. Bentley and M. BenYaala and S. Bera and M. Berbel and F. Bergamin and B. K. Berger and S. Bernuzzi and M. Beroiz and C. P. L. Berry and D. Bersanetti and A. Bertolini and J. Betzwieser and D. Beveridge and R. Bhandare and A. V. Bhandari and U. Bhardwaj and R. Bhatt and D. Bhattacharjee and S. Bhaumik and A. Bianchi and I. A. Bilenko and M. Bilicki and G. Billingsley and S. Bini and O. Birnholtz and S. Biscans and M. Bischi and S. Biscoveanu and A. Bisht and B. Biswas and M. Bitossi and M. -A. Bizouard and J. K. Blackburn and C. D. Blair and D. G. Blair and R. M. Blair and F. Bobba and N. Bode and M. Bo{\"
[…]
