### 2026-05-06 84e6357
paper draft update
