+++ paper/paper.md
@@ -88,8 +88,7 @@ $ler$ is designed as an end-to-end population simulator with a clear separation 
 
 A core design choice is modularity with stable interfaces between modules that handle source populations, lens populations, image-property calculations, and rate estimation. Users can select built-in models such as BBH, BNS, or NSBH populations and lens models such as SIS (singular isothermal sphere), SIE (singular isothermal ellipsoid), or EPL+Shear (elliptical power-law with external shear), or provide their own distributions and detection criteria. Detector selection is handled through a backend interface so that signal-to-noise and detection probability calculations can be delegated to *gwsnr* or replaced by alternative implementations. This architecture supports extension and testing while keeping the scientific assumptions visible in configuration rather than embedded in ad hoc scripts.
 
-Performance is obtained by compiling the repeatedly evaluated parts of the pipeline with *Numba* `njit` and parallelizing the main Monte Carlo loops. $ler$ uses *lenstronomy* [@Birrer:2021] as an optional backend, but it provides an in-house analytical EPL+Shear solver and caustic and cross-section calculator that are compiled with *Numba* and run in parallel, following the same mathematical formulation as *lenstronomy* and Tessore (2015) [@Tessore:2015]. This leads to two practical trade-offs. First, the rate estimates are Monte Carlo approximations, so accuracy is controlled by sample size rather than closed-form expressions, and $ler$ reports convergence statistics from repeated batches. Second, just-in-time compilation adds overhead on first use, which is about 2 s for the first unlensed batch and about 25 s for the first lensed batch that includes image-property calculations. After compilation, the compiled kernels reduce per-sample cost and make large simulations feasible. In benchmarks that disable just-in-time compilation and parallelism, these choices provide speed gains of about $180\times$ for unlensed populations and $5189\times$ for lensed populations. On a standard six-core processor, a batch of $10^5$ samples takes about 300 ms for unlensed calculations and about 30 s for lensed calculations, and repeated batches converge to a stable rate estimate.
-
+Performance is obtained by compiling the repeatedly evaluated parts of the pipeline with *Numba* `njit` and parallelizing the main Monte Carlo loops. $ler$ uses *lenstronomy* [@Birrer:2021] as an optional backend, but it provides an in-house analytical EPL+Shear solver and caustic and cross-section calculator that are compiled with *Numba* and run in parallel, following the same mathematical formulation as *lenstronomy* and Tessore (2015) [@Tessore:2015]. This leads to two practical trade-offs. First, the rate estimates are Monte Carlo approximations, so accuracy is controlled by sample size rather than closed-form expressions, and $ler$ reports convergence statistics from repeated batches. Second, just-in-time compilation adds overhead on first use, which is about 2 s for the first unlensed batch and about 25 s for the first lensed batch that includes image-property calculations. After compilation, the compiled kernels reduce per-sample cost and make large simulations feasible. In benchmarks that disable just-in-time compilation and parallelism, these choices provide speed gains of about $180\times$ for unlensed populations and $5189\times$ for lensed populations. On a standard use of six-core processor (on Apple M2 Pro, 16GB RAM), a batch of $10^5$ samples takes about 300 ms for unlensed calculations and about 30 s for lensed calculations, and repeated batches converge to a stable rate estimate.
 # Mathematics
 
 The mathematical workflow in $ler$ can be summarised as Monte Carlo estimates of detectable event rates for unlensed and strongly lensed populations, obtained by averaging detection probabilities over simulated events drawn from the relevant population models. For unlensed events, the detector-frame rate of detectable events is
