### 2026-05-06 1f79843
doc updates and md files
