+++ ler/rates/ler.py
@@ -150,75 +150,6 @@ class LeR(LensGalaxyParameterDistribution):
     >>> lensed_params = ler.lensed_cbc_statistics()
     >>> ler.lensed_rate()
     >>> ler.rate_ratio()
-
-
-    Instance Methods
-    ----------
-    LeR class has the following methods: \n
-    +-----------------------------------------------------+------------------------------------------------+
-    | Method                                              | Description                                    |
-    +=====================================================+================================================+
-    | :meth:`~unlensed_cbc_statistics`                    | Generate unlensed GW source parameters         |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~unlensed_sampling_routine`                  | Generate unlensed parameters with batching     |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~unlensed_rate`                              | Calculate the unlensed detection rate          |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~lensed_cbc_statistics`                      | Generate lensed GW source parameters           |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~lensed_sampling_routine`                    | Generate lensed parameters with batching       |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~lensed_rate`                                | Calculate the lensed detection rate            |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~rate_function`                              | General helper for rate calculation            |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~rate_ratio`                                 | Calculate lensed/unlensed rate ratio           |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~rate_comparison_with_rate_calculation`      | Calculate and compare lensed/unlensed rates    |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~selecting_n_unlensed_detectable_events`     | Select n unlensed detectable events            |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~selecting_n_lensed_detectable_events`       | Select n lensed detectable events              |
-    +-----------------------------------------------------+------------------------------------------------+
-
-
-    Instance Attributes
-    ----------
-    LeR class has the following attributes: \n
-    +------------------------------------------------+------------------+-------+------------------------------------------------+
-    | Attribute                                      | Type             | Unit  | Description                                    |
-    +================================================+==================+=======+================================================+
-    | :meth:`~npool`                                 | ``int``          |       | Number of parallel processing cores            |
-    +------------------------------------------------+------------------+-------+------------------------------------------------+
-    | :meth:`~z_min`                                 | ``float``        |       | Minimum source redshift                        |
-    +------------------------------------------------+------------------+-------+------------------------------------------------+
-    | :meth:`~z_max`                                 | ``float``        |       | Maximum source redshift                        |
-    +------------------------------------------------+------------------+-------+------------------------------------------------+
-    | :meth:`~event_type`                            | ``str``          |       | Type of CBC event (BBH, BNS, NSBH)             |
-    +------------------------------------------------+------------------+-------+------------------------------------------------+
-    | :meth:`~lens_type`                             | ``str``          |       | Type of lens galaxy model                      |
-    +------------------------------------------------+------------------+-------+------------------------------------------------+
-    | :meth:`~cosmo`                                 | ``Cosmology``    |       | Astropy cosmology object                       |
-    +------------------------------------------------+------------------+-------+------------------------------------------------+
-    | :meth:`~json_file_names`                       | ``dict``         |       | JSON file names for parameter storage          |
[…]
