+++ ler/gw_source_population/cbc_source_parameter_distribution.py
@@ -108,88 +108,6 @@ class CBCSourceParameterDistribution(CBCSourceRedshiftDistribution):
     >>> cbc = CBCSourceParameterDistribution(event_type='BBH')
     >>> params = cbc.gw_parameters_rvs(size=1000)
     >>> print(list(params.keys()))
-
-    Instance Methods
-    ----------
-    CBCSourceParameterDistribution has the following methods: \n
-    +-----------------------------------------------------+------------------------------------------------+
-    | Method                                              | Description                                    |
-    +=====================================================+================================================+
-    | :meth:`~gw_parameters_rvs`                       | Sample all GW parameters for compact binaries  |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~powerlaw_plus_peak`.       | Sample BBH masses with PowerLaw+PEAK model     |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~binary_masses_BBH_popIII_lognormal`         | Sample pop III BBH masses from lognormal       |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~binary_masses_BBH_primordial_lognormal`     | Sample primordial BBH masses from lognormal    |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~binary_masses_NSBH_broken_powerlaw`         | Sample NSBH masses from broken powerlaw        |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~binary_masses_uniform`                      | Sample masses from uniform distribution        |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~binary_masses_BNS_bimodal`                  | Sample BNS masses from bimodal distribution    |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~constant_values_n_size`                     | Return array of constant values                |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~uniform`                            | Sample from uniform distribution               |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~sampler_cosine`                             | Sample from cosine distribution                |
-    +-----------------------------------------------------+------------------------------------------------+
-    | :meth:`~sampler_sine`                               | Sample from sine distribution                  |
-    +-----------------------------------------------------+------------------------------------------------+
-
-    Instance Attributes
-    ----------
-    CBCSourceParameterDistribution has the following attributes: \n
-    +------------------------------------------------+------------------------+-------+------------------------------------------------+
-    | Attribute                                      | Type                   | Unit  | Description                                    |
-    +================================================+========================+=======+================================================+
-    | :attr:`~z_min`                                 | ``float``              |       | Minimum redshift of source population          |
-    +------------------------------------------------+------------------------+-------+------------------------------------------------+
-    | :attr:`~z_max`                                 | ``float``              |       | Maximum redshift of source population          |
-    +------------------------------------------------+------------------------+-------+------------------------------------------------+
-    | :attr:`~cosmo`                                 | ``astropy.cosmology``  |       | Cosmology for distance calculations            |
-    +------------------------------------------------+------------------------+-------+------------------------------------------------+
-    | :attr:`~spin_zero`                             | ``bool``               |       | Whether to ignore spin parameters              |
-    +------------------------------------------------+------------------------+-------+------------------------------------------------+
-    | :attr:`~spin_precession`                       | ``bool``               |       | Whether to use precessing spins                |
-    +------------------------------------------------+------------------------+-------+------------------------------------------------+
-    | :attr:`~directory`                             | ``str``                |       | Directory for interpolator files               |
-    +------------------------------------------------+------------------------+-------+------------------------------------------------+
[…]
