### 2026-05-06 hemantaph/ler#12 @hemantaph
> The summary of this paper read a bit mixed in terms of the purpose of the software. I understand the code can be in principle used for modeling an unlensed population, but the majority of the codebase is geared toward modeling a population of lensed gravitational wave event. For example, the summary page of the doc starts with strong lensing of gravitational wave. I think the summary should more accurately reflect the main intended usage of the code instead of encapsulating all the possible use cases. This helps the user better identify the purpose of the software.

**Reply:** Thank you for pointing this out. I have revised the code structure, examples, tests, documentation, and paper so that the roles of both unlensed and lensed event populations are described more clearly.

> Regarding statement of need, I believe the paper can be benefited from similar focusing as the summary.

**Reply:** Noted. I have revised the Statement of need accordingly. I have also rewritten the paper to follow the current [JOSS paper format](https://joss.readthedocs.io/en/latest/paper.html) more closely:

- [JOSS paper format](https://joss.readthedocs.io/en/latest/paper.html)
- [[JOSS example paper](https://joss.readthedocs.io/en/latest/example_paper.html)](https://joss.readthedocs.io/en/latest/example_paper.html)

> In both the summary and the statement of need, the authors mentioned the calculation needed is computational intensive, but there is not quantitative description of the computational load. Considering some of the example contains these estimates and the speedup provided by this code, I believe quoting an estimate in the paper can help the readers understand the strength of this package. For example there are some estimate hidden in this notebook.

**Reply:** Thank you for the suggestion. I have added quantitative information on the speed gain in the paper. I have also added additional validation material through `pytest` tests and external examples:

- [[pytest tests](https://github.com/hemantaph/ler/tree/main/tests)](https://github.com/hemantaph/ler/tree/main/tests)
- [[lertest repository](https://github.com/hemantaph/lertest)](https://github.com/hemantaph/lertest)

The `lertest` repository is now cited in the paper. It also includes a `uv` environment setup to reduce package-conflict issues. In addition, `ler` has been tested in a [[Google Colab notebook](https://colab.research.google.com/drive/14OMlNVn-cbF6Cddu3NaKVPYNj96a13g6?usp=sharing)](https://colab.research.google.com/drive/14OMlNVn-cbF6Cddu3NaKVPYNj96a13g6?usp=sharing), although it runs more slowly there.

> Also, I am not exactly sure whether equations section is needed for the JOSS paper. They seem quite out of place and were not referenced directly in the paper. Furthermore, more precise versions of the equations are detailed in the docs, so I will suggest to remove the equations.

**Reply:** Thank you for this comment. I have revised the mathematics section so that it is more directly connected to the software design and the main purpose of the package. Since these equations are central to the algorithm and software design, I have kept a concise mathematics section in the paper. This is also consistent with the JOSS example paper, which includes a mathematics section where it is relevant.
