### 2026-05-06 hemantaph/ler#13 @hemantaph
> I do not see any test in the current main branch. Here is the JOSS review guideline regarding testing:

**Reply:** I have now added extensive unit and integration tests using `pytest`:

- [[pytest tests](https://github.com/hemantaph/ler/tree/main/tests)](https://github.com/hemantaph/ler/tree/main/tests)

These tests are also automated through a GitHub workflow.
