### 2026-05-06 hemantaph/ler#14 @hemantaph
> Currently there is no community guideline I could find on the repository. Please adjust accordingly. As a starting point, I can open a PR pasting a template for the community guide I used in the past, and the authors are welcome to either modified it or craft one from scratch.

**Reply:** Thank you for offering to help with this. I have now added a contribution guide:

- [[CONTRIBUTING.md](https://github.com/hemantaph/ler/blob/main/CONTRIBUTING.md)](https://github.com/hemantaph/ler/blob/main/CONTRIBUTING.md)
