### 2026-05-06 bfdf9cd
docs and paper update
