+++ paper/paper.md
@@ -68,7 +68,7 @@ bibliography: paper.bib
 
 Gravitational waves (GWs) are ripples in spacetime produced by accelerating massive objects. Since 2015, the LIGO-Virgo-KAGRA (LVK) collaboration [@LIGO:2015; @VIRGO:2015; @KAGRA:2021] has routinely detected GW transients from compact-binary mergers, including binary black holes (BBHs), binary neutron stars (BNSs), and neutron star-black hole pairs (NSBHs), during observing runs [@GWTC4:2026; @GWTC3:2023; @GWTC2:2021; @GWTC1:2019]. Most of these signals reach detectors unobstructed, forming a baseline unlensed population, but a small fraction can encounter massive intervening objects such as galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott:2021; @Wierda:2021].
 
-Accurate modeling of unlensed and lensed gravitational-wave populations is essential for astrophysical interpretation, inferring merger rate histories, and validating future lensing events [@Abbott:2021; @Janquart:2023; @ligolensing:2023; @Wierda:2021]. The $ler$ software is a Python framework that provides a unified pipeline to model these populations statistically and calculate their detectable rates. Producing precise results requires computationally demanding tasks such as large-scale sampling of source and lens attributes, solving lens equations, and simulating detector responses. $ler$ reduces these costs through optimization and parallelization, achieving up to a thousandfold speedup [@lertestgithub:2026] over conventional methods. Source code, validation examples, and documentation are available through the project repository [@lergithub:2026], supporting examples [@lertestgithub:2026], and documentation site [@lerdoc:2026].
+Accurate modeling of unlensed and lensed gravitational-wave populations is essential for astrophysical interpretation, inferring merger rate histories, and validating future lensing events [@Abbott:2021; @Janquart:2023; @ligolensing:2023; @Wierda:2021]. The $ler$ software is a Python framework that provides a unified pipeline to model these populations statistically and calculate their detectable rates. Producing precise results requires computationally demanding tasks such as large-scale sampling of source and lens attributes, solving lens equations, and simulating detector responses. $ler$ reduces these costs through optimization and parallelization, achieving up to a thousandfold speedup [@lertestgithub:2026] relative to an un-optimized baseline implementation (futher details provided in the \ref{sec:software-design} section). Source code, validation examples, and documentation are available through the project repository [@lergithub:2026], supporting examples [@lertestgithub:2026], and documentation site [@lerdoc:2026].
 
 # Statement of need
 
@@ -82,7 +82,7 @@ Joint population studies of unlensed and strongly lensed gravitational waves req
 
 Previous strong-lensing population studies often relied on pipelines tightly coupled to specific lens models or detection criteria [@Haris:2018; @Wierda:2021]. Many studies report rate estimates based on limited sample sizes without showing convergence diagnostics or uncertainty estimates. They also commonly neglect uncertainty in population-model hyperparameters, which should propagate into the detectable population and rate predictions. $ler$ addresses these gaps by providing an integrated, high-throughput workflow linking population draws, lensing cross-sections, image properties, and detector selection in one reproducible pipeline.
 
-# Software design
+# Software design \label{sec:software-design}
 
[…]
