### 2026-06-05 2514c7d
paper update
