### 2026-06-05 1825a8c
paper update
