+++ requirements_sphinx.txt
@@ -13,6 +13,7 @@ sphinx-jupyterbook-latex==1.0.0
 myst-parser==3.0.1
 linkify-it-py==2.0.3
 sphinx-rtd-dark-mode==1.3.0
+sphinxcontrib-mermaid==2.0.2
 
 # html5lib
 # setuptools
@@ -28,4 +29,5 @@ sphinx-rtd-dark-mode==1.3.0
 # myst-parser
 # linkify-it-py
 # sphinx-rtd-dark-mode
+# sphinxcontrib-mermaid
 # # npm install -g @mermaid-js/mermaid-cli
