+++ pyproject.toml
@@ -62,6 +62,7 @@ docs = [
     "myst-parser>=3.0.1",
     "linkify-it-py>=2.0.3",
     "sphinx-rtd-dark-mode>=1.3.0",
+    "sphinxcontrib-mermaid>=2.0.2",
 ]
 
 [tool.setuptools]
