### 2026-06-10 401d039
uv lock fix; paper update
