+++ paper/paper.md
@@ -78,7 +78,7 @@ Accurate modeling of unlensed and lensed gravitational-wave populations is essen
 
 # State of the field
 
-Joint population studies of unlensed and strongly lensed gravitational waves require end-to-end simulations that combine source and lens-galaxy models with detector selection effects. Common inference packages such as `Bilby` [@Ashton:2019] and cosmology and rate tools such as `gwcosmo` [@gwcosmo:2022] provide general-purpose likelihood evaluation, sampling, and population inference utilities. However, they do not provide a single, configuration-driven pipeline to generate unlensed and strongly lensed populations jointly at scale, nor to propagate lensing image properties into detectability. Lens-modelling tools such as lenstronomy [@Birrer:2021] implement fast analytical routines for lensing calculations but are not designed as high-throughput population-level simulators.
+Joint population studies of unlensed and strongly lensed gravitational waves require end-to-end simulations that combine source and lens-galaxy models with detector selection effects. Common inference packages such as `Bilby` [@Ashton:2019] and cosmology and rate tools such as `gwcosmo` [@gwcosmo:2022] provide general-purpose likelihood evaluation, sampling, and population inference utilities. However, they do not provide a single, configuration-driven pipeline to generate unlensed and strongly lensed populations jointly at scale, nor to propagate lensing image properties into detectability. Lens-modelling tools such as `lenstronomy` [@Birrer:2021] implement fast analytical routines for lensing calculations but are not designed as high-throughput population-level simulators.
 
 Previous strong-lensing population studies often relied on pipelines tightly coupled to specific lens models or detection criteria [@Haris:2018; @Wierda:2021]. Many studies report rate estimates based on limited sample sizes without showing convergence diagnostics or uncertainty estimates. They also commonly neglect uncertainty in population-model hyperparameters, which should propagate into the detectable population and rate predictions. `ler` addresses these gaps by providing an integrated, high-throughput workflow linking population draws, lensing cross-sections, image properties, and detector selection in one reproducible pipeline.
 
