### 2026-06-24 ef6b285
paper update
