### 2026-07-14 b5942db
update UIB acknowledgements
 - Updated acknowledgements section for UIB (N. Singh and D. Keitel)
