+++ paper/paper.md
@@ -123,6 +123,6 @@ No generative AI tools were used to develop the software, write this manuscript,
 
 # Acknowledgements
 
-The authors thank their academic advisors for guidance and support. Hemantakumar Phurailatpam acknowledges the Department of Physics at The Chinese University of Hong Kong for the Postgraduate Studentship that facilitated this research. Hemantakumar Phurailatpam and Otto A. Hannuksela acknowledge support from the Research Grants Council of Hong Kong (Project Nos. CUHK 14304622 and 14307923), the start-up grant from The Chinese University of Hong Kong, and the Direct Grant for Research from the Research Committee of The Chinese University of Hong Kong. The authors also acknowledge support from the Netherlands Organisation for Scientific Research (NWO). N. Singh and D. Keitel are supported by Universitat de les Illes Balears (UIB); the Spanish Agencia Estatal de Investigación grants CNS2022-135440, PID2022-138626NB-I00, RED2022-134204-E, RED2022-134411-T, funded by MICIU/AEI/10.13039/501100011033, the European Union NextGenerationEU/PRTR, and the ERDF/EU; and the Comunitat Autònoma de les Illes Balears through the Direcció General de Recerca, Innovació I Transformació Digital with funds from the Tourist Stay Tax Law (PDR2020/11 - ITS2017-006) as well as through the Conselleria d’Economia, Hisenda i Innovació with grant numbers SINCO2022/6719 (European Union NextGenerationEU/PRTR-C17.I1) and SINCO2022/18146 (co-financed by the European Union and FEDER Operational Program 2021-2027 of the Balearic Islands). The authors acknowledge the computational resources provided by the LIGO Laboratory, supported by National Science Foundation Grants No. PHY-0757058 and No. PHY-0823459.
+The authors thank their academic advisors for guidance and support. Hemantakumar Phurailatpam acknowledges the Department of Physics at The Chinese University of Hong Kong for the Postgraduate Studentship that facilitated this research. Hemantakumar Phurailatpam and Otto A. Hannuksela acknowledge support from the Research Grants Council of Hong Kong (Project Nos. CUHK 14304622 and 14307923), the start-up grant from The Chinese University of Hong Kong, and the Direct Grant for Research from the Research Committee of The Chinese University of Hong Kong. The authors also acknowledge support from the Netherlands Organisation for Scientific Research (NWO). N. Singh is supported through the Conselleria d'Educació i Universitats del Govern de les Illes Balears via an Vicenç Mut postdoctoral grant (POSTDOC2024_55) with funds from the European Social Fund+ in the framework of the Balearic Islands ESF+ Program 2021-2027. N. Singh and D. Keitel are supported by the Universitat de les Illes Balears (UIB) with funds from the Programa de Foment de la Recerca i la Innovació de la UIB 2024-2026 (supported by the yearly plan of the Tourist Stay Tax ITS2023-086); the Spanish Agencia Estatal de Investigación grants PID2022-138626NB-I00, RED2024-153978-E, RED2024-153735-E, funded by MICIU/AEI/10.13039/501100011033 and the ERDF/EU; and the Comunitat Autònoma de les Illes Balears through the Conselleria d'Educació i Universitats with funds from the European Union - European Regional Development Fund (ERDF) (SINCO2022/18146 - Plataforma HiTech-IAC3-BIO).  The authors acknowledge the computational resources provided by the LIGO Laboratory, supported by National Science Foundation Grants No. PHY-0757058 and No. PHY-0823459.
 
 # References
