+++ ler/utils/gwsnr_training_data_generator.py
@@ -18,6 +18,15 @@ class TrainingDataGenerator():
         self.z_max = z_max
         self.verbose = verbose
 
+        pdet_kwargs = dict(
+            snr_th=10.0,
+            snr_th_net=10.0,
+            pdet_type="boolean",
+            distribution_type="noncentral_chi2",
+            include_optimal_snr=True,
+            include_observed_snr=False,
+        )
+
         self.ler_init_args = dict(
             event_type="BBH",
             cosmology=None,
@@ -36,7 +45,9 @@ class TrainingDataGenerator():
             sampling_frequency=2048.0,
             waveform_approximant="IMRPhenomXPHM",
             minimum_frequency=20.0,
-            snr_type="interpolation_aligned_spins",
+            snr_method="inner_product",
+            snr_type="optimal_snr",
+            pdet_kwargs=pdet_kwargs,
             psds=None,
             ifos=None,
             interpolator_dir="./interpolator_json",
@@ -46,6 +57,17 @@ class TrainingDataGenerator():
             mtot_cut=False,
         )
         self.ler_init_args.update(kwargs)
+        if "spin_precessing" in self.ler_init_args:
+            self.ler_init_args["spin_precession"] = self.ler_init_args["spin_precessing"]
+        pdet_kwargs.update(self.ler_init_args.get("pdet_kwargs") or {})
+        self.ler_init_args["pdet_kwargs"] = pdet_kwargs
+
+    @staticmethod
+    def _snr_key(gw_param):
+        for key in ("optimal_snr_net", "snr_net"):
+            if key in gw_param:
+                return key
+        raise KeyError("GW parameter dictionary must contain 'optimal_snr_net' or 'snr_net'.")
 
     def gw_parameters_generator(self, 
         size, 
@@ -63,9 +85,9 @@ class TrainingDataGenerator():
 
         args = self.ler_init_args.copy()
         if snr_recalculation:
-            snr_type = 'interpolation_aligned_spins'
+            snr_method = 'interpolation_aligned_spins'
         else:
-            snr_type = 'inner_product'
+            snr_method = args['snr_method']
 
         # 
         from ler.rates import GWRATES
@@ -87,12 +109,16 @@ class TrainingDataGenerator():
             mtot_max=args['mtot_max'],
             ratio_min=args['ratio_min'],
             ratio_max=args['ratio_max'],
+            spin_max=args['spin_max'],
             mtot_resolution=args['mtot_resolution'],
             ratio_resolution=args['ratio_resolution'],
+            spin_resolution=args['spin_resolution'],
             sampling_frequency=args['sampling_frequency'],
             waveform_approximant=args['waveform_approximant'],
             minimum_frequency=args['minimum_frequency'],
-            snr_type=snr_type,
+            snr_method=snr_method,
+            snr_type=args['snr_type'],
+            pdet_kwargs=args['pdet_kwargs'],
             psds=args['psds'],
             ifos=args['ifos'],
             interpolator_dir=args['interpolator_dir'],
@@ -102,6 +128,33 @@ class TrainingDataGenerator():
             mtot_cut=args['mtot_cut'],
         )
         ler.batch_size = batch_size
+        snr_recalculator = None
+        if snr_recalculation:
+            from gwsnr import GWSNR
+            snr_recalculator = GWSNR(
+                npool=self.npool,
+                snr_method='inner_product',
+                snr_type=args['snr_type'],
+                pdet_kwargs=args['pdet_kwargs'],
+                mtot_min=args['mtot_min'],
+                mtot_max=args['mtot_max'],
+                ratio_min=args['ratio_min'],
+                ratio_max=args['ratio_max'],
+                spin_max=args['spin_max'],
+                mtot_resolution=args['mtot_resolution'],
+                ratio_resolution=args['ratio_resolution'],
[…]
