+++ README.md
@@ -1,6 +1,6 @@
 # LeR
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8087461.svg)](https://doi.org/10.5281/zenodo.8087461)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8087461.svg)](https://doi.org/10.5281/zenodo.8087460)
 [![PyPI version](https://badge.fury.io/py/ler.svg)](https://badge.fury.io/py/ler)
 [![DOCS](https://img.shields.io/badge/docs-GitHub%20Pages-orange)](https://ler.hemantaph.com/)
 
