### 2026-07-19 6a1f43c
mu = 100000.0 # effectively infinite magnification on the critical
