+++ paper/paper.bib
@@ -1,21 +1,32 @@
-@misc{Hannuksela:2026,
-      title={Strong gravitational-wave lensing posterior odds},
-      author={Otto A. Hannuksela and K. Haris and Justin Janquart and Harsh Narola and Hemantakumar Phurailatpam and Jolien D. E. Creighton and Chris Van Den Broeck},
-      year={2026},
-      eprint={2510.15463},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.HE},
-      url={https://arxiv.org/abs/2510.15463},
+@article{Hannuksela:2026,
+doi = {10.3847/1538-4357/ae5816},
+url = {https://doi.org/10.3847/1538-4357/ae5816},
+year = {2026},
+month = {apr},
+publisher = {The American Astronomical Society},
+volume = {1002},
+number = {1},
+pages = {42},
+author = {Hannuksela, Otto A. and Haris, K. and Janquart, Justin and Narola, Harsh and Phurailatpam, Hemantakumar and Creighton, Jolien D. E. and Van Den Broeck, Chris},
+title = {Strong Gravitational-wave Lensing Posterior Odds},
+journal = {The Astrophysical Journal},
+abstract = {Like light, gravitational waves are gravitationally lensed by intervening massive astrophysical objects—such as galaxies, clusters, black holes, and stars—resulting in a variety of potentially observable gravitational-wave lensing signatures. Searches for gravitational-wave lensing by the LIGO–Virgo–KAGRA collaboration have begun. One common method focuses on strong gravitational-wave lensing, which produces multiple “images”: repeated copies of the same gravitational wave that differ only in amplitude, arrival time, and overall “Morse phase.” The literature identifies two separate approaches to identifying such repeated gravitational-wave events, based on frequentist and Bayesian approaches. Several works have discussed selection effects and identified challenges similar to the well-known “birthday problem”—namely, the rapidly increasing likelihood of false alarms in an ever-growing catalog of event pairs. Here, we discuss these problems, unify the different approaches in Bayesian language, and derive the posterior odds for strong lensing. In particular, the Bayes factor and prior odds are sensitive to the number of gravitational-wave events in the data, but the posterior odds is insensitive to it once strong-lensing time delays are accounted for. We confirm the Lo et al. finding that selection effects enter the Bayes factor as an overall normalization constant. However, this factor cancels out in the posterior odds and does not affect frequentist approaches to strong-lensing detection.}
 }
 
-@misc{More:2025,
-      title={Gravitational lensing: towards combining the multi-messengers},
-      author={Anupreeta More and Hemantakumar Phurailatpam},
-      year={2025},
-      eprint={2502.02536},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.CO},
-      url={https://arxiv.org/abs/2502.02536},
+@article{More:2025,
+    author = {More, Anupreeta and Phurailatpam, Hemanta},
+    title = {Gravitational lensing: towards combining the multi-messengers},
+    journal = {Philosophical Transactions of the Royal Society A: Mathematical, Physical and Engineering Sciences},
+    volume = {383},
+    number = {2295},
+    pages = {20240127},
+    year = {2025},
+    month = {05},
[…]
