+++ paper/paper.bib
@@ -45,12 +45,13 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
   year         = {2026}
 }
 
-@misc{lertestgithub:2026,
-  author       = {Phurailatpam, Hemantakumar and collaborators},
-  title        = {{lertest}: Environment and validation examples for {ler}},
-  howpublished = {\url{https://github.com/hemantaph/lertest}},
-  note         = {Accessed 2026-05-06},
-  year         = {2026}
+@software{lertest:2026,
+  author    = {Phurailatpam, Hemantakumar and collaborators},
+  title     = {{lertest}: Environment and validation examples for {ler}},
+  year      = {2026},
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.21768512},
+  url       = {https://doi.org/10.5281/zenodo.21768512}
 }
 
 @misc{lergithub:2026,
@@ -61,14 +62,17 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
   year         = {2026}
 }
 
-@misc{GWTC4:2026,
-      title={GWTC-4.0: Updating the Gravitational-Wave Transient Catalog with Observations from the First Part of the Fourth LIGO-Virgo-KAGRA Observing Run},
-      author={{LIGO Scientific Collaboration et al.}},
-      year={2025},
-      eprint={2508.18082},
-      archivePrefix={arXiv},
-      primaryClass={gr-qc},
-      url={https://arxiv.org/abs/2508.18082},
+@article{GWTC4:2026,
+  author  = {{The LIGO Scientific Collaboration, the Virgo Collaboration, and the KAGRA Collaboration}},
+  title   = {{GWTC-4.0}: Updating the Gravitational-Wave Transient Catalog with Observations from the First Part of the Fourth {LIGO--Virgo--KAGRA} Observing Run},
+  journal = {The Astrophysical Journal Letters},
+  volume  = {1004},
+  number  = {2},
+  pages   = {L22},
+  year    = {2026},
+  month   = jun,
+  doi     = {10.3847/2041-8213/ae2c74},
+  url     = {https://doi.org/10.3847/2041-8213/ae2c74}
 }
 
 @article{GWTC3:2023,
@@ -154,6 +158,7 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
       eprint={1807.07062},
       archivePrefix={arXiv},
       primaryClass={gr-qc},
+      doi={10.48550/arXiv.1807.07062},
       url={https://arxiv.org/abs/1807.07062},
 }
 
@@ -164,6 +169,7 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
       eprint={2412.09888},
       archivePrefix={arXiv},
       primaryClass={astro-ph.IM},
+      doi={10.48550/arXiv.2412.09888},
       url={https://arxiv.org/abs/2412.09888},
 }
 
@@ -235,48 +241,96 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
     doi = {10.3847/1538-4357/ac1bb4}
 }
 
-@article{Wempe:2022,
-    author = "{Wempe}, Ewoud and Koopmans, L{\'e}on V.E. and Wierda, A. Renske A.C. and Hannuksela, Otto Akseli and Agnello, Alberto and Bonvin, Cyril and Bucciarelli, Bendetta and Camera, Chiara and Czoske, Oliver and Finke, Carsten and others",
-    title = {A lensing multi-messenger channel: Combining LIGO-Virgo-Kagra lensed gravitational-wave measurements with Euclid observations},
[…]
