+++ paper/paper.md
@@ -63,7 +63,7 @@ affiliations:
     index: 8
   - name: IAC3, Universitat de les Illes Balears, Crta. Valldemossa km 7.5, E-07122 Palma, Spain
     index: 9
-date: 6 May 2026
+date: 3 August 2026
 bibliography: paper.bib
 ---
 
@@ -71,13 +71,13 @@ bibliography: paper.bib
 
 Gravitational waves (GWs) are ripples in spacetime produced by accelerating massive objects. Since 2015, the LIGO-Virgo-KAGRA (LVK) collaboration [@LIGO:2015; @VIRGO:2015; @KAGRA:2021] has routinely detected GW transients from compact-binary mergers, including binary black holes (BBHs), binary neutron stars (BNSs), and neutron star-black hole pairs (NSBHs), during observing runs [@GWTC4:2026; @GWTC3:2023; @GWTC2:2021; @GWTC1:2019]. Most of these signals reach detectors unobstructed, forming a baseline unlensed population, but a small fraction can encounter massive intervening objects such as galaxies or galaxy clusters. These objects act as lenses, splitting the signals into multiple images that are magnified, time-shifted, and phase-shifted through strong gravitational lensing [@Abbott:2021; @Wierda:2021].
 
-Accurate modeling of unlensed and lensed gravitational-wave populations is essential for astrophysical interpretation, inferring merger rate histories, and validating future lensing events [@Abbott:2021; @Janquart:2023; @ligolensing:2024; @Wierda:2021]. The `ler` software is a Python framework that provides a unified pipeline to model these populations statistically and calculate their detectable rates. Producing precise results requires computationally demanding tasks such as large-scale sampling of source and lens attributes, solving lens equations, and simulating detector responses. `ler` reduces these costs through optimization and parallelization, achieving up to a thousandfold speedup [@lertest:2026] relative to an un-optimized baseline implementation (further details provided in the Software design section). Source code, validation examples, and documentation are available through the project repository, [supporting examples](https://github.com/hemantaph/lertest), and [documentation site](https://ler.hemantaph.com).
+Accurate modeling of unlensed and lensed gravitational-wave populations is essential for astrophysical interpretation, inferring merger rate histories, and validating future lensing events [@Abbott:2021; @Janquart:2023; @ligolensing:2024; @Wierda:2021]. The `ler` software is a Python framework that provides a unified pipeline to model these populations statistically and calculate their detectable rates. Producing precise results requires computationally demanding tasks such as large-scale sampling of source and lens attributes, solving lens equations, and simulating detector responses. `ler` reduces these costs through optimization and parallelization, achieving up to a thousandfold speedup relative to an un-optimized baseline implementation (further details provided in the Software design section). Source code, validation examples, and documentation are available through the project repository, [supporting examples](https://github.com/hemantaph/lertest), and [documentation site](https://ler.hemantaph.com).
 
 # Statement of need
 
[…]
