+++ paper/paper.bib
@@ -254,7 +254,7 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
   url     = {https://doi.org/10.1093/mnras/stae1023}
 }
 
-@article{numpy:2022,
+@article{numpy:2020,
     author = {Harris, Charles R. and Millman, K. Jarrod and {van der Walt}, St{\'e}fan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and {van Kerkwijk}, Marten H. and Brett, Matthew and Haldane, Allan and Fern{\'a}ndez del R{\'i}o, Jaime and Wiebe, Mark and Peterson, Pearu and G{\'e}rard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
     title = {Array programming with {NumPy}},
     journal = {Nature},
@@ -268,7 +268,7 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
     publisher = {Springer Nature}
 }
 
-@inproceedings{numba:2022,
+@inproceedings{numba:2015,
     author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
     title = {Numba: A {LLVM}-based Python {JIT} compiler},
     booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in {HPC}},
