### 2026-08-03 d42af50
reply to editor comments
