+++ paper/paper.md
@@ -113,7 +113,7 @@ where $\mathcal{N}_{\mathrm{L}}$ is the total intrinsic merger rate in the detec
 
 # Research impact statement
 
-`ler` has established a measurable impact within the gravitational-wave community through its integration into research workflows and its growing user base. The package supports a diverse community ranging from students to faculty members at multiple institutions, with development driven by researchers across various organizations and peer-reviewed within the LVK Scientific Collaboration [@lerlvkpnpreview:2024]. Active community engagement is maintained through bug reports and feature requests on the GitHub issue tracker and through collaboration communication channels such as LIGO Mattermost.
+`ler` has established a measurable impact within the gravitational-wave community through its integration into research workflows and its growing user base. The package supports a diverse community ranging from students to faculty members at multiple institutions, with development driven by researchers across various organizations and [peer-reviewed within the LVK Scientific Collaboration](https://dcc.ligo.org/LIGO-P2400274). Active community engagement is maintained through bug reports and feature requests on the GitHub issue tracker and through collaboration communication channels such as LIGO Mattermost.
 
 The software has facilitated several scientific publications related to gravitational-wave lensing statistics. It has been used to analyze the distribution of time delays and magnification ratios for the validation of lensing candidates [@Janquart:2023] and to forecast the detection of sub-threshold lensed events with future observatories [@Ng:2025]. Additional research applications include the generation of lensed-event statistics and the calculation of posterior odds [@Hannuksela:2026], as well as forecasting electromagnetic counterparts for strongly lensed gamma-ray bursts [@More:2025]. These implementations demonstrate that the tool provides a robust framework for high-level population studies and astrophysical forecasting.
 
