+++ paper/paper.bib
@@ -29,22 +29,6 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
     eprint = {https://royalsocietypublishing.org/rsta/article-pdf/doi/10.1098/rsta.2024.0127/2818173/rsta.2024.0127.pdf},
 }
 
-@misc{lerlvkpnpreview:2024,
-  author       = {{LIGO-Virgo-KAGRA Collaboration}},
-  title        = {LIGO DCC Document LIGO-P2400274: {P\&P} internal review of {ler}},
-  howpublished = {\url{https://dcc.ligo.org/LIGO-P2400274}},
-  note         = {Internal LVK DCC document (requires collaboration access). Accessed 2026-05-06},
-  year         = {2024}
-}
-
-@misc{lerpypi:2026,
-  author       = {Phurailatpam, Hemantakumar and collaborators},
-  title        = {{ler}: PyPI package page},
-  howpublished = {\url{https://pypi.org/project/ler/}},
-  note         = {Accessed 2026-05-06},
-  year         = {2026}
-}
-
 @software{lertest:2026,
   author    = {Phurailatpam, Hemantakumar and collaborators},
   title     = {{lertest}: Environment and validation examples for {ler}},
@@ -52,14 +36,6 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
   publisher = {Zenodo},
   doi       = {10.5281/zenodo.21768512},
   url       = {https://doi.org/10.5281/zenodo.21768512}
-}
-
-@misc{lergithub:2026,
-  author       = {Phurailatpam, Hemantakumar and collaborators},
-  title        = {{ler}: Source code for {ler}},
-  howpublished = {\url{https://github.com/hemantaph/ler}},
-  note         = {Accessed 2026-05-06},
-  year         = {2026}
 }
 
 @article{GWTC4:2026,
@@ -127,15 +103,6 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
    publisher={Cambridge University Press (CUP)},
    author={Thrane, Eric and Talbot, Colm},
    year={2019}
-}
-
-@misc{lerdoc:2026,
-  author = {Phurailatpam, Hemantakumar and More, Anupreeta and Narola, Harsh and Ng, Chung Yin and Janquart, Justin and Van Den Broeck, Chris and Hannuksela, Otto Akseli and Singh, Neha and Keitel, David},
-  title = {ler: A Python package for gravitational-wave lensed and unlensed event rate calculations},
-  year = {2026},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://ler.hemantaph.com}},
 }
 
 @article{Tessore:2015,
