### 2026-08-06 35e74c5
Initial tweaks for JOSS submission ahead of publication
