+++ paper/paper.bib
@@ -52,7 +52,7 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
 }
 
 @article{GWTC3:2023,
-   title={GWTC-3: Compact Binary Coalescences Observed by LIGO and Virgo during the Second Part of the Third Observing Run},
+   title={GWTC-3: Compact Binary Coalescences Observed by LIGO and {Virgo} during the Second Part of the Third Observing Run},
    volume={13},
    ISSN={2160-3308},
    url={http://dx.doi.org/10.1103/PhysRevX.13.041039},
@@ -60,13 +60,13 @@ abstract = {Like light, gravitational waves are gravitationally lensed by interv
    number={4},
    journal={Physical Review X},
    publisher={American Physical Society (APS)},
-   author={{R. Abbott et al.}},
+   author = {Abbott, R. and Abbott, T. D. and Acernese, F. and Ackley, K. and Adams, C. and Adhikari, N. and Adhikari, R. X. and Adya, V. B. and Affeldt, C. and Agarwal, D. and Agathos, M. and Agatsuma, K. and Aggarwal, N. and Aguiar, O. D. and Aiello, L. and Ain, A. and Ajith, P. and Akcay, S. and Akutsu, T. and Albanesi, S. and Allocca, A. and Altin, P. A. and Amato, A. and Anand, C. and Anand, S. and Ananyeva, A. and Anderson, S. B. and Anderson, W. G. and Ando, M. and Andrade, T. and Andres, N. and Andri\ifmmode \acute{c}\else \'{c}\fi{}, T. and Angelova, S. V. and Ansoldi, S. and Antelis, J. M. and Antier, S. and Appert, S. and Arai, Koji and Arai, Koya and Arai, Y. and Araki, S. and Araya, A. and Araya, M. C. and Areeda, J. S. and Ar\`ene, M. and Aritomi, N. and Arnaud, N. and Arogeti, M. and Aronson, S. M. and Arun, K. G. and Asada, H. and Asali, Y. and Ashton, G. and Aso, Y. and Assiduo, M. and Aston, S. M. and Astone, P. and Aubin, F. and Austin, C. and Babak, S. and Badaracco, F. and Bader, M. K. M. and Badger, C. and Bae, S. and Bae, Y. and Baer, A. M. and Bagnasco, S. and Bai, Y. and Baiotti, L. and Baird, J. and Bajpai, R. and Ball, M. and Ballardin, G. and Ballmer, S. W. and Balsamo, A. and Baltus, G. and Banagiri, S. and Bankar, D. and Barayoga, J. C. and Barbieri, C. and Barish, B. C. and Barker, D. and Barneo, P. and Barone, F. and Barr, B. and Barsotti, L. and Barsuglia, M. and Barta, D. and Bartlett, J. and Barton, M. A. and Bartos, I. and Bassiri, R. and Basti, A. and Bawaj, M. and Bayley, J. C. and Baylor, A. C. and Bazzan, M. and B\'ecsy, B. and Bedakihale, V. M. and Bejger, M. and Belahcene, I. and Benedetto, V. and Beniwal, D. and Bennett, T. F. and Bentley, J. D. and BenYaala, M. and Bergamin, F. and Berger, B. K. and Bernuzzi, S. and Berry, C. P. L. and Bersanetti, D. and Bertolini, A. and Betzwieser, J. and Beveridge, D. and Bhandare, R. and Bhardwaj, U. and Bhattacharjee, D. and Bhaumik, S. and Bilenko
[…]
