### 2026-08-16 #7420 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
