### 2026-08-16 #7420 @warrickball
Thanks for the merge, @hemantaph. With those final changes, your paper has now been published in JOSS!

Thanks once again to @sibirrer & @Jammy2211 for reviewing. JOSS simply wouldn't be possible without its community of volunteers.
