### 2026-08-17 #7420 @hemantaph
@warrickball @sibirrer @Jammy2211
Thank you to the editor and reviewers; this means a lot to me. I also thank all the co-authors.

I will respond to the JOSS effort.

> 🎉🎉🎉 Congratulations on your paper acceptance! 🎉🎉🎉
>
> If you would like to include a link to your paper from your README use the following
>
> **code snippets**
>
> ```
> Markdown:
>
> HTML:
> <a style="border-width:0" href="https://doi.org/10.21105/joss.07420">
> </a>
>
> reStructuredText:
>    :target: https://doi.org/10.21105/joss.07420
> ```
>
> This is how it will look in your documentation:
>
> [![DOI](https://camo.githubusercontent.com/195a928dbe86d1a1fd2a97bbdab30d45fee333cec7b63aff2a21d93f6fa6b82f/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f31302e32313130352f6a6f73732e30373432302f7374617475732e737667)](https://doi.org/10.21105/joss.07420)
>
> **We need your help!**
>
> The Journal of Open Source Software is a community-run journal and relies upon volunteer effort. If you'd like to support us please consider doing either one (or both) of the the following:
>
> * Volunteering to review for us sometime in the future. You can add your name to the reviewer list here: https://reviewers.joss.theoj.org/join
> * Making a small donation to support our running costs here: https://numfocus.org/donate-to-open-journals
